fetch COVID hospitalization data for specific facilities

Obtains the COVID-19 reported patient impact and hospital capacity data by facility. This dataset is provided by the US Department of Health & Human Services via healthdata.gov.

covid_hosp_facility(hospital_pks, collection_weeks, publication_dates = NULL)

Arguments

hospital_pks

A character string of facility unique identifiers

collection_weeks

Takes in the form epirange(startdate,enddate), where startdate and enddate are of the form YYYYMMDD (can be passed as string or numeric).

publication_dates

An optional epirange publication dates to fetch, using epirange(startdate,enddate).

Value

an instance of epidata_call

Details

Starting October 1, 2022, some facilities are only required to report annually. The companion function covid_hosp_facility_lookup() can be used to look up facility identifiers in a variety of ways.

References

API documentation: https://cmu-delphi.github.io/delphi-epidata/api/covid_hosp_facility.html

See also the official description and data dictionary at <healthdata.gov> for more information.

See also

Examples

call <- covid_hosp_facility(hospital_pks = "100075", collection_weeks = epirange(20200101, 20200501))
fetch_csv(call)
#> [1] "hospital_pk,state,ccn,hospital_name,address,city,zip,hospital_subtype,fips_code,geocoded_hospital_address,hhs_ids,publication_date,collection_week,is_metro_micro,total_beds_7_day_sum,all_adult_hospital_beds_7_day_sum,all_adult_hospital_inpatient_beds_7_day_sum,inpatient_beds_used_7_day_sum,all_adult_hospital_inpatient_bed_occupied_7_day_sum,total_adult_patients_hosp_confirmed_suspected_covid_7d_sum,total_adult_patients_hospitalized_confirmed_covid_7_day_sum,total_pediatric_patients_hosp_confirmed_suspected_covid_7d_sum,total_pediatric_patients_hospitalized_confirmed_covid_7_day_sum,inpatient_beds_7_day_sum,total_icu_beds_7_day_sum,total_staffed_adult_icu_beds_7_day_sum,icu_beds_used_7_day_sum,staffed_adult_icu_bed_occupancy_7_day_sum,staffed_icu_adult_patients_confirmed_suspected_covid_7d_sum,staffed_icu_adult_patients_confirmed_covid_7_day_sum,total_patients_hospitalized_confirmed_influenza_7_day_sum,icu_patients_confirmed_influenza_7_day_sum,total_patients_hosp_confirmed_influenza_and_covid_7d_sum,total_beds_7_day_coverage,all_adult_hospital_beds_7_day_coverage,all_adult_hospital_inpatient_beds_7_day_coverage,inpatient_beds_used_7_day_coverage,all_adult_hospital_inpatient_bed_occupied_7_day_coverage,total_adult_patients_hosp_confirmed_suspected_covid_7d_cov,total_adult_patients_hospitalized_confirmed_covid_7_day_coverage,total_pediatric_patients_hosp_confirmed_suspected_covid_7d_cov,total_pediatric_patients_hosp_confirmed_covid_7d_cov,inpatient_beds_7_day_coverage,total_icu_beds_7_day_coverage,total_staffed_adult_icu_beds_7_day_coverage,icu_beds_used_7_day_coverage,staffed_adult_icu_bed_occupancy_7_day_coverage,staffed_icu_adult_patients_confirmed_suspected_covid_7d_cov,staffed_icu_adult_patients_confirmed_covid_7_day_coverage,total_patients_hospitalized_confirmed_influenza_7_day_coverage,icu_patients_confirmed_influenza_7_day_coverage,total_patients_hosp_confirmed_influenza_and_covid_7d_cov,previous_day_admission_adult_covid_confirmed_7_day_sum,previous_day_admission_adult_covid_confirmed_18_19_7_day_sum,previous_day_admission_adult_covid_confirmed_20_29_7_day_sum,previous_day_admission_adult_covid_confirmed_30_39_7_day_sum,previous_day_admission_adult_covid_confirmed_40_49_7_day_sum,previous_day_admission_adult_covid_confirmed_50_59_7_day_sum,previous_day_admission_adult_covid_confirmed_60_69_7_day_sum,previous_day_admission_adult_covid_confirmed_70_79_7_day_sum,previous_day_admission_adult_covid_confirmed_80plus_7_day_sum,previous_day_admission_adult_covid_confirmed_unknown_7_day_sum,previous_day_admission_pediatric_covid_confirmed_7_day_sum,previous_day_covid_ed_visits_7_day_sum,previous_day_admission_adult_covid_suspected_7_day_sum,previous_day_admission_adult_covid_suspected_18_19_7_day_sum,previous_day_admission_adult_covid_suspected_20_29_7_day_sum,previous_day_admission_adult_covid_suspected_30_39_7_day_sum,previous_day_admission_adult_covid_suspected_40_49_7_day_sum,previous_day_admission_adult_covid_suspected_50_59_7_day_sum,previous_day_admission_adult_covid_suspected_60_69_7_day_sum,previous_day_admission_adult_covid_suspected_70_79_7_day_sum,previous_day_admission_adult_covid_suspected_80plus_7_day_sum,previous_day_admission_adult_covid_suspected_unknown_7_day_sum,previous_day_admission_pediatric_covid_suspected_7_day_sum,previous_day_total_ed_visits_7_day_sum,previous_day_admission_influenza_confirmed_7_day_sum,is_corrected,previous_day_admission_adult_covid_confirmed_7_day_coverage,previous_day_admission_adult_covid_suspected_7_day_coverage,previous_day_admission_pediatric_covid_confirmed_7_day_coverage,previous_day_admission_pediatric_covid_suspected_7_day_coverage,previous_week_patients_covid_vaccinated_doses_all_7_day,previous_week_patients_covid_vaccinated_doses_all_7_day_sum,previous_week_patients_covid_vaccinated_doses_one_7_day,previous_week_patients_covid_vaccinated_doses_one_7_day_sum,previous_week_personnel_covid_vaccd_doses_administered_7_day,previous_week_personnel_covid_vaccd_doses_administered_7_day_sum,total_personnel_covid_vaccinated_doses_all_7_day,total_personnel_covid_vaccinated_doses_all_7_day_sum,total_personnel_covid_vaccinated_doses_none_7_day,total_personnel_covid_vaccinated_doses_none_7_day_sum,total_personnel_covid_vaccinated_doses_one_7_day,total_personnel_covid_vaccinated_doses_one_7_day_sum,total_beds_7_day_avg,all_adult_hospital_beds_7_day_avg,all_adult_hospital_inpatient_beds_7_day_avg,inpatient_beds_used_7_day_avg,all_adult_hospital_inpatient_bed_occupied_7_day_avg,total_adult_patients_hosp_confirmed_suspected_covid_7d_avg,total_adult_patients_hospitalized_confirmed_covid_7_day_avg,total_pediatric_patients_hosp_confirmed_suspected_covid_7d_avg,total_pediatric_patients_hospitalized_confirmed_covid_7_day_avg,inpatient_beds_7_day_avg,total_icu_beds_7_day_avg,total_staffed_adult_icu_beds_7_day_avg,icu_beds_used_7_day_avg,staffed_adult_icu_bed_occupancy_7_day_avg,staffed_icu_adult_patients_confirmed_suspected_covid_7d_avg,staffed_icu_adult_patients_confirmed_covid_7_day_avg,total_patients_hospitalized_confirmed_influenza_7_day_avg,icu_patients_confirmed_influenza_7_day_avg,total_patients_hosp_confirmed_influenza_and_covid_7d_avg\n100075,FL,100075,ST JOSEPHS HOSPITAL,3001 W MARTIN LUTHER KING JR BLVD,TAMPA,33677,Short Term,12057,POINT (-82.492263 27.983390),\"[C100075-D, C100075-A, C100075-B, C100075-F, C100075-C, C100075-E]\",20230731,20200322,1,,,,,,,,,,,160,,,,,,,,,0,0,0,0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,,,,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,80.0,,,,,,,,\n100075,FL,100075,ST JOSEPHS HOSPITAL,3001 W MARTIN LUTHER KING JR BLVD,TAMPA,33677,Short Term,12057,POINT (-82.492263 27.983390),\"[C100075-D, C100075-A, C100075-B, C100075-F, C100075-C, C100075-E]\",20230623,20200327,1,,,,,,,,,,,615,,,,,,,,,0,0,0,0,0,0,0,0,0,0,7,0,0,0,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,,,,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,87.9,,,,,,,,\n100075,FL,100075,ST JOSEPHS HOSPITAL,3001 W MARTIN LUTHER KING JR BLVD,TAMPA,33677,Short Term,12057,POINT (-82.492263 27.983390),\"[C100075-D, C100075-A, C100075-B, C100075-F, C100075-C, C100075-E]\",20230731,20200329,1,,,,,,,,,,,654,,,,,,,,,0,0,0,0,0,0,0,0,0,0,7,0,0,0,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,,,,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,93.4,,,,,,,,\n100075,FL,100075,ST JOSEPHS HOSPITAL,3001 W MARTIN LUTHER KING JR BLVD,TAMPA,33677,Short Term,12057,POINT (-82.492263 27.983390),\"[C100075-D, C100075-A, C100075-B, C100075-F, C100075-C, C100075-E]\",20230623,20200403,1,,,,,,,,,,,527,,,,,,,,,0,0,0,0,0,0,0,0,0,0,6,0,0,0,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,,,,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,87.8,,,,,,,,\n100075,FL,100075,ST JOSEPHS HOSPITAL,3001 W MARTIN LUTHER KING JR BLVD,TAMPA,33677,Short Term,12057,POINT (-82.492263 27.983390),\"[C100075-D, C100075-A, C100075-B, C100075-F, C100075-C, C100075-E]\",20230731,20200405,1,,,,,,,,,,,521,,,,,,,,,0,0,0,0,0,0,0,0,0,0,6,0,0,0,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,,,,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,86.8,,,,,,,,\n100075,FL,100075,ST JOSEPHS HOSPITAL,3001 W MARTIN LUTHER KING JR BLVD,TAMPA,33677,Short Term,12057,POINT (-82.492263 27.983390),\"[C100075-D, C100075-A, C100075-B, C100075-F, C100075-C, C100075-E]\",20230623,20200410,1,,,,,,,,,,,668,,,,,,,,,0,0,0,0,0,0,0,0,0,0,7,0,0,0,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,,,,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,95.4,,,,,,,,\n100075,FL,100075,ST JOSEPHS HOSPITAL,3001 W MARTIN LUTHER KING JR BLVD,TAMPA,33677,Short Term,12057,POINT (-82.492263 27.983390),\"[C100075-D, C100075-A, C100075-B, C100075-F, C100075-C, C100075-E]\",20230731,20200412,1,,,,,,,,,,,660,,,,,,,,,0,0,0,0,0,0,0,0,0,0,7,0,0,0,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,,,,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,94.3,,,,,,,,\n100075,FL,100075,ST JOSEPHS HOSPITAL,3001 W MARTIN LUTHER KING JR BLVD,TAMPA,33677,Short Term,12057,POINT (-82.492263 27.983390),\"[C100075-D, C100075-A, C100075-B, C100075-F, C100075-C, C100075-E]\",20230623,20200417,1,,,,,,,,,,,666,,,,,,,,,0,0,0,0,0,0,0,0,0,0,7,0,0,0,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,,,,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,95.1,,,,,,,,\n100075,FL,100075,ST JOSEPHS HOSPITAL,3001 W MARTIN LUTHER KING JR BLVD,TAMPA,33677,Short Term,12057,POINT (-82.492263 27.983390),\"[C100075-D, C100075-A, C100075-B, C100075-F, C100075-C, C100075-E]\",20230731,20200419,1,,,,,,,,,,,665,,,,,,,,,0,0,0,0,0,0,0,0,0,0,7,0,0,0,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,,,,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,95.0,,,,,,,,\n100075,FL,100075,ST JOSEPHS HOSPITAL,3001 W MARTIN LUTHER KING JR BLVD,TAMPA,33677,Short Term,12057,POINT (-82.492263 27.983390),\"[C100075-D, C100075-A, C100075-B, C100075-F, C100075-C, C100075-E]\",20230623,20200424,1,,,,,,,,,,,661,,,,,,,,,0,0,0,0,0,0,0,0,0,0,7,0,0,0,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,,,,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,94.4,,,,,,,,\n100075,FL,100075,ST JOSEPHS HOSPITAL,3001 W MARTIN LUTHER KING JR BLVD,TAMPA,33677,Short Term,12057,POINT (-82.492263 27.983390),\"[C100075-D, C100075-A, C100075-B, C100075-F, C100075-C, C100075-E]\",20230731,20200426,1,,,,,,,,,,,660,,,,,,,,,0,0,0,0,0,0,0,0,0,0,7,0,0,0,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,,,,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,94.3,,,,,,,,\n100075,FL,100075,ST JOSEPHS HOSPITAL,3001 W MARTIN LUTHER KING JR BLVD,TAMPA,33677,Short Term,12057,POINT (-82.492263 27.983390),\"[C100075-D, C100075-A, C100075-B, C100075-F, C100075-C, C100075-E]\",20230623,20200501,1,,,,,,,,,,,661,,,,,,,,,0,0,0,0,0,0,0,0,0,0,7,0,0,0,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,,,,0,0,0,0,0,,,,,,,,,,,,,,,,,,,,,,,94.4,,,,,,,,\n"
#> attr(,"class")
#> [1] "epidata_csv" "character"