Fetch COVID Hospitalization Data by State

Obtains the COVID-19 reported patient impact and hospital capacity data by state. This dataset is provided by the US Department of Health & Human Services via healthdata.gov.

covid_hosp_state_timeseries(states, dates, issues = NULL)

Arguments

states

A two letter character string state abbreviation.

dates

Takes in the form epirange(startdate,enddate), where startdate and enddate are of the form YYYYMMDD (can be passed as string or numeric).

issues

An optional parameter that takes the form YYYYMMDD. If issues is not specified, then the most recent issue is used by default.

Value

an instance of epidata_call

Details

Starting October 1, 2022, some facilities are only required to report annually.

References

API Documentation: https://cmu-delphi.github.io/delphi-epidata/api/covid_hosp.html.

See also the official description and data dictionary at <healthdata.gov> for more information.

Examples

# \donttest{
call <- covid_hosp_state_timeseries(states = "fl", dates = epirange(20200101, 20200501))
fetch_csv(call)
#> [1] "state,geocoded_state,issue,date,critical_staffing_shortage_today_yes,critical_staffing_shortage_today_no,critical_staffing_shortage_today_not_reported,critical_staffing_shortage_anticipated_within_week_yes,critical_staffing_shortage_anticipated_within_week_no,critical_staffing_shortage_anticipated_within_week_not_reported,hospital_onset_covid,hospital_onset_covid_coverage,inpatient_beds,inpatient_beds_coverage,inpatient_beds_used,inpatient_beds_used_coverage,inpatient_beds_used_covid,inpatient_beds_used_covid_coverage,previous_day_admission_adult_covid_confirmed,previous_day_admission_adult_covid_confirmed_coverage,previous_day_admission_adult_covid_suspected,previous_day_admission_adult_covid_suspected_coverage,previous_day_admission_pediatric_covid_confirmed,previous_day_admission_pediatric_covid_confirmed_coverage,previous_day_admission_pediatric_covid_suspected,previous_day_admission_pediatric_covid_suspected_coverage,staffed_adult_icu_bed_occupancy,staffed_adult_icu_bed_occupancy_coverage,staffed_icu_adult_patients_confirmed_suspected_covid,staffed_icu_adult_patients_confirmed_suspected_covid_coverage,staffed_icu_adult_patients_confirmed_covid,staffed_icu_adult_patients_confirmed_covid_coverage,total_adult_patients_hosp_confirmed_suspected_covid,total_adult_patients_hosp_confirmed_suspected_covid_coverage,total_adult_patients_hosp_confirmed_covid,total_adult_patients_hosp_confirmed_covid_coverage,total_pediatric_patients_hosp_confirmed_suspected_covid,total_pediatric_patients_hosp_confirmed_suspected_covid_coverage,total_pediatric_patients_hosp_confirmed_covid,total_pediatric_patients_hosp_confirmed_covid_coverage,total_staffed_adult_icu_beds,total_staffed_adult_icu_beds_coverage,inpatient_beds_utilization_coverage,inpatient_beds_utilization_numerator,inpatient_beds_utilization_denominator,percent_of_inpatients_with_covid_coverage,percent_of_inpatients_with_covid_numerator,percent_of_inpatients_with_covid_denominator,inpatient_bed_covid_utilization_coverage,inpatient_bed_covid_utilization_numerator,inpatient_bed_covid_utilization_denominator,adult_icu_bed_covid_utilization_coverage,adult_icu_bed_covid_utilization_numerator,adult_icu_bed_covid_utilization_denominator,adult_icu_bed_utilization_coverage,adult_icu_bed_utilization_numerator,adult_icu_bed_utilization_denominator,deaths_covid,deaths_covid_coverage,icu_patients_confirmed_influenza,icu_patients_confirmed_influenza_coverage,on_hand_supply_therapeutic_a_casirivimab_imdevimab_courses,on_hand_supply_therapeutic_b_bamlanivimab_courses,on_hand_supply_therapeutic_c_bamlanivimab_etesevimab_courses,previous_day_admission_adult_covid_confirmed_18_19,previous_day_admission_adult_covid_confirmed_18_19_coverage,previous_day_admission_adult_covid_confirmed_20_29,previous_day_admission_adult_covid_confirmed_20_29_coverage,previous_day_admission_adult_covid_confirmed_30_39,previous_day_admission_adult_covid_confirmed_30_39_coverage,previous_day_admission_adult_covid_confirmed_40_49,previous_day_admission_adult_covid_confirmed_40_49_coverage,previous_day_admission_adult_covid_confirmed_50_59,previous_day_admission_adult_covid_confirmed_50_59_coverage,previous_day_admission_adult_covid_confirmed_60_69,previous_day_admission_adult_covid_confirmed_60_69_coverage,previous_day_admission_adult_covid_confirmed_70_79,previous_day_admission_adult_covid_confirmed_70_79_coverage,previous_day_admission_adult_covid_confirmed_80plus,previous_day_admission_adult_covid_confirmed_80plus_coverage,previous_day_admission_adult_covid_confirmed_unknown,previous_day_admission_adult_covid_confirmed_unknown_coverage,previous_day_admission_adult_covid_suspected_18_19,previous_day_admission_adult_covid_suspected_18_19_coverage,previous_day_admission_adult_covid_suspected_20_29,previous_day_admission_adult_covid_suspected_20_29_coverage,previous_day_admission_adult_covid_suspected_30_39,previous_day_admission_adult_covid_suspected_30_39_coverage,previous_day_admission_adult_covid_suspected_40_49,previous_day_admission_adult_covid_suspected_40_49_coverage,previous_day_admission_adult_covid_suspected_50_59,previous_day_admission_adult_covid_suspected_50_59_coverage,previous_day_admission_adult_covid_suspected_60_69,previous_day_admission_adult_covid_suspected_60_69_coverage,previous_day_admission_adult_covid_suspected_70_79,previous_day_admission_adult_covid_suspected_70_79_coverage,previous_day_admission_adult_covid_suspected_80plus,previous_day_admission_adult_covid_suspected_80plus_coverage,previous_day_admission_adult_covid_suspected_unknown,previous_day_admission_adult_covid_suspected_unknown_coverage,previous_day_admission_influenza_confirmed,previous_day_admission_influenza_confirmed_coverage,previous_day_deaths_covid_and_influenza,previous_day_deaths_covid_and_influenza_coverage,previous_day_deaths_influenza,previous_day_deaths_influenza_coverage,previous_week_therapeutic_a_casirivimab_imdevimab_courses_used,previous_week_therapeutic_b_bamlanivimab_courses_used,previous_week_therapeutic_c_bamlanivimab_etesevimab_courses_used,total_patients_hospitalized_confirmed_influenza,total_patients_hospitalized_confirmed_influenza_coverage,total_patients_hospitalized_confirmed_influenza_covid,total_patients_hospitalized_confirmed_influenza_covid_coverage,inpatient_beds_utilization,percent_of_inpatients_with_covid,inpatient_bed_covid_utilization,adult_icu_bed_covid_utilization,adult_icu_bed_utilization\nFL,,20230904,20200312,0,0,1,0,0,1,0,1,38,1,10,1,0,1,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,1,10,38,1,0,10,1,0,38,,,,,,,0,1,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.263157894736842,0.0,0.0,,\nFL,,20230904,20200313,0,0,1,0,0,1,0,1,38,1,10,1,0,1,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,1,10,38,1,0,10,1,0,38,,,,,,,0,1,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.263157894736842,0.0,0.0,,\nFL,,20230904,20200314,0,0,1,0,0,1,0,1,38,1,12,1,0,1,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,1,12,38,1,0,12,1,0,38,,,,,,,1,1,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.31578947368421,0.0,0.0,,\nFL,,20230904,20200315,0,0,1,0,0,1,0,1,38,1,12,1,1,1,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,1,12,38,1,1,12,1,1,38,,,,,,,0,1,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.31578947368421,0.083333333333333,0.026315789473684,,\nFL,,20230904,20200316,0,0,1,0,0,1,0,1,38,1,9,1,1,1,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,1,9,38,1,1,9,1,1,38,,,,,,,0,1,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.236842105263158,0.111111111111111,0.026315789473684,,\nFL,,20230904,20200317,0,0,1,0,0,1,0,1,38,1,9,1,1,1,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,1,9,38,1,1,9,1,1,38,,,,,,,0,1,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.236842105263158,0.111111111111111,0.026315789473684,,\nFL,,20230904,20200318,0,0,3,0,0,3,0,3,358,3,191,3,26,3,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,3,191,358,3,26,191,3,26,358,,,,,,,0,3,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.533519553072626,0.136125654450262,0.072625698324022,,\nFL,,20230904,20200319,0,0,3,0,0,3,0,3,358,3,167,3,22,3,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,3,167,358,3,22,167,3,22,358,,,,,,,0,3,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.466480446927374,0.131736526946108,0.06145251396648,,\nFL,,20230904,20200320,0,0,3,0,0,3,0,3,358,3,159,3,24,3,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,3,159,358,3,24,159,3,24,358,,,,,,,0,3,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.44413407821229,0.150943396226415,0.067039106145251,,\nFL,,20230904,20200321,0,0,3,0,0,3,0,3,358,3,150,3,24,3,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,3,150,358,3,24,150,3,24,358,,,,,,,0,3,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.418994413407821,0.16,0.067039106145251,,\nFL,,20230904,20200322,0,0,3,0,0,3,0,3,358,3,144,3,31,3,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,3,144,358,3,31,144,3,31,358,,,,,,,0,3,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.402234636871508,0.215277777777778,0.08659217877095,,\nFL,,20230904,20200323,0,0,4,0,0,4,0,3,835,4,399,4,41,4,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,4,399,835,4,41,399,4,41,835,,,,,,,0,4,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.477844311377246,0.102756892230576,0.049101796407186,,\nFL,,20230904,20200324,0,0,4,0,0,4,0,3,835,4,376,4,50,4,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,4,376,835,4,50,376,4,50,835,,,,,,,0,4,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.450299401197605,0.132978723404255,0.059880239520958,,\nFL,,20230904,20200325,0,0,4,0,0,4,0,3,835,4,372,4,49,4,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,4,372,835,4,49,372,4,49,835,,,,,,,0,4,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.445508982035928,0.131720430107527,0.058682634730539,,\nFL,,20230904,20200326,0,0,4,0,0,4,0,3,835,4,390,4,65,4,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,4,390,835,4,65,390,4,65,835,,,,,,,0,4,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.467065868263473,0.166666666666667,0.077844311377246,,\nFL,,20230904,20200327,0,0,261,0,0,261,0,3,835,4,388,4,32,3,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,4,388,835,3,32,169,3,32,358,,,,,,,0,4,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.464670658682635,0.189349112426036,0.089385474860335,,\nFL,,20230904,20200328,0,0,261,0,0,261,0,3,835,4,340,4,624,235,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,4,340,835,3,29,147,3,29,358,,,,,,,0,4,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.407185628742515,0.197278911564626,0.081005586592179,,\nFL,,20230904,20200329,0,0,261,0,0,261,0,4,1304,6,340,5,899,238,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,5,340,847,5,75,340,5,75,847,,,,,,,1,5,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.401416765053129,0.220588235294118,0.088547815820543,,\nFL,,20230904,20200330,0,0,261,0,0,261,0,16,3599,18,736,15,1002,244,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,15,736,1842,15,142,736,18,172,3599,,,,,,,0,18,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.39956568946797,0.192934782608696,0.047791053070297,,\nFL,,20230904,20200331,0,0,261,0,0,261,9,19,3863,20,1806,18,1251,247,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,18,1806,3187,18,303,1806,20,324,3863,,,,,,,0,20,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.566677125823659,0.167774086378738,0.083872637846234,,\nFL,,20230904,20200401,0,0,263,0,0,263,9,29,6344,31,3257,29,1478,252,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,29,3257,5668,29,504,3257,31,526,6344,,,,,,,2,31,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.574629498941426,0.15474362910654,0.082912988650694,,\nFL,,20230904,20200402,0,0,263,0,0,263,12,34,6947,36,3541,34,1635,254,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,34,3541,6342,33,479,3487,35,502,6865,,,,,,,3,35,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.558341217281615,0.137367364496702,0.073124544792425,,\nFL,,20230904,20200403,0,0,263,0,0,263,8,33,6352,35,2951,34,1593,255,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,34,2951,5895,33,335,2889,34,355,6270,,,,,,,4,34,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.500593723494487,0.115957078573901,0.056618819776715,,\nFL,,20230904,20200404,0,0,263,0,0,263,12,28,6941,31,2517,28,1623,255,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,28,2517,4859,26,288,2222,28,334,5301,,,,,,,3,29,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.518007820539206,0.12961296129613,0.063006979815129,,\nFL,,20230904,20200405,0,0,263,0,0,263,14,30,8482,33,4025,32,1936,260,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,32,4025,8091,30,460,3369,31,486,7319,,,,,,,5,32,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.497466320603139,0.136539032353814,0.066402514004645,,\nFL,,20230904,20200406,0,0,263,0,0,263,1,41,10725,43,5581,42,2049,261,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,42,5581,10334,41,708,5527,42,734,10643,,,,,,,5,42,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.540061931488291,0.128098425909173,0.068965517241379,,\nFL,,20230904,20200407,0,0,263,0,0,263,3,46,10876,48,6135,48,2199,261,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,48,6135,10876,47,661,6072,47,661,10794,,,,,,,5,47,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.564086061051857,0.108860342555995,0.061237724661849,,\nFL,,20230904,20200408,0,0,263,0,0,263,0,48,11035,50,6197,50,2150,261,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,50,6197,11035,49,665,6136,49,665,10953,,,,,,,11,49,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.561576801087449,0.108376792698827,0.060713959645759,,\nFL,,20230904,20200409,0,0,263,0,0,263,1,47,10596,49,5810,48,2091,261,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,48,5810,10382,47,558,5750,48,564,10514,,,,,,,4,48,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.559622423425159,0.09704347826087,0.053642762031577,,\nFL,,20230904,20200410,0,0,270,0,0,270,2,48,10929,50,6196,50,2168,262,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,50,6196,10929,49,590,6128,49,590,10847,,,,,,,5,49,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.566932015737945,0.096279373368146,0.0543929197013,,\nFL,,20230904,20200411,0,0,270,0,0,270,0,38,9480,41,4391,40,2234,262,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,40,4391,8399,39,461,4323,40,488,9398,,,,,,,3,40,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.522800333373021,0.106638908165626,0.051925941689721,,\nFL,,20230904,20200412,0,0,270,0,0,270,1,38,9551,41,4369,41,2174,261,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,40,4348,8470,39,381,4107,39,401,8992,,,,,,,3,39,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.513341204250295,0.092768444119795,0.044595195729537,,\nFL,,20230904,20200413,0,0,271,0,0,271,8,88,22001,102,11259,100,2651,268,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,96,10919,20900,100,1189,11259,102,1195,22001,,,,,,,16,90,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.52244019138756,0.105604405364597,0.05431571292214,,\nFL,,20230904,20200414,0,0,271,0,0,271,18,98,22662,106,12148,106,2622,268,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,100,11751,21530,106,1243,12148,106,1250,22662,,,,,,,165,99,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.545796562935439,0.102321369772802,0.055158414967787,,\nFL,,20230904,20200415,0,0,271,0,0,271,15,116,26625,117,14426,120,2622,263,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,111,13850,25491,120,1518,14426,117,1460,26625,,,,,,,172,117,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.543329018084814,0.105226674060724,0.054835680751174,,\nFL,,20230904,20200416,0,0,271,0,0,271,81,134,35605,135,17348,137,3508,266,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,129,17048,34471,137,2579,17348,135,2481,35605,,,,,,,167,135,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.494560645180006,0.148662670048421,0.069681224547114,,\nFL,,20230904,20200417,0,0,271,0,0,271,43,130,33453,131,16250,132,3505,266,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,125,15990,32319,132,2544,16250,131,2434,33453,,,,,,,163,131,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.494755407036109,0.156553846153846,0.072758795922638,,\nFL,,20230904,20200418,0,0,271,0,0,271,42,118,30461,119,14672,119,3125,262,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,113,13799,29327,119,1892,14672,119,1859,30461,,,,,,,161,119,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.470522044532342,0.128953107960742,0.061028856570697,,\nFL,,20230904,20200419,0,0,270,0,0,270,42,119,30474,120,14104,120,3271,263,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,114,13668,29340,120,2043,14104,120,2001,30474,,,,,,,19,119,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.465848670756646,0.144852524106636,0.065662531994487,,\nFL,,20230904,20200420,0,0,270,0,0,270,51,129,32423,131,15405,129,3369,262,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,125,15224,31289,128,2207,15383,130,2109,32387,,,,,,,30,130,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.486560772156349,0.143470064356757,0.065118720474264,,\nFL,,20230904,20200421,0,0,270,0,0,270,51,133,32383,133,15887,130,3645,267,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,127,15705,31255,130,2547,15887,133,2467,32383,,,,,,,31,133,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.502479603263478,0.160319758292944,0.076181947318037,,\nFL,,20230904,20200422,0,0,270,0,0,270,58,134,32441,134,16504,136,2773,265,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,128,15734,31312,136,1716,16504,134,1613,32441,,,,,,,45,134,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.502491057741441,0.103974793989336,0.049721032027373,,\nFL,,20230904,20200423,0,0,270,0,0,270,52,130,31286,130,17064,141,2846,262,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,129,15661,31173,141,1899,17064,130,1685,31286,,,,,,,35,130,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.502389888685722,0.111286919831224,0.053857955635108,,\nFL,,20230904,20200424,0,0,270,0,0,270,44,129,31524,129,17667,141,2782,261,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,128,15841,31411,141,1855,17667,129,1616,31524,,,,,,,40,129,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.50431377542899,0.104998018905304,0.05126253013577,,\nFL,,20230904,20200425,0,0,270,0,0,270,52,119,29809,119,15583,126,2685,259,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,118,14351,29696,126,1733,15583,119,1550,29809,,,,,,,32,119,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.483263739224138,0.111210934993262,0.051997718809755,,\nFL,,20230904,20200426,0,0,270,0,0,270,52,118,29698,118,15363,124,2781,269,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,117,14145,29585,124,1721,15363,118,1552,29698,,,,,,,28,118,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.478113909075545,0.112022391460001,0.052259411408176,,\nFL,,20230904,20200427,0,0,270,0,0,270,52,127,30551,127,16339,135,2749,270,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,126,15062,30438,135,1787,16339,127,1601,30551,,,,,,,41,126,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.494841973848479,0.109370218495624,0.052404176622696,,\nFL,,20230904,20200428,0,0,270,0,0,270,175,138,31457,131,16647,133,2948,270,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,125,15358,30326,133,1897,16647,131,1771,31457,,,,,,,39,131,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.506430125964519,0.113954466270199,0.056299074927679,,\nFL,,20230904,20200429,0,0,270,0,0,270,50,142,31595,133,16974,137,3748,269,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,127,15494,30461,137,2764,16974,133,2591,31595,,,,,,,40,133,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.50865040543646,0.162837280546718,0.082006646621301,,\nFL,,20230904,20200430,0,0,270,0,0,270,59,141,31358,132,16740,136,3814,269,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,126,15279,30224,136,2801,16740,132,2615,31358,,,,,,,45,132,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.505525410269984,0.167323775388292,0.083391797946298,,\nFL,,20230904,20200501,0,0,270,0,0,270,63,138,30744,129,16332,133,3737,270,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,123,14919,29613,133,2660,16332,129,2500,30744,,,,,,,38,129,,0,,,,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,0,,,,,0,,,0.503799007192787,0.162870438403135,0.081316679677335,,\n"
#> attr(,"class")
#> [1] "epidata_csv" "character"  
# }