fetch covidcast data

fetch covidcast data

covidcast(
  data_source,
  signals,
  time_type,
  geo_type,
  time_values,
  geo_values,
  as_of = NULL,
  issues = NULL,
  lag = NULL
)

Arguments

data_source

A character string representing a data source to query: (1) fb-survey; (2) jhu-csse; (3) google-symptoms; (4) doctor-visits; (5) quidel; (6) hhs. For more information about each data source, see https://cmu-delphi.github.io/delphi-epidata/api/covidcast_signals.html.

signals

A character string representing signals from a specific source to query. A list of available signals for each data source can be found in https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals. Each data source may have different signals.

time_type

Temporal resolution of the data. Most signals are available at time_type = "day" resolution. Some signals are only available at the time_type = "week" resolution.

geo_type

A character string that specifies geographic types: (1) county; (2) hrr, hospital referral region; (3) hhs; (4) msa, metropolitan statistical area; (5) dma, designated market areas; (6) state; (7) nation. See https://cmu-delphi.github.io/delphi-epidata/api/covidcast_geography.html for details on which types are available for each data source.

time_values

Takes in the form epirange(startdate,enddate), where startdate and enddate are of the form YYYYMMDD (can be passed as string or numeric).

geo_values

A character string that specified which geographies to return. "*" fetches all geographies. To fetch specific geographies, specify their IDs as a vector or a list of strings. More information can be found in https://cmu-delphi.github.io/delphi-epidata/api/covidcast_geography.html.

as_of

data source to fetch

issues

data source to fetch

lag

data source to fetch

Value

an instance of epidata_call

References

COVIDcast API documentation: https://cmu-delphi.github.io/delphi-epidata/api/covidcast.html

Documentation for all COVIDcast sources and signals: https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals

COVIDcast public dashboard: https://delphi.cmu.edu/covidcast/

See also

Examples

call <- covidcast(
  data_source = "jhu-csse",
  signals = "confirmed_7dav_incidence_prop",
  time_type = "day",
  geo_type = "state",
  time_values = epirange(20200601, 20200801),
  geo_values = "ca,fl"
)
fetch_csv(call)
#> [1] "geo_value,signal,source,geo_type,time_type,time_value,direction,issue,lag,missing_value,missing_stderr,missing_sample_size,value,stderr,sample_size\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200601,,20230310,1012,0,5,5,6.8431078,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200601,,20230303,1005,0,5,5,3.3418087,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200602,,20230310,1011,0,5,5,6.8256897,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200602,,20230303,1004,0,5,5,3.4127991,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200603,,20230310,1010,0,5,5,6.6649358,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200603,,20230303,1003,0,5,5,4.0293642,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200604,,20230310,1009,0,5,5,6.9806376,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200604,,20230303,1002,0,5,5,4.534185,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200605,,20230310,1008,0,5,5,6.976646,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200605,,20230303,1001,0,5,5,4.5953156,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200606,,20230310,1007,0,5,5,6.6689275,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200606,,20230303,1000,0,5,5,4.820776,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200607,,20230310,1006,0,5,5,6.7422283,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200607,,20230303,999,0,5,5,5.1106536,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200608,,20230310,1005,0,5,5,6.6765479,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200608,,20230303,998,0,5,5,5.3071919,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200609,,20230310,1004,0,5,5,6.8184322,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200609,,20230303,997,0,5,5,5.6220476,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200610,,20230310,1003,0,5,5,7.1373999,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200610,,20230303,996,0,5,5,5.6575428,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200611,,20230310,1002,0,5,5,7.1301423,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200611,,20230303,995,0,5,5,5.8409347,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200612,,20230310,1001,0,5,5,7.3743576,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200612,,20230303,994,0,5,5,6.2333541,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200613,,20230310,1000,0,5,5,7.5351115,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200613,,20230303,993,0,5,5,7.095099,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200614,,20230310,999,0,5,5,7.5405546,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200614,,20230303,992,0,5,5,7.6446175,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200615,,20230310,998,0,5,5,7.5623272,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200615,,20230303,991,0,5,5,8.165214,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200616,,20230310,997,0,5,5,7.9081476,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200616,,20230303,990,0,5,5,9.2741108,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200617,,20230310,996,0,5,5,8.4339543,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200617,,20230303,989,0,5,5,10.0885287,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200618,,20230310,995,0,5,5,9.0330619,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200618,,20230303,988,0,5,5,11.0804228,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200619,,20230310,994,0,5,5,9.1869212,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200619,,20230303,987,0,5,5,12.3424749,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200620,,20230310,993,0,5,5,9.7138165,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200620,,20230303,986,0,5,5,13.3074188,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200621,,20230310,992,0,5,5,10.304578,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200621,,20230303,985,0,5,5,14.278936,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200622,,20230310,991,0,5,5,11.6109301,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200622,,20230303,984,0,5,5,15.0466844,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200623,,20230310,990,0,5,5,13.1096062,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200623,,20230303,983,0,5,5,15.3773157,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200624,,20230310,989,0,5,5,13.5134867,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200624,,20230303,982,0,5,5,17.2841975,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200625,,20230310,988,0,5,5,13.8106818,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200625,,20230303,981,0,5,5,18.4653994,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200626,,20230310,987,0,5,5,14.6409412,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200626,,20230303,980,0,5,5,21.8308716,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200627,,20230310,986,0,5,5,14.4903478,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200627,,20230303,979,0,5,5,25.4697884,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200628,,20230310,985,0,5,5,15.2110187,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200628,,20230303,978,0,5,5,28.7800458,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200629,,20230310,984,0,5,5,15.6609844,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200629,,20230303,977,0,5,5,30.3181717,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200630,,20230310,983,0,5,5,16.4633024,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200630,,20230303,976,0,5,5,32.1632656,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200701,,20230310,982,0,5,5,17.0747477,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200701,,20230303,975,0,5,5,32.8547649,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200702,,20230310,981,0,5,5,17.9587126,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200702,,20230303,974,0,5,5,36.2103773,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200703,,20230310,980,0,5,5,17.5682585,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200703,,20230303,973,0,5,5,36.5692734,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200704,,20230310,979,0,5,5,17.9601641,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200704,,20230303,972,0,5,5,37.8004315,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200705,,20230310,978,0,5,5,19.7505923,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200705,,20230303,971,0,5,5,38.8054719,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200706,,20230303,970,0,5,5,19.4690008,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200706,,20230303,970,0,5,5,39.508803,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200707,,20230310,976,0,5,5,19.8456657,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200707,,20230303,969,0,5,5,40.3330808,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200708,,20230310,975,0,5,5,20.0379897,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200708,,20230303,968,0,5,5,42.5850549,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200709,,20230310,974,0,5,5,20.1258056,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200709,,20230303,967,0,5,5,41.8133627,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200710,,20230310,973,0,5,5,21.1770562,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200710,,20230303,966,0,5,5,43.0918477,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200711,,20230310,972,0,5,5,22.2261295,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200711,,20230303,965,0,5,5,42.3701117,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200712,,20230310,971,0,5,5,20.9524362,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200712,,20230303,964,0,5,5,45.8151195,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200713,,20230310,970,0,5,5,21.4898549,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200713,,20230303,963,0,5,5,49.94834,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200714,,20230310,969,0,5,5,21.4735255,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200714,,20230303,962,0,5,5,51.1624078,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200715,,20230310,968,0,5,5,21.7195552,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200715,,20230303,961,0,5,5,51.288613,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200716,,20230310,967,0,5,5,22.3778115,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200716,,20230303,960,0,5,5,54.5949265,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200717,,20230310,966,0,5,5,22.5897308,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200717,,20230303,959,0,5,5,54.616618,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200718,,20230310,965,0,5,5,22.7004079,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200718,,20230303,958,0,5,5,54.5955838,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200719,,20230310,964,0,5,5,22.3886978,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200719,,20230303,957,0,5,5,52.7406302,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200720,,20230310,963,0,5,5,22.9493405,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200720,,20230303,956,0,5,5,51.2439153,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200721,,20230310,962,0,5,5,23.3365288,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200721,,20230303,955,0,5,5,51.4056158,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200722,,20230310,961,0,5,5,24.4411221,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200722,,20230303,954,0,5,5,51.1453175,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200723,,20230310,960,0,5,5,24.0423218,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200723,,20230303,953,0,5,5,48.7027209,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200724,,20230310,959,0,5,5,24.4549113,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200724,,20230303,952,0,5,5,49.3455787,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200725,,20230310,958,0,5,5,24.269119,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200725,,20230303,951,0,5,5,50.5754221,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200726,,20230310,957,0,5,5,24.2408147,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200726,,20230303,950,0,5,5,48.515385,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200727,,20230310,956,0,5,5,23.0520343,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200727,,20230303,949,0,5,5,47.5589862,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200728,,20230310,955,0,5,5,22.2696746,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200728,,20230303,948,0,5,5,47.4209493,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200729,,20230310,954,0,5,5,21.9746567,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200729,,20230303,947,0,5,5,47.1981182,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200730,,20230310,953,0,5,5,21.7986621,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200730,,20230303,946,0,5,5,47.0055238,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200731,,20230310,952,0,5,5,20.8366789,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200731,,20230303,945,0,5,5,44.7463191,,\nca,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200801,,20230310,951,0,5,5,21.0417036,,\nfl,confirmed_7dav_incidence_prop,jhu-csse,state,day,20200801,,20230303,944,0,5,5,43.065555,,\n"
#> attr(,"class")
#> [1] "epidata_csv" "character"