Fetch ECDC data

Obtain information on influenza-like-illness from the European Centre for Disease Prevention and Control.

ecdc_ili(regions, epiweeks, issues = NULL, lag = NULL)

Arguments

regions

character vector. The regions to fetch.

epiweeks

epirange. The epiweeks to fetch.

issues

epirange. Optionally, the issues to fetch. If not set, the most recent issue is returned. Mutually exclusive with lag.

lag

integer. Optionally, the lag of the issues to fetch. If not set, the most recent issue is returned. Mutually exclusive with issues.

Value

epidata_call

Details

API docs: https://cmu-delphi.github.io/delphi-epidata/api/ecdc_ili.html.

The list of location argument can be found in https://github.com/cmu-delphi/delphi-epidata/blob/main/labels/ecdc_regions.txt.

Examples

# \donttest{
call <- ecdc_ili(regions = "austria", epiweeks = epirange(201201, 202001))
fetch_classic(call)
#> $epidata
#>    release_date  region issue    epiweek lag incidence_rate
#> 1          <NA> Austria  <NA> 2018-09-30  76       439.5328
#> 2          <NA> Austria  <NA> 2018-10-07  75       439.4062
#> 3          <NA> Austria  <NA> 2018-10-14  74       602.6987
#> 4          <NA> Austria  <NA> 2018-10-21  73       459.8340
#> 5          <NA> Austria  <NA> 2018-10-28  72       490.6953
#> 6          <NA> Austria  <NA> 2018-11-04  71       654.2370
#> 7          <NA> Austria  <NA> 2018-11-11  70       697.1184
#> 8          <NA> Austria  <NA> 2018-11-18  69       763.7961
#> 9          <NA> Austria  <NA> 2018-11-25  68       707.1851
#> 10         <NA> Austria  <NA> 2018-12-02  67       654.4775
#> 11         <NA> Austria  <NA> 2018-12-09  66       686.2745
#> 12         <NA> Austria  <NA> 2018-12-16  65       772.0588
#> 13         <NA> Austria  <NA> 2019-01-06  62       786.8775
#> 14         <NA> Austria  <NA> 2019-01-13  61       855.3868
#> 15         <NA> Austria  <NA> 2019-01-20  60      1022.3642
#> 16         <NA> Austria  <NA> 2019-01-27  59      1264.8444
#> 17         <NA> Austria  <NA> 2019-02-03  58      1367.0933
#> 18         <NA> Austria  <NA> 2019-02-10  57      1209.1503
#> 19         <NA> Austria  <NA> 2019-02-17  56      1156.0458
#> 20         <NA> Austria  <NA> 2019-02-24  55      1122.2999
#> 21         <NA> Austria  <NA> 2019-03-03  54       947.8168
#> 22         <NA> Austria  <NA> 2019-03-10  53       702.8754
#> 23         <NA> Austria  <NA> 2019-03-17  52       618.2950
#> 24         <NA> Austria  <NA> 2019-03-24  51       560.7015
#> 25         <NA> Austria  <NA> 2019-03-31  50       352.5394
#> 26         <NA> Austria  <NA> 2019-04-07  49         0.0000
#> 27         <NA> Austria  <NA> 2019-09-29  24       603.4575
#> 28         <NA> Austria  <NA> 2019-10-06  23       667.9469
#> 29         <NA> Austria  <NA> 2019-10-13  22       742.0457
#> 30         <NA> Austria  <NA> 2019-10-20  21       673.1721
#> 31         <NA> Austria  <NA> 2019-10-27  20       560.8205
#> 32         <NA> Austria  <NA> 2019-11-03  19       466.8708
#> 33         <NA> Austria  <NA> 2019-11-10  18       755.2870
#> 34         <NA> Austria  <NA> 2019-11-17  17       777.2590
#> 35         <NA> Austria  <NA> 2019-11-24  16       837.2336
#> 36         <NA> Austria  <NA> 2019-12-01  15       930.4386
#> 37         <NA> Austria  <NA> 2019-12-08  14       967.4829
#> 38         <NA> Austria  <NA> 2019-12-15  13      1157.3962
#> 39         <NA> Austria  <NA> 2019-12-22  12       504.0990
#> 40         <NA> Austria  <NA> 2019-12-29  11       406.6063
#> 
#> $result
#> [1] 1
#> 
#> $message
#> [1] "success"
#> 
# }