epidatr: Client for Delphi's 'Epidata' API

The Delphi 'Epidata' API provides real-time access to epidemiological surveillance data for influenza, 'COVID-19', and other diseases for the USA at various geographical resolutions, both from official government sources such as the Center for Disease Control (CDC) and Google Trends and private partners such as Facebook and Change 'Healthcare'. It is built and maintained by the Carnegie Mellon University Delphi research group. To cite this API: David C. Farrow, Logan C. Brooks, Aaron 'Rumack', Ryan J. 'Tibshirani', 'Roni' 'Rosenfeld' (2015). Delphi 'Epidata' API. https://github.com/cmu-delphi/delphi-epidata.

Package options

The delphi.epidata.key option specifies the API key to be used when making requests to the Epidata API.

See also

Author

Maintainer: David Weber davidweb@andrew.cmu.edu [contributor]

Authors:

Logan Brooks lcbrooks@andrew.cmu.edu

Dmitry Shemetov dshemeto@andrew.cmu.edu

Samuel Gratzl sam@sgratzl.com

Other contributors:

Nat DeFries [contributor]

Alex Reinhart [contributor]

Daniel J. McDonald [contributor]

Kean Ming Tan [contributor]

Will Townes [contributor]

George Haff [contributor]

Kathryn Mazaitis [contributor]