Helper: Extract Data From Epx-File

A helper function of read.EpiData, to extract the xml-like data from an epx-file.

epx.extract(x)

Arguments

x

An EpiData epx-file

Value

A list, containing the following elements:

$epx:

The entire xml_document from the epx-file.

$infoEpiData:

Information about EpiDatas set up.

$infoStudy

Information about the study.

$infoSeparators

Used separators.

$infoDataSets

Summary information about the data sets.

$infoParentDataSet

Information about parent data sets in relational data bases.

$perDataSet

Information from each data set.

$infoKeyVars

Key variables in relational data bases.

Examples

# This demonstrates the effects of all six helper functions to \code{\link{read.EpiData}}.

(x <- epx.example("SomeFakeData.epx"))
#> [1] "/home/runner/work/_temp/Library/epidatR/extdata/SomeFakeData.epx"

# epx.extract() extracts the information from an EpiData-epx file
(info <- epx.extract(x))
#> $epx
#> {xml_document}
#> <EpiData schemaLocation="http://www.epidata.dk/XML/2.1 http://www.epidata.dk/XML/2.1/epx.xsd" version="6" lang="en" dateSeparator="/" decimalSeparator="," timeSeparator="." cycle="2" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
#> [1] <Settings dateSeparator="/" timeSeparator="." decimalSeparator=","/>
#> [2] <ProjectSettings autoIncStart="1" timedBackupInterval="10" backupOnShutdo ...
#> [3] <StudyInfo>\n  <Abstract>\n    <Text xml:lang="en">Anhand dieses Beispiel ...
#> [4] <ValueLabelSets>\n  <ValueLabelSet id="_miss" type="ftInteger" scope="vls ...
#> [5] <DataFiles>\n  <DataFile id="ds1" created="2019/05/21 20.20.55" modified= ...
#> [6] <DataFileRelations>\n  <DataFileRelation id="relation_id_1" dataFileRef=" ...
#> 
#> $infoEpiData
#> $infoEpiData[[1]]
#>                                                        schemaLocation 
#> "http://www.epidata.dk/XML/2.1 http://www.epidata.dk/XML/2.1/epx.xsd" 
#>                                                               version 
#>                                                                   "6" 
#>                                                                  lang 
#>                                                                  "en" 
#>                                                         dateSeparator 
#>                                                                   "/" 
#>                                                      decimalSeparator 
#>                                                                   "," 
#>                                                         timeSeparator 
#>                                                                   "." 
#>                                                                 cycle 
#>                                                                   "2" 
#>                                                             xmlns:xsi 
#>                           "http://www.w3.org/2001/XMLSchema-instance" 
#> 
#> 
#> $infoStudy
#> {xml_nodeset (1)}
#> [1] <StudyInfo>\n  <Abstract>\n    <Text xml:lang="en">Anhand dieses Beispiel ...
#> 
#> $infoSeparators
#>    dateSeparator    timeSeparator decimalSeparator 
#>              "/"              "."              "," 
#> 
#> $infoDataSets
#> {xml_nodeset (1)}
#> [1] <DataFile id="ds1" created="2019/05/21 20.20.55" modified="2023/11/27 19. ...
#> 
#> $infoRelations
#>   data.set    labels offspring.data.sets key
#> 1      ds1 Dataset 1                        
#> 
#> $perDataSet
#> $perDataSet$ds1
#> $perDataSet$ds1$records
#> [1] "ID=1;date=03/11/2023;age=65;gender=2;height=189;weight=78;notes=\"No notes\""
#> [2] "ID=2;date=10/11/2023;age=-1;gender=-1;height=160;weight=52"                  
#> [3] "ID=3;date=30/10/2023;age=18;gender=1;height=-1;weight=60"                    
#> [4] "ID=4;age=39;gender=2;height=170;weight=81;notes=\"This is fake data\""       
#> 
#> $perDataSet$ds1$recordStatus
#> [1] "Verified" "Deleted"  "Normal"   "Normal"  
#> 
#> $perDataSet$ds1$datFields
#> {xml_nodeset (7)}
#> [1] <Field id="ID" top="56" left="134" type="ftAutoInc" length="6" decimals=" ...
#> [2] <Field id="date" top="88" left="134" type="ftDMYDate" length="10" decimal ...
#> [3] <Field id="age" top="122" left="134" type="ftInteger" length="3" decimals ...
#> [4] <Field id="gender" top="155" left="134" type="ftInteger" length="2" decim ...
#> [5] <Field id="height" top="188" left="134" type="ftInteger" length="3" decim ...
#> [6] <Field id="weight" top="221" left="134" type="ftFloat" length="8" decimal ...
#> [7] <Field id="notes" top="274" left="134" type="ftMemo" length="20" decimals ...
#> 
#> $perDataSet$ds1$fieldNames
#> [1] "ID"     "date"   "age"    "gender" "height" "weight" "notes" 
#> 
#> $perDataSet$ds1$fieldLabels
#> [1] "Index"       "Date"        "Age (years)" "Gender"      "Height (cm)"
#> [6] "Weight (kg)" "Notes"      
#> 
#> $perDataSet$ds1$fieldTypes
#> [1] "ftAutoInc" "ftDMYDate" "ftInteger" "ftInteger" "ftInteger" "ftFloat"  
#> [7] "ftMemo"   
#> 
#> $perDataSet$ds1$fieldValLabSets
#> [1] NA        NA        "_miss"   "_gender" "_miss"   "_miss"   NA       
#> 
#> $perDataSet$ds1$valLabelSets
#> {xml_nodeset (3)}
#> [1] <ValueLabelSet id="_miss" type="ftInteger" scope="vlsInternal">\n  <Value ...
#> [2] <ValueLabelSet id="_gender" type="ftInteger" scope="vlsInternal">\n  <Val ...
#> [3] <ValueLabelSet id="_gebDatum" type="ftDMYDate" scope="vlsInternal">\n  <V ...
#> 
#> $perDataSet$ds1$valLabels
#> $perDataSet$ds1$valLabels[[1]]
#> {xml_nodeset (1)}
#> [1] <ValueLabel missing="true" order="2" value="-1">\n  <Label xml:lang="en"> ...
#> 
#> $perDataSet$ds1$valLabels[[2]]
#> {xml_nodeset (4)}
#> [1] <ValueLabel order="2" value="1">\n  <Label xml:lang="en">female</Label>\n ...
#> [2] <ValueLabel order="3" value="2">\n  <Label xml:lang="en">male</Label>\n</ ...
#> [3] <ValueLabel order="4" value="3">\n  <Label xml:lang="en">other</Label>\n< ...
#> [4] <ValueLabel missing="true" order="5" value="-1">\n  <Label xml:lang="en"> ...
#> 
#> $perDataSet$ds1$valLabels[[3]]
#> {xml_nodeset (1)}
#> [1] <ValueLabel missing="true" order="2" value="-287861">\n  <Label xml:lang= ...
#> 
#> 
#> $perDataSet$ds1$infoSeparators
#>    dateSeparator    timeSeparator decimalSeparator 
#>              "/"              "."              "," 
#> 
#> 
#> 

# epx.read() transforms the information into a data.frame
(dat <- lapply(info$perDataSet, epx.read))
#> $ds1
#>   ID       date age gender height weight             notes
#> 1  1 03/11/2023  65      2    189     78          No notes
#> 2  2 10/11/2023  -1     -1    160     52                NA
#> 3  3 30/10/2023  18      1     -1     60                NA
#> 4  4         NA  39      2    170     81 This is fake data
#> 
# Combine dat and info in a data-set-wise list
# This is a necessary detour to handle EpiData files with multiple data sets.
perDataSet <- mapply(function(dat, info) list(list(dat = dat, info = info)), dat, info[[7]])

# epx.labels() replaces value codes with value labels
(datLab <- epx.labels(perDataSet$ds1$dat, perDataSet$ds1$info))
#>   ID       date     age  gender  height weight             notes
#> 1  1 03/11/2023      65    male     189     78          No notes
#> 2  2 10/11/2023 no data no data     160     52                NA
#> 3  3 30/10/2023      18  female no data     60                NA
#> 4  4         NA      39    male     170     81 This is fake data

# epx.missing() replaces definde missing values with NA
(datMis <- epx.missing(perDataSet$ds1$dat, perDataSet$ds1$info))
#> Warning: NAs introduced by coercion
#>   ID       date  age gender height weight             notes
#> 1  1 03/11/2023   65      2    189     78          No notes
#> 2  2 10/11/2023 <NA>   <NA>    160     52              <NA>
#> 3  3 30/10/2023   18      1   <NA>     60              <NA>
#> 4  4       <NA>   39      2    170     81 This is fake data

# epx.class() sets variable classes according to the field types defined in EpiData
(datClass <- epx.class(perDataSet$ds1$dat, perDataSet$ds$info))
#>   ID       date age gender height weight             notes
#> 1  1 2023-11-03  65      2    189     78          No notes
#> 2  2 2023-11-10  -1     -1    160     52                NA
#> 3  3 2023-10-30  18      1     -1     60                NA
#> 4  4       <NA>  39      2    170     81 This is fake data

# Bringing it all together
read.EpiData(epx.example("SomeFakeData.epx"))
#>     status ID       date age gender height weight             notes
#> 1 Verified  1 2023-11-03  65   male    189     78          No notes
#> 2  Deleted  2 2023-11-10  NA   <NA>    160     52              <NA>
#> 3   Normal  3 2023-10-30  18 female     NA     60              <NA>
#> 4   Normal  4       <NA>  39   male    170     81 This is fake data