Fetch FluSurv hospitalization data

Obtain information on flu hospitalization rates from the Center of Disease Control.

flusurv(locations, epiweeks, issues = NULL, lag = NULL)

Arguments

locations

character vector. Character strings indicating location.

epiweeks

epirange. The epiweeks to fetch.

issues

epirange. Optionally, the issues to fetch. If not set, the most recent issue is returned. Mutually exclusive with lag.

lag

integer. Optionally, the lag of the issues to fetch. If not set, the most recent issue is returned. Mutually exclusive with issues.

Value

epidata_call

Details

API docs: https://cmu-delphi.github.io/delphi-epidata/api/flusurv.html. See also https://gis.cdc.gov/GRASP/Fluview/FluHospRates.html.

The list of location argument can be found in https://github.com/cmu-delphi/delphi-epidata/blob/main/labels/flusurv_locations.txt.

Examples

# \donttest{
call <- flusurv(locations = "CA", epiweeks = epirange(201201, 202001))
fetch_classic(call)
#> $epidata
#>     release_date location issue    epiweek lag rate_age_0 rate_age_1 rate_age_2
#> 1     2017-05-19       CA  <NA> 2012-01-01 279   0.000000   0.000000   0.000000
#> 2     2017-05-19       CA  <NA> 2012-01-08 278   0.000000   0.000000   0.000000
#> 3     2017-05-19       CA  <NA> 2012-01-15 277   0.497966   0.000000   0.000000
#> 4     2017-05-19       CA  <NA> 2012-01-22 276   0.000000   0.000000   0.061794
#> 5     2017-05-19       CA  <NA> 2012-01-29 275   0.000000   0.196172   0.247176
#> 6     2017-05-19       CA  <NA> 2012-02-05 274   1.493897   0.196172   0.000000
#> 7     2017-05-19       CA  <NA> 2012-02-12 273   0.497966   0.000000   0.308969
#> 8     2017-05-19       CA  <NA> 2012-02-19 272   0.995932   0.196172   0.247176
#> 9     2017-05-19       CA  <NA> 2012-02-26 271   0.995932   0.392343   0.308969
#> 10    2017-05-19       CA  <NA> 2012-03-04 270   1.991863   0.588515   0.185382
#> 11    2017-05-19       CA  <NA> 2012-03-11 269   1.493897   0.392343   0.988702
#> 12    2017-05-19       CA  <NA> 2012-03-18 268   2.489829   0.980858   0.556145
#> 13    2017-05-19       CA  <NA> 2012-03-25 267   2.987795   0.980858   0.617939
#> 14    2017-05-19       CA  <NA> 2012-04-01 266   0.000000   0.392343   0.432557
#> 15    2017-05-19       CA  <NA> 2012-04-08 265   1.493897   0.392343   0.308969
#> 16    2017-05-19       CA  <NA> 2012-04-15 264   0.497966   0.588515   0.185382
#> 17    2017-05-19       CA  <NA> 2012-04-22 263   0.995932   0.196172   0.185382
#> 18    2017-05-19       CA  <NA> 2012-09-30 240   0.000000   0.195165   0.000000
#> 19    2017-05-19       CA  <NA> 2012-10-07 239   0.000000   0.000000   0.000000
#> 20    2017-05-19       CA  <NA> 2012-10-14 238   0.000000   0.000000   0.000000
#> 21    2017-05-19       CA  <NA> 2012-10-21 237   0.000000   0.000000   0.000000
#> 22    2017-05-19       CA  <NA> 2012-10-28 236   0.499683   0.000000   0.000000
#> 23    2017-05-19       CA  <NA> 2012-11-04 235   0.000000   0.000000   0.000000
#> 24    2017-05-19       CA  <NA> 2012-11-11 234   0.000000   0.000000   0.000000
#> 25    2017-05-19       CA  <NA> 2012-11-18 233   0.000000   0.000000   0.000000
#> 26    2017-05-19       CA  <NA> 2012-11-25 232   0.000000   0.195165   0.000000
#> 27    2017-05-19       CA  <NA> 2012-12-02 231   0.000000   0.195165   0.000000
#> 28    2017-05-19       CA  <NA> 2012-12-09 230   0.000000   0.000000   0.061117
#> 29    2017-05-19       CA  <NA> 2012-12-16 229   0.999365   0.000000   0.000000
#> 30    2017-05-19       CA  <NA> 2012-12-23 228   0.999365   0.000000   0.122234
#> 31    2017-05-19       CA  <NA> 2012-12-30 227   0.999365   0.000000   0.611171
#> 32    2017-05-19       CA  <NA> 2013-01-06 226   2.998096   0.390329   0.733406
#> 33    2017-05-19       CA  <NA> 2013-01-13 225   4.497144   0.780658   1.283460
#> 34    2017-05-19       CA  <NA> 2013-01-20 224   6.495875   0.975823   2.077983
#> 35    2017-05-19       CA  <NA> 2013-01-27 223   5.496510   1.170988   1.405694
#> 36    2017-05-19       CA  <NA> 2013-02-03 222   3.497779   0.780658   1.405694
#> 37    2017-05-19       CA  <NA> 2013-02-10 221   2.498414   0.390329   0.916757
#> 38    2017-05-19       CA  <NA> 2013-02-17 220   1.998731   0.585494   0.855640
#> 39    2017-05-19       CA  <NA> 2013-02-24 219   2.498414   0.000000   0.427820
#> 40    2017-05-19       CA  <NA> 2013-03-03 218   0.000000   0.195165   0.488937
#> 41    2017-05-19       CA  <NA> 2013-03-10 217   0.499683   0.000000   0.305586
#> 42    2017-05-19       CA  <NA> 2013-03-17 216   3.497779   0.390329   0.061117
#> 43    2017-05-19       CA  <NA> 2013-03-24 215   1.499048   0.585494   0.122234
#> 44    2017-05-19       CA  <NA> 2013-03-31 214   0.499683   0.390329   0.122234
#> 45    2017-05-19       CA  <NA> 2013-04-07 213   0.000000   0.195165   0.061117
#> 46    2017-05-19       CA  <NA> 2013-04-14 212   0.999365   0.000000   0.061117
#> 47    2017-05-19       CA  <NA> 2013-04-21 211   0.499683   0.195165   0.061117
#> 48    2017-05-19       CA  <NA> 2013-09-29 188   0.998597   0.000000   0.000000
#> 49    2017-05-19       CA  <NA> 2013-10-06 187   0.000000   0.000000   0.000000
#> 50    2017-05-19       CA  <NA> 2013-10-13 186   0.499298   0.000000   0.000000
#> 51    2017-05-19       CA  <NA> 2013-10-20 185   0.000000   0.000000   0.000000
#> 52    2017-05-19       CA  <NA> 2013-10-27 184   0.499298   0.000000   0.000000
#> 53    2017-05-19       CA  <NA> 2013-11-03 183   0.499298   0.000000   0.000000
#> 54    2017-05-19       CA  <NA> 2013-11-10 182   0.499298   0.000000   0.060372
#> 55    2017-05-19       CA  <NA> 2013-11-17 181   0.000000   0.000000   0.060372
#> 56    2017-05-19       CA  <NA> 2013-11-24 180   0.998597   0.000000   0.060372
#> 57    2017-05-19       CA  <NA> 2013-12-01 179   0.998597   0.000000   0.181115
#> 58    2017-05-19       CA  <NA> 2013-12-08 178   0.998597   0.000000   0.362230
#> 59    2017-05-19       CA  <NA> 2013-12-15 177   0.499298   0.000000   0.724460
#> 60    2017-05-19       CA  <NA> 2013-12-22 176   1.997194   0.775532   1.086689
#> 61    2017-05-19       CA  <NA> 2013-12-29 175   2.995791   0.969415   2.294122
#> 62    2017-05-19       CA  <NA> 2014-01-05 174   4.992985   0.581649   2.897839
#> 63    2017-05-19       CA  <NA> 2014-01-12 173   4.493686   0.581649   2.233751
#> 64    2017-05-19       CA  <NA> 2014-01-19 172   4.493686   1.163298   1.569663
#> 65    2017-05-19       CA  <NA> 2014-01-26 171   1.497895   1.163298   1.086689
#> 66    2017-05-19       CA  <NA> 2014-02-02 170   3.495089   0.387766   0.724460
#> 67    2017-05-19       CA  <NA> 2014-02-09 169   0.000000   0.000000   0.181115
#> 68    2017-05-19       CA  <NA> 2014-02-16 168   1.497895   0.000000   0.000000
#> 69    2017-05-19       CA  <NA> 2014-02-23 167   0.998597   0.000000   0.181115
#> 70    2017-05-19       CA  <NA> 2014-03-02 166   1.997194   0.193883   0.000000
#> 71    2017-05-19       CA  <NA> 2014-03-09 165   0.000000   0.193883   0.000000
#> 72    2017-05-19       CA  <NA> 2014-03-16 164   0.499298   0.000000   0.060372
#> 73    2017-05-19       CA  <NA> 2014-03-23 163   0.998597   0.000000   0.000000
#> 74    2017-05-19       CA  <NA> 2014-03-30 162   0.499298   0.000000   0.000000
#> 75    2017-05-19       CA  <NA> 2014-04-06 161   0.499298   0.000000   0.000000
#> 76    2017-05-19       CA  <NA> 2014-04-13 160   0.499298   0.000000   0.000000
#> 77    2017-05-19       CA  <NA> 2014-04-20 159   1.497895   0.000000   0.000000
#> 78    2017-05-19       CA  <NA> 2014-09-28 136   0.496367   0.000000   0.000000
#> 79    2017-05-19       CA  <NA> 2014-10-05 135   0.000000   0.000000   0.000000
#> 80    2017-05-19       CA  <NA> 2014-10-12 134   0.496367   0.000000   0.000000
#> 81    2017-05-19       CA  <NA> 2014-10-19 133   0.496367   0.000000   0.000000
#> 82    2017-05-19       CA  <NA> 2014-10-26 132   0.000000   0.000000   0.000000
#> 83    2017-05-19       CA  <NA> 2014-11-02 131   0.496367   0.000000   0.000000
#> 84    2017-05-19       CA  <NA> 2014-11-09 130   0.000000   0.000000   0.000000
#> 85    2017-05-19       CA  <NA> 2014-11-16 129   0.000000   0.000000   0.000000
#> 86    2017-05-19       CA  <NA> 2014-11-23 128   0.000000   0.192967   0.059534
#> 87    2017-05-19       CA  <NA> 2014-11-30 127   0.992733   0.000000   0.059534
#> 88    2017-05-19       CA  <NA> 2014-12-07 126   0.496367   0.000000   0.059534
#> 89    2017-05-19       CA  <NA> 2014-12-14 125   1.489100   0.000000   0.059534
#> 90    2017-05-19       CA  <NA> 2014-12-21 124   0.496367   0.771870   0.357205
#> 91    2017-05-19       CA  <NA> 2014-12-28 123   1.985466   0.578902   0.595342
#> 92    2017-05-19       CA  <NA> 2015-01-04 122   4.467299   1.736707   1.012081
#> 93    2017-05-19       CA  <NA> 2015-01-11 121   3.474566   1.543740   1.131149
#> 94    2017-05-19       CA  <NA> 2015-01-18 120   4.963666   1.157805   0.833478
#> 95    2017-05-19       CA  <NA> 2015-01-25 119   6.949132   1.350772   1.369286
#> 96    2017-05-19       CA  <NA> 2015-02-01 118   2.481833   1.543740   1.250217
#> 97    2017-05-19       CA  <NA> 2015-02-08 117   5.956399   0.578902   1.012081
#> 98    2017-05-19       CA  <NA> 2015-02-15 116   3.474566   0.385935   0.476273
#> 99    2017-05-19       CA  <NA> 2015-02-22 115   1.985466   0.771870   0.238137
#> 100   2017-05-19       CA  <NA> 2015-03-01 114   0.496367   0.192967   0.178602
#> 101   2017-05-19       CA  <NA> 2015-03-08 113   1.985466   0.192967   0.297671
#> 102   2017-05-19       CA  <NA> 2015-03-15 112   0.000000   0.000000   0.000000
#> 103   2017-05-19       CA  <NA> 2015-03-22 111   0.496367   0.192967   0.178602
#> 104   2017-05-19       CA  <NA> 2015-03-29 110   1.489100   0.000000   0.059534
#> 105   2017-05-19       CA  <NA> 2015-04-05 109   0.496367   0.000000   0.238137
#> 106   2017-05-19       CA  <NA> 2015-04-12 108   1.489100   0.000000   0.000000
#> 107   2017-05-19       CA  <NA> 2015-04-19 107   0.000000   0.000000   0.178602
#> 108   2017-05-19       CA  <NA> 2015-04-26 106   0.000000   0.385935   0.178602
#> 109   2017-05-19       CA  <NA> 2015-10-04  83   0.491969   0.000000   0.058742
#> 110   2017-05-19       CA  <NA> 2015-10-11  82   0.491969   0.000000   0.058742
#> 111   2017-05-19       CA  <NA> 2015-10-18  81   0.000000   0.192189   0.000000
#> 112   2017-05-19       CA  <NA> 2015-10-25  80   0.000000   0.000000   0.000000
#> 113   2017-05-19       CA  <NA> 2015-11-01  79   0.000000   0.000000   0.000000
#> 114   2017-05-19       CA  <NA> 2015-11-08  78   0.000000   0.000000   0.058742
#> 115   2017-05-19       CA  <NA> 2015-11-15  77   0.000000   0.000000   0.000000
#> 116   2017-05-19       CA  <NA> 2015-11-22  76   0.491969   0.000000   0.058742
#> 117   2017-05-19       CA  <NA> 2015-11-29  75   0.000000   0.384379   0.000000
#> 118   2017-05-19       CA  <NA> 2015-12-06  74   0.491969   0.192189   0.058742
#> 119   2017-05-19       CA  <NA> 2015-12-13  73   0.491969   0.192189   0.000000
#> 120   2017-05-19       CA  <NA> 2015-12-20  72   1.475906   0.192189   0.058742
#> 121   2017-05-19       CA  <NA> 2015-12-27  71   0.000000   0.000000   0.176226
#> 122   2017-05-19       CA  <NA> 2016-01-03  70   1.475906   0.192189   0.293710
#> 123   2017-05-19       CA  <NA> 2016-01-10  69   1.475906   0.384379   0.411194
#> 124   2017-05-19       CA  <NA> 2016-01-17  68   0.491969   0.768758   0.176226
#> 125   2017-05-19       CA  <NA> 2016-01-24  67   1.967874   0.000000   0.411194
#> 126   2017-05-19       CA  <NA> 2016-01-31  66   2.951812   0.960947   0.293710
#> 127   2017-05-19       CA  <NA> 2016-02-07  65   1.967874   0.768758   1.057356
#> 128   2017-05-19       CA  <NA> 2016-02-14  64   3.443780   1.153137   0.763646
#> 129   2017-05-19       CA  <NA> 2016-02-21  63   2.951812   1.153137   0.822388
#> 130   2017-05-19       CA  <NA> 2016-02-28  62   2.459843   0.768758   0.646162
#> 131   2017-05-19       CA  <NA> 2016-03-06  61   3.443780   0.960947   0.998614
#> 132   2017-05-19       CA  <NA> 2016-03-13  60   2.459843   1.153137   1.116098
#> 133   2017-05-19       CA  <NA> 2016-03-20  59   3.935749   0.576568   0.646162
#> 134   2017-05-19       CA  <NA> 2016-03-27  58   0.491969   0.000000   0.587420
#> 135   2017-05-19       CA  <NA> 2016-04-03  57   1.475906   0.384379   0.293710
#> 136   2017-05-19       CA  <NA> 2016-04-10  56   0.491969   0.384379   0.234968
#> 137   2017-05-19       CA  <NA> 2016-04-17  55   0.983937   0.384379   0.117484
#> 138   2017-05-19       CA  <NA> 2016-04-24  54   0.983937   0.192189   0.352452
#> 139   2017-10-13       CA  <NA> 2016-10-02  52   0.000000   0.000000   0.000000
#> 140   2017-10-20       CA  <NA> 2016-10-09  52   0.000000   0.400000   0.000000
#> 141   2017-10-27       CA  <NA> 2016-10-16  52   0.000000   0.000000   0.100000
#> 142   2017-11-03       CA  <NA> 2016-10-23  52   0.000000   0.200000   0.000000
#> 143   2017-11-09       CA  <NA> 2016-10-30  52   0.000000   0.200000   0.100000
#> 144   2017-11-17       CA  <NA> 2016-11-06  52   1.000000   0.000000   0.000000
#> 145   2017-11-27       CA  <NA> 2016-11-13  52   0.500000   0.000000   0.000000
#> 146   2017-12-01       CA  <NA> 2016-11-20  52   1.500000   0.000000   0.100000
#> 147   2017-12-08       CA  <NA> 2016-11-27  52   0.500000   0.000000   0.100000
#> 148   2017-12-15       CA  <NA> 2016-12-04  52   2.500000   0.000000   0.600000
#> 149   2017-12-22       CA  <NA> 2016-12-11  52   1.500000   1.000000   0.500000
#> 150   2017-12-29       CA  <NA> 2016-12-18  52   1.000000   0.800000   1.000000
#> 151   2018-01-05       CA  <NA> 2016-12-25  52   2.900000   0.400000   1.300000
#> 152   2018-01-12       CA  <NA> 2017-01-01  52   3.400000   0.600000   2.000000
#> 153   2018-01-19       CA  <NA> 2017-01-08  52   2.900000   1.700000   2.100000
#> 154   2018-01-26       CA  <NA> 2017-01-15  52   3.400000   1.000000   1.200000
#> 155   2018-02-02       CA  <NA> 2017-01-22  52   2.000000   0.600000   1.100000
#> 156   2018-02-09       CA  <NA> 2017-01-29  52   2.000000   0.400000   1.100000
#> 157   2018-02-16       CA  <NA> 2017-02-05  52   2.000000   1.200000   0.500000
#> 158   2018-02-14       CA  <NA> 2017-02-12  52   2.000000   0.200000   0.600000
#> 159   2018-03-02       CA  <NA> 2017-02-19  52   0.500000   0.800000   0.200000
#> 160   2018-03-09       CA  <NA> 2017-02-26  52   2.000000   0.200000   0.200000
#> 161   2018-03-16       CA  <NA> 2017-03-05  52   0.500000   0.400000   0.300000
#> 162   2018-03-23       CA  <NA> 2017-03-12  52   0.000000   0.000000   0.100000
#> 163   2018-03-30       CA  <NA> 2017-03-19  52   1.000000   0.000000   0.100000
#> 164   2018-04-06       CA  <NA> 2017-03-26  52   1.000000   0.000000   0.200000
#> 165   2018-04-13       CA  <NA> 2017-04-02  52   1.500000   0.200000   0.100000
#> 166   2018-04-20       CA  <NA> 2017-04-09  52   0.000000   0.000000   0.100000
#> 167   2018-04-27       CA  <NA> 2017-04-16  52   1.000000   0.000000   0.100000
#> 168   2018-05-04       CA  <NA> 2017-04-23  52   0.500000   0.000000   0.100000
#> 169   2018-10-12       CA  <NA> 2017-10-01  52   0.000000   0.000000   0.100000
#> 170   2018-10-19       CA  <NA> 2017-10-08  52   0.000000   0.000000   0.000000
#> 171   2018-10-26       CA  <NA> 2017-10-15  52   0.000000   0.000000   0.000000
#> 172   2018-11-02       CA  <NA> 2017-10-22  52   0.000000   0.000000   0.100000
#> 173   2018-11-09       CA  <NA> 2017-10-29  52   0.500000   0.200000   0.100000
#> 174   2018-11-16       CA  <NA> 2017-11-05  52   0.000000   0.200000   0.300000
#> 175   2018-11-26       CA  <NA> 2017-11-12  52   1.000000   0.400000   0.200000
#> 176   2018-11-30       CA  <NA> 2017-11-19  52   2.500000   0.000000   0.200000
#> 177   2018-12-07       CA  <NA> 2017-11-26  52   1.000000   0.000000   0.100000
#> 178   2018-12-14       CA  <NA> 2017-12-03  52   0.000000   0.600000   0.200000
#> 179   2018-12-21       CA  <NA> 2017-12-10  52   1.500000   0.400000   0.500000
#> 180   2018-12-28       CA  <NA> 2017-12-17  52   2.500000   0.200000   1.400000
#> 181   2019-01-04       CA  <NA> 2017-12-24  52   3.900000   1.100000   2.700000
#> 182   2019-01-11       CA  <NA> 2017-12-31  52   4.400000   1.700000   2.700000
#> 183   2019-01-18       CA  <NA> 2018-01-07  52   4.400000   2.100000   2.700000
#> 184   2019-01-25       CA  <NA> 2018-01-14  52   4.400000   1.000000   2.000000
#> 185   2019-02-01       CA  <NA> 2018-01-21  52   3.400000   1.300000   2.000000
#> 186   2019-02-08       CA  <NA> 2018-01-28  52   3.900000   1.300000   1.700000
#> 187   2019-02-15       CA  <NA> 2018-02-04  52   4.400000   0.800000   1.000000
#> 188   2019-02-22       CA  <NA> 2018-02-11  52   5.400000   0.200000   0.600000
#> 189   2019-03-01       CA  <NA> 2018-02-18  52   2.500000   1.100000   0.900000
#> 190   2019-03-08       CA  <NA> 2018-02-25  52   1.500000   1.100000   0.800000
#> 191   2019-03-15       CA  <NA> 2018-03-04  52   1.500000   1.300000   1.300000
#> 192   2019-03-25       CA  <NA> 2018-03-11  52   2.000000   0.800000   1.000000
#> 193   2019-03-29       CA  <NA> 2018-03-18  52   1.500000   0.400000   0.600000
#> 194   2019-04-05       CA  <NA> 2018-03-25  52   1.500000   0.400000   0.600000
#> 195   2019-04-12       CA  <NA> 2018-04-01  52   0.500000   0.200000   0.400000
#> 196   2019-04-19       CA  <NA> 2018-04-08  52   0.500000   0.000000   0.100000
#> 197   2019-04-26       CA  <NA> 2018-04-15  52   0.500000   0.000000   0.200000
#> 198   2019-05-03       CA  <NA> 2018-04-22  52   1.000000   0.200000   0.400000
#> 199   2019-10-11       CA  <NA> 2018-09-30  52   1.000000   0.000000   0.100000
#> 200   2019-10-18       CA  <NA> 2018-10-07  52   0.000000   0.000000   0.100000
#> 201   2019-10-25       CA  <NA> 2018-10-14  52   0.500000   0.000000   0.000000
#> 202   2019-11-01       CA  <NA> 2018-10-21  52   0.000000   0.000000   0.000000
#> 203   2019-11-08       CA  <NA> 2018-10-28  52   0.000000   0.000000   0.100000
#> 204   2019-11-15       CA  <NA> 2018-11-04  52   0.500000   0.200000   0.000000
#> 205   2019-11-22       CA  <NA> 2018-11-11  52   0.000000   0.200000   0.200000
#> 206   2019-12-02       CA  <NA> 2018-11-18  52   0.500000   0.200000   0.100000
#> 207   2019-12-06       CA  <NA> 2018-11-25  52   0.000000   0.400000   0.100000
#> 208   2019-12-13       CA  <NA> 2018-12-02  52   1.000000   0.400000   0.100000
#> 209   2019-12-20       CA  <NA> 2018-12-09  52   0.500000   0.200000   0.200000
#> 210   2019-12-27       CA  <NA> 2018-12-16  52   0.000000   0.800000   0.800000
#> 211   2020-01-03       CA  <NA> 2018-12-23  52   3.500000   0.800000   0.600000
#> 212   2020-01-10       CA  <NA> 2018-12-30  52   5.500000   0.600000   1.700000
#> 213   2020-01-17       CA  <NA> 2019-01-06  52   5.000000   0.200000   1.500000
#> 214   2020-01-24       CA  <NA> 2019-01-13  52   2.500000   0.800000   1.100000
#> 215   2020-01-31       CA  <NA> 2019-01-20  52   1.000000   1.000000   1.100000
#> 216   2020-02-07       CA  <NA> 2019-01-27  52   1.500000   1.200000   0.900000
#> 217   2020-02-14       CA  <NA> 2019-02-03  52   1.000000   1.000000   0.900000
#> 218   2020-02-21       CA  <NA> 2019-02-10  52   2.500000   2.500000   1.100000
#> 219   2020-02-28       CA  <NA> 2019-02-17  52   1.500000   1.200000   1.400000
#> 220   2020-03-06       CA  <NA> 2019-02-24  52   1.000000   1.000000   1.300000
#> 221   2020-03-13       CA  <NA> 2019-03-03  52   4.000000   2.100000   0.800000
#> 222   2020-03-22       CA  <NA> 2019-03-10  52   2.000000   1.300000   1.200000
#> 223   2020-03-27       CA  <NA> 2019-03-17  52   2.500000   1.000000   1.100000
#> 224   2020-04-03       CA  <NA> 2019-03-24  52   1.000000   1.300000   0.800000
#> 225   2020-04-10       CA  <NA> 2019-03-31  52   2.000000   0.000000   0.500000
#> 226   2020-04-17       CA  <NA> 2019-04-07  52   0.500000   0.000000   0.200000
#> 227   2020-04-24       CA  <NA> 2019-04-14  52   0.000000   0.400000   0.100000
#> 228   2020-05-01       CA  <NA> 2019-04-21  52   0.500000   0.400000   0.200000
#> 229   2020-10-12       CA  <NA> 2019-09-29  52   0.000000   0.000000   0.100000
#> 230   2020-10-16       CA  <NA> 2019-10-06  52   0.000000   0.200000   0.100000
#> 231   2020-10-23       CA  <NA> 2019-10-13  52   0.500000   0.000000   0.100000
#> 232   2020-10-23       CA  <NA> 2019-10-20  51   1.000000   0.200000   0.200000
#> 233   2020-11-06       CA  <NA> 2019-10-27  52   1.000000   0.400000   0.400000
#> 234   2020-11-06       CA  <NA> 2019-11-03  51   0.500000   0.400000   0.800000
#> 235   2020-11-06       CA  <NA> 2019-11-10  50   1.500000   1.200000   0.700000
#> 236   2020-11-06       CA  <NA> 2019-11-17  49   0.000000   0.400000   0.400000
#> 237   2020-11-06       CA  <NA> 2019-11-24  48   1.000000   0.600000   0.300000
#> 238   2020-11-06       CA  <NA> 2019-12-01  47   2.000000   1.000000   0.600000
#> 239   2020-11-06       CA  <NA> 2019-12-08  46   2.500000   0.800000   0.800000
#> 240   2020-11-06       CA  <NA> 2019-12-15  45   1.500000   1.200000   0.800000
#> 241   2020-11-06       CA  <NA> 2019-12-22  44   3.000000   1.200000   1.100000
#> 242   2020-11-06       CA  <NA> 2019-12-29  43   8.600000   0.800000   1.500000
#>     rate_age_3 rate_age_4 rate_overall
#> 1     0.452067   1.180799     0.234198
#> 2     0.150689   1.180799     0.175648
#> 3     0.150689   0.944639     0.175648
#> 4     0.301378   1.889279     0.322022
#> 5     0.753444   2.597758     0.614769
#> 6     0.000000   1.180799     0.263472
#> 7     0.452067   2.361599     0.556219
#> 8     0.602755   2.361599     0.614769
#> 9     1.054822   3.778558     0.936790
#> 10    1.054822   5.431677     1.170988
#> 11    1.356200   7.557115     1.815031
#> 12    1.054822   6.612476     1.580834
#> 13    2.561711   5.903996     1.844306
#> 14    0.602755   3.306238     0.790417
#> 15    1.958955   3.306238     1.083164
#> 16    0.452067   4.014717     0.790417
#> 17    0.301378   2.361599     0.526945
#> 18    0.000000   0.000000     0.028854
#> 19    0.000000   0.225727     0.028854
#> 20    0.000000   0.000000     0.000000
#> 21    0.000000   0.000000     0.000000
#> 22    0.000000   0.000000     0.028854
#> 23    0.000000   0.451454     0.057707
#> 24    0.148362   0.000000     0.028854
#> 25    0.000000   0.225727     0.028854
#> 26    0.148362   0.225727     0.086561
#> 27    0.148362   0.000000     0.057707
#> 28    0.445085   0.225727     0.144269
#> 29    0.000000   1.805816     0.288537
#> 30    0.445085   2.482997     0.519367
#> 31    1.038531   6.546083     1.384978
#> 32    1.038531  11.737805     2.279443
#> 33    2.670508  13.317894     3.202762
#> 34    4.747570  26.861514     5.857303
#> 35    5.192655  16.252345     4.241495
#> 36    2.818870  16.478072     3.635567
#> 37    3.115593  16.703799     3.375884
#> 38    2.225424  10.383442     2.366004
#> 39    1.186893  12.414986     2.164028
#> 40    0.445085   8.351899     1.413832
#> 41    0.296723   4.514540     0.807904
#> 42    0.000000   2.482997     0.605928
#> 43    0.741808   0.451454     0.432806
#> 44    0.148362   0.902908     0.288537
#> 45    0.445085   0.677181     0.230830
#> 46    0.148362   0.677181     0.201976
#> 47    0.000000   0.000000     0.086561
#> 48    0.000000   0.000000     0.056810
#> 49    0.145821   0.000000     0.028405
#> 50    0.000000   0.432633     0.085215
#> 51    0.145821   0.432633     0.085215
#> 52    0.000000   0.000000     0.028405
#> 53    0.000000   0.432633     0.085215
#> 54    0.145821   0.000000     0.085215
#> 55    0.145821   0.216316     0.085215
#> 56    0.291643   0.432633     0.198834
#> 57    0.583286   0.432633     0.312454
#> 58    0.583286   0.865265     0.454478
#> 59    1.458215   2.812112     1.022577
#> 60    3.791359   6.922122     2.386012
#> 61    5.978681   8.436336     3.664232
#> 62    7.145253  11.464764     4.629999
#> 63    5.103752   9.950550     3.692637
#> 64    4.083002   7.138438     2.897300
#> 65    3.062251   5.407908     2.073558
#> 66    2.624787   4.542642     1.704294
#> 67    1.020750   1.730530     0.511288
#> 68    0.291643   2.163163     0.426074
#> 69    0.000000   0.865265     0.255644
#> 70    0.000000   0.648949     0.227239
#> 71    0.000000   0.865265     0.142025
#> 72    0.145821   0.648949     0.170429
#> 73    0.000000   0.000000     0.056810
#> 74    0.000000   0.648949     0.113620
#> 75    0.000000   0.000000     0.028405
#> 76    0.145821   0.000000     0.056810
#> 77    0.145821   0.432633     0.170429
#> 78    0.000000   0.000000     0.027956
#> 79    0.143508   0.000000     0.027956
#> 80    0.143508   0.207954     0.083867
#> 81    0.143508   0.415908     0.111823
#> 82    0.000000   0.207954     0.027956
#> 83    0.000000   0.207954     0.055911
#> 84    0.287015   0.000000     0.055911
#> 85    0.000000   0.623861     0.083867
#> 86    0.000000   0.207954     0.083867
#> 87    0.000000   0.831815     0.195689
#> 88    0.143508   1.871584     0.335468
#> 89    0.143508   2.079538     0.419334
#> 90    0.430523   3.119307     0.810713
#> 91    1.578584   8.526107     1.928938
#> 92    3.731199  24.954458     5.059969
#> 93    4.735752  28.281719     5.674993
#> 94    3.731199  29.321488     5.507259
#> 95    4.305229  31.401026     6.290017
#> 96    2.296122  32.024888     5.702949
#> 97    1.865599  19.963566     3.941744
#> 98    1.148061  10.397691     2.096672
#> 99    1.148061   6.030661     1.369826
#> 100   0.287015   6.238615     1.034358
#> 101   1.004553   3.119307     0.894580
#> 102   0.287015   2.911353     0.447290
#> 103   0.287015   2.079538     0.475246
#> 104   1.004553   1.039769     0.447290
#> 105   0.287015   1.039769     0.335468
#> 106   0.287015   0.207954     0.167734
#> 107   0.143508   0.831815     0.223645
#> 108   0.000000   0.623861     0.223645
#> 109   0.000000   0.400154     0.110200
#> 110   0.000000   0.600230     0.137750
#> 111   0.000000   0.600230     0.110200
#> 112   0.000000   0.000000     0.000000
#> 113   0.000000   0.400154     0.055100
#> 114   0.000000   0.200077     0.055100
#> 115   0.000000   0.400154     0.055100
#> 116   0.000000   0.200077     0.082650
#> 117   0.142041   0.600230     0.165299
#> 118   0.284082   0.800307     0.247949
#> 119   0.142041   0.200077     0.110200
#> 120   0.142041   2.000768     0.440799
#> 121   0.000000   1.800691     0.330599
#> 122   0.994286   4.201613     1.019347
#> 123   0.426122   2.600999     0.771397
#> 124   0.142041   1.800691     0.495898
#> 125   0.994286   3.801460     1.019347
#> 126   1.846531   3.001152     1.212196
#> 127   1.278367   6.602535     1.873394
#> 128   2.982857   5.001921     1.983593
#> 129   2.414694   6.602535     2.093793
#> 130   3.551020   8.203150     2.369292
#> 131   1.704490   8.403227     2.286642
#> 132   1.846531   6.002305     2.011143
#> 133   1.420408   5.402074     1.625445
#> 134   1.988571   4.601767     1.322396
#> 135   1.562449   4.601767     1.212196
#> 136   0.710204   2.600999     0.688748
#> 137   0.426122   1.200461     0.413249
#> 138   0.852245   1.200461     0.578548
#> 139   0.000000   0.600000     0.100000
#> 140   0.400000   1.400000     0.300000
#> 141   0.400000   1.800000     0.400000
#> 142   0.100000   2.300000     0.400000
#> 143   0.100000   2.300000     0.400000
#> 144   0.900000   2.100000     0.500000
#> 145   0.700000   1.800000     0.400000
#> 146   0.700000   1.400000     0.500000
#> 147   0.300000   3.100000     0.600000
#> 148   1.000000   7.000000     1.600000
#> 149   1.100000   8.000000     1.800000
#> 150   2.800000  16.100000     3.400000
#> 151   6.400000  40.100000     7.700000
#> 152  10.100000  63.000000    12.000000
#> 153  10.000000  52.100000    10.600000
#> 154   8.000000  34.600000     7.300000
#> 155   3.400000  24.100000     4.800000
#> 156   4.100000  22.900000     4.700000
#> 157   2.700000  15.000000     3.100000
#> 158   1.700000   8.200000     1.900000
#> 159   1.700000   6.400000     1.500000
#> 160   1.000000   6.000000     1.300000
#> 161   1.000000   5.400000     1.200000
#> 162   0.600000   3.300000     0.600000
#> 163   0.400000   3.700000     0.700000
#> 164   0.400000   3.300000     0.700000
#> 165   0.700000   3.500000     0.800000
#> 166   0.600000   2.700000     0.500000
#> 167   0.900000   1.600000     0.500000
#> 168   0.400000   2.700000     0.500000
#> 169   0.100000   1.100000     0.200000
#> 170   0.100000   0.700000     0.100000
#> 171   0.400000   2.600000     0.500000
#> 172   0.900000   2.400000     0.500000
#> 173   0.300000   1.700000     0.400000
#> 174   0.300000   2.400000     0.600000
#> 175   0.300000   2.100000     0.500000
#> 176   0.700000   3.400000     0.900000
#> 177   0.600000   3.200000     0.700000
#> 178   0.600000   5.200000     1.000000
#> 179   2.300000   9.100000     2.100000
#> 180   5.300000  26.500000     5.700000
#> 181  12.900000  74.900000    15.000000
#> 182  18.700000  99.700000    19.800000
#> 183   8.500000  59.000000    12.000000
#> 184   8.000000  42.400000     9.000000
#> 185   6.200000  25.000000     6.100000
#> 186   4.400000  18.700000     4.800000
#> 187   4.100000  12.500000     3.400000
#> 188   3.100000  12.900000     3.100000
#> 189   2.800000  14.900000     3.400000
#> 190   6.400000  20.200000     4.800000
#> 191   3.800000  24.500000     5.100000
#> 192   3.700000  20.900000     4.400000
#> 193   2.300000  10.500000     2.400000
#> 194   1.400000   9.000000     2.000000
#> 195   1.300000   5.800000     1.300000
#> 196   1.300000   4.900000     1.000000
#> 197   0.600000   2.800000     0.600000
#> 198   0.700000   2.800000     0.800000
#> 199   0.000000   0.700000     0.200000
#> 200   0.300000   0.600000     0.200000
#> 201   0.100000   0.700000     0.200000
#> 202   0.000000   1.300000     0.200000
#> 203   0.100000   0.400000     0.100000
#> 204   0.000000   0.700000     0.200000
#> 205   0.000000   0.200000     0.200000
#> 206   0.700000   1.300000     0.400000
#> 207   1.300000   2.500000     0.700000
#> 208   0.300000   1.100000     0.400000
#> 209   1.300000   2.000000     0.700000
#> 210   1.900000   3.300000     1.300000
#> 211   2.400000   4.700000     1.700000
#> 212   4.900000  10.000000     3.600000
#> 213   5.400000  10.700000     3.600000
#> 214   4.300000   7.100000     2.600000
#> 215   2.900000   6.200000     2.200000
#> 216   4.300000   7.100000     2.500000
#> 217   2.000000   8.200000     2.200000
#> 218   3.100000   9.800000     3.100000
#> 219   4.000000  15.600000     4.000000
#> 220   6.100000  24.000000     5.500000
#> 221   6.300000  17.300000     4.600000
#> 222   3.700000  14.000000     3.600000
#> 223   4.700000  16.400000     4.100000
#> 224   2.300000  11.600000     2.800000
#> 225   1.300000   8.400000     1.800000
#> 226   0.900000   4.900000     1.000000
#> 227   1.400000   2.400000     0.700000
#> 228   0.600000   2.700000     0.700000
#> 229   0.000000   0.900000     0.200000
#> 230   0.100000   0.500000     0.200000
#> 231   0.100000   0.400000     0.100000
#> 232   0.100000   1.400000     0.400000
#> 233   0.600000   1.800000     0.700000
#> 234   1.400000   0.900000     0.900000
#> 235   1.100000   1.600000     1.000000
#> 236   0.900000   3.000000     0.900000
#> 237   1.700000   1.800000     0.900000
#> 238   1.700000   4.600000     1.500000
#> 239   1.900000   4.000000     1.600000
#> 240   2.200000   6.300000     2.000000
#> 241   5.600000   7.200000     3.000000
#> 242   5.700000  15.300000     4.700000
#> 
#> $result
#> [1] 1
#> 
#> $message
#> [1] "success"
#> 
# }