Fetch FluView data

Obtains information on outpatient inluenza-like-illness (ILI) from U.S. Outpatient Influenza-like Illness Surveillance Network (ILINet).

fluview(regions, epiweeks, issues = NULL, lag = NULL, auth = NULL)

Arguments

regions

character vector. The locations to fetch. Can we any string IDs in national, HHS region, census division, most states and territories, and so on. Full list link below.

epiweeks

timeset. The epiweeks to fetch in the form epirange(startweek,endweek), where startweek and endweek are of the form YYYYWW (string or numeric).

issues

timeset. Optionally, the issues to fetch. If not set, the most recent issue is returned. Mutually exclusive with lag.

lag

integer. Optionally, the lag of the issues to fetch. If not set, the most recent issue is returned. Mutually exclusive with issues.

auth

string. Optionally, restricted access key (not the same as API key).

Value

epidata_call

Details

API docs: https://cmu-delphi.github.io/delphi-epidata/api/fluview.html. For more information on ILINet, see https://gis.cdc.gov/grasp/fluview/fluportaldashboard.html.

The full list of location inputs can be accsssed at https://github.com/cmu-delphi/delphi-epidata/blob/main/src/acquisition/fluview/fluview_locations.py.

Examples

fluview(regions = "nat", epiweeks = epirange(201201, 202001)) %>% fetch()
#> # A tibble: 418 × 16
#>    release_date region issue      epiweek      lag num_ili num_patients
#>    <chr>        <chr>  <date>     <date>     <int>   <int>        <int>
#>  1 2017-10-24   nat    2017-10-01 2012-01-01   300   11992       678631
#>  2 2017-10-24   nat    2017-10-01 2012-01-08   299   11543       747894
#>  3 2017-10-24   nat    2017-10-01 2012-01-15   298   11939       724623
#>  4 2017-10-24   nat    2017-10-01 2012-01-22   297   13209       784244
#>  5 2017-10-24   nat    2017-10-01 2012-01-29   296   14448       775298
#>  6 2017-10-24   nat    2017-10-01 2012-02-05   295   14624       784516
#>  7 2017-10-24   nat    2017-10-01 2012-02-12   294   15929       789193
#>  8 2017-10-24   nat    2017-10-01 2012-02-19   293   16137       767022
#>  9 2017-10-24   nat    2017-10-01 2012-02-26   292   16518       788242
#> 10 2017-10-24   nat    2017-10-01 2012-03-04   291   15763       749198
#> # ℹ 408 more rows
#> # ℹ 9 more variables: num_providers <int>, num_age_0 <int>, num_age_1 <int>,
#> #   num_age_2 <int>, num_age_3 <int>, num_age_4 <int>, num_age_5 <int>,
#> #   wili <dbl>, ili <dbl>