Fetch FluView virological data from clinical labs

API docs: https://cmu-delphi.github.io/delphi-epidata/api/fluview_clinical.html

fluview_clinical(regions, epiweeks, issues = NULL, lag = NULL)

Arguments

regions

character vector. The regions to fetch.

epiweeks

timeset. The epiweeks to fetch in the form epirange(startweek,endweek), where startweek and endweek are of the form YYYYWW (string or numeric).

issues

timeset. Optionally, the issues to fetch. If not set, the most recent issue is returned. Mutually exclusive with lag.

lag

integer. Optionally, the lag of the issues to fetch. If not set, the most recent issue is returned. Mutually exclusive with issues.

Value

epidata_call

Examples

# \donttest{
fluview_clinical(regions = "nat", epiweeks = epirange(201601, 201701)) %>% fetch_classic()
#> $epidata
#>     release_date region      issue    epiweek lag total_specimens total_a
#> 1     2018-10-08    nat 2018-09-23 2016-10-02 103           13380     120
#> 2     2018-10-08    nat 2018-09-23 2016-10-02 103           13380     120
#> 3     2018-10-09    nat 2018-09-23 2016-10-02 103           13380     120
#> 4     2018-10-09    nat 2018-09-23 2016-10-02 103           13380     120
#> 5     2018-10-09    nat 2018-09-23 2016-10-02 103           13380     120
#> 6     2018-10-10    nat 2018-09-23 2016-10-02 103           13380     120
#> 7     2018-10-10    nat 2018-09-23 2016-10-02 103           13380     120
#> 8     2018-10-10    nat 2018-09-23 2016-10-02 103           13380     120
#> 9     2018-10-11    nat 2018-09-23 2016-10-02 103           13380     120
#> 10    2018-10-11    nat 2018-09-23 2016-10-02 103           13380     120
#> 11    2018-10-11    nat 2018-09-23 2016-10-02 103           13380     120
#> 12    2018-10-12    nat 2018-09-23 2016-10-02 103           13380     120
#> 13    2018-10-08    nat 2018-09-23 2016-10-09 102           14053     108
#> 14    2018-10-08    nat 2018-09-23 2016-10-09 102           14053     108
#> 15    2018-10-09    nat 2018-09-23 2016-10-09 102           14053     108
#> 16    2018-10-09    nat 2018-09-23 2016-10-09 102           14053     108
#> 17    2018-10-09    nat 2018-09-23 2016-10-09 102           14053     108
#> 18    2018-10-10    nat 2018-09-23 2016-10-09 102           14053     108
#> 19    2018-10-10    nat 2018-09-23 2016-10-09 102           14053     108
#> 20    2018-10-10    nat 2018-09-23 2016-10-09 102           14053     108
#> 21    2018-10-11    nat 2018-09-23 2016-10-09 102           14053     108
#> 22    2018-10-11    nat 2018-09-23 2016-10-09 102           14053     108
#> 23    2018-10-11    nat 2018-09-23 2016-10-09 102           14053     108
#> 24    2018-10-12    nat 2018-09-23 2016-10-09 102           14053     108
#> 25    2018-10-08    nat 2018-09-23 2016-10-16 101           15110     115
#> 26    2018-10-08    nat 2018-09-23 2016-10-16 101           15110     115
#> 27    2018-10-09    nat 2018-09-23 2016-10-16 101           15110     115
#> 28    2018-10-09    nat 2018-09-23 2016-10-16 101           15110     115
#> 29    2018-10-09    nat 2018-09-23 2016-10-16 101           15110     115
#> 30    2018-10-10    nat 2018-09-23 2016-10-16 101           15110     115
#> 31    2018-10-10    nat 2018-09-23 2016-10-16 101           15110     115
#> 32    2018-10-10    nat 2018-09-23 2016-10-16 101           15110     115
#> 33    2018-10-11    nat 2018-09-23 2016-10-16 101           15110     115
#> 34    2018-10-11    nat 2018-09-23 2016-10-16 101           15110     115
#> 35    2018-10-11    nat 2018-09-23 2016-10-16 101           15110     115
#> 36    2018-10-12    nat 2018-09-23 2016-10-16 101           15110     115
#> 37    2018-10-08    nat 2018-09-23 2016-10-23 100           15312     143
#> 38    2018-10-08    nat 2018-09-23 2016-10-23 100           15312     143
#> 39    2018-10-09    nat 2018-09-23 2016-10-23 100           15312     143
#> 40    2018-10-09    nat 2018-09-23 2016-10-23 100           15312     143
#> 41    2018-10-09    nat 2018-09-23 2016-10-23 100           15312     143
#> 42    2018-10-10    nat 2018-09-23 2016-10-23 100           15312     143
#> 43    2018-10-10    nat 2018-09-23 2016-10-23 100           15312     143
#> 44    2018-10-10    nat 2018-09-23 2016-10-23 100           15312     143
#> 45    2018-10-11    nat 2018-09-23 2016-10-23 100           15312     143
#> 46    2018-10-11    nat 2018-09-23 2016-10-23 100           15312     143
#> 47    2018-10-11    nat 2018-09-23 2016-10-23 100           15312     143
#> 48    2018-10-12    nat 2018-09-23 2016-10-23 100           15312     143
#> 49    2018-10-08    nat 2018-09-23 2016-10-30  99           16652     201
#> 50    2018-10-08    nat 2018-09-23 2016-10-30  99           16652     201
#> 51    2018-10-09    nat 2018-09-23 2016-10-30  99           16652     201
#> 52    2018-10-09    nat 2018-09-23 2016-10-30  99           16652     201
#> 53    2018-10-09    nat 2018-09-23 2016-10-30  99           16652     201
#> 54    2018-10-10    nat 2018-09-23 2016-10-30  99           16652     201
#> 55    2018-10-10    nat 2018-09-23 2016-10-30  99           16652     201
#> 56    2018-10-10    nat 2018-09-23 2016-10-30  99           16652     201
#> 57    2018-10-11    nat 2018-09-23 2016-10-30  99           16652     201
#> 58    2018-10-11    nat 2018-09-23 2016-10-30  99           16652     201
#> 59    2018-10-11    nat 2018-09-23 2016-10-30  99           16652     201
#> 60    2018-10-12    nat 2018-09-23 2016-10-30  99           16652     201
#> 61    2018-10-08    nat 2018-09-23 2016-11-06  98           17811     271
#> 62    2018-10-08    nat 2018-09-23 2016-11-06  98           17811     271
#> 63    2018-10-09    nat 2018-09-23 2016-11-06  98           17811     271
#> 64    2018-10-09    nat 2018-09-23 2016-11-06  98           17811     271
#> 65    2018-10-09    nat 2018-09-23 2016-11-06  98           17811     271
#> 66    2018-10-10    nat 2018-09-23 2016-11-06  98           17811     271
#> 67    2018-10-10    nat 2018-09-23 2016-11-06  98           17811     271
#> 68    2018-10-10    nat 2018-09-23 2016-11-06  98           17811     271
#> 69    2018-10-11    nat 2018-09-23 2016-11-06  98           17811     271
#> 70    2018-10-11    nat 2018-09-23 2016-11-06  98           17811     271
#> 71    2018-10-11    nat 2018-09-23 2016-11-06  98           17811     271
#> 72    2018-10-12    nat 2018-09-23 2016-11-06  98           17811     271
#> 73    2018-10-08    nat 2018-09-23 2016-11-13  97           18948     347
#> 74    2018-10-08    nat 2018-09-23 2016-11-13  97           18948     347
#> 75    2018-10-09    nat 2018-09-23 2016-11-13  97           18948     347
#> 76    2018-10-09    nat 2018-09-23 2016-11-13  97           18948     347
#> 77    2018-10-09    nat 2018-09-23 2016-11-13  97           18948     347
#> 78    2018-10-10    nat 2018-09-23 2016-11-13  97           18948     347
#> 79    2018-10-10    nat 2018-09-23 2016-11-13  97           18948     347
#> 80    2018-10-10    nat 2018-09-23 2016-11-13  97           18948     347
#> 81    2018-10-11    nat 2018-09-23 2016-11-13  97           18948     347
#> 82    2018-10-11    nat 2018-09-23 2016-11-13  97           18948     347
#> 83    2018-10-11    nat 2018-09-23 2016-11-13  97           18948     347
#> 84    2018-10-12    nat 2018-09-23 2016-11-13  97           18948     347
#> 85    2018-10-08    nat 2018-09-23 2016-11-20  96           18520     451
#> 86    2018-10-08    nat 2018-09-23 2016-11-20  96           18520     451
#> 87    2018-10-09    nat 2018-09-23 2016-11-20  96           18520     451
#> 88    2018-10-09    nat 2018-09-23 2016-11-20  96           18520     451
#> 89    2018-10-09    nat 2018-09-23 2016-11-20  96           18520     451
#> 90    2018-10-10    nat 2018-09-23 2016-11-20  96           18520     451
#> 91    2018-10-10    nat 2018-09-23 2016-11-20  96           18520     451
#> 92    2018-10-10    nat 2018-09-23 2016-11-20  96           18520     451
#> 93    2018-10-11    nat 2018-09-23 2016-11-20  96           18520     451
#> 94    2018-10-11    nat 2018-09-23 2016-11-20  96           18520     451
#> 95    2018-10-11    nat 2018-09-23 2016-11-20  96           18520     451
#> 96    2018-10-12    nat 2018-09-23 2016-11-20  96           18520     451
#> 97    2018-10-08    nat 2018-09-23 2016-11-27  95           21542     532
#> 98    2018-10-08    nat 2018-09-23 2016-11-27  95           21542     532
#> 99    2018-10-09    nat 2018-09-23 2016-11-27  95           21542     532
#> 100   2018-10-09    nat 2018-09-23 2016-11-27  95           21542     532
#> 101   2018-10-09    nat 2018-09-23 2016-11-27  95           21542     532
#> 102   2018-10-10    nat 2018-09-23 2016-11-27  95           21542     532
#> 103   2018-10-10    nat 2018-09-23 2016-11-27  95           21542     532
#> 104   2018-10-10    nat 2018-09-23 2016-11-27  95           21542     532
#> 105   2018-10-11    nat 2018-09-23 2016-11-27  95           21542     532
#> 106   2018-10-11    nat 2018-09-23 2016-11-27  95           21542     532
#> 107   2018-10-11    nat 2018-09-23 2016-11-27  95           21542     532
#> 108   2018-10-12    nat 2018-09-23 2016-11-27  95           21542     532
#> 109   2018-10-08    nat 2018-09-23 2016-12-04  94           20564     696
#> 110   2018-10-08    nat 2018-09-23 2016-12-04  94           20564     696
#> 111   2018-10-09    nat 2018-09-23 2016-12-04  94           20564     696
#> 112   2018-10-09    nat 2018-09-23 2016-12-04  94           20564     696
#> 113   2018-10-09    nat 2018-09-23 2016-12-04  94           20564     696
#> 114   2018-10-10    nat 2018-09-23 2016-12-04  94           20564     696
#> 115   2018-10-10    nat 2018-09-23 2016-12-04  94           20564     696
#> 116   2018-10-10    nat 2018-09-23 2016-12-04  94           20564     696
#> 117   2018-10-11    nat 2018-09-23 2016-12-04  94           20564     696
#> 118   2018-10-11    nat 2018-09-23 2016-12-04  94           20564     696
#> 119   2018-10-11    nat 2018-09-23 2016-12-04  94           20564     696
#> 120   2018-10-12    nat 2018-09-23 2016-12-04  94           20564     696
#> 121   2018-10-08    nat 2018-09-23 2016-12-11  93           23476    1359
#> 122   2018-10-08    nat 2018-09-23 2016-12-11  93           23476    1359
#> 123   2018-10-09    nat 2018-09-23 2016-12-11  93           23476    1359
#> 124   2018-10-09    nat 2018-09-23 2016-12-11  93           23476    1359
#> 125   2018-10-09    nat 2018-09-23 2016-12-11  93           23476    1359
#> 126   2018-10-10    nat 2018-09-23 2016-12-11  93           23476    1359
#> 127   2018-10-10    nat 2018-09-23 2016-12-11  93           23476    1359
#> 128   2018-10-10    nat 2018-09-23 2016-12-11  93           23476    1359
#> 129   2018-10-11    nat 2018-09-23 2016-12-11  93           23476    1359
#> 130   2018-10-11    nat 2018-09-23 2016-12-11  93           23476    1359
#> 131   2018-10-11    nat 2018-09-23 2016-12-11  93           23476    1359
#> 132   2018-10-12    nat 2018-09-23 2016-12-11  93           23476    1359
#> 133   2018-10-08    nat 2018-09-23 2016-12-18  92           26775    2556
#> 134   2018-10-08    nat 2018-09-23 2016-12-18  92           26775    2556
#> 135   2018-10-09    nat 2018-09-23 2016-12-18  92           26775    2556
#> 136   2018-10-09    nat 2018-09-23 2016-12-18  92           26775    2556
#> 137   2018-10-09    nat 2018-09-23 2016-12-18  92           26775    2556
#> 138   2018-10-10    nat 2018-09-23 2016-12-18  92           26775    2556
#> 139   2018-10-10    nat 2018-09-23 2016-12-18  92           26775    2556
#> 140   2018-10-10    nat 2018-09-23 2016-12-18  92           26775    2556
#> 141   2018-10-11    nat 2018-09-23 2016-12-18  92           26775    2556
#> 142   2018-10-11    nat 2018-09-23 2016-12-18  92           26775    2556
#> 143   2018-10-11    nat 2018-09-23 2016-12-18  92           26775    2556
#> 144   2018-10-12    nat 2018-09-23 2016-12-18  92           26775    2556
#> 145   2018-10-08    nat 2018-09-23 2016-12-25  91           33273    4270
#> 146   2018-10-08    nat 2018-09-23 2016-12-25  91           33273    4270
#> 147   2018-10-09    nat 2018-09-23 2016-12-25  91           33273    4270
#> 148   2018-10-09    nat 2018-09-23 2016-12-25  91           33273    4270
#> 149   2018-10-09    nat 2018-09-23 2016-12-25  91           33273    4270
#> 150   2018-10-10    nat 2018-09-23 2016-12-25  91           33273    4270
#> 151   2018-10-10    nat 2018-09-23 2016-12-25  91           33273    4270
#> 152   2018-10-10    nat 2018-09-23 2016-12-25  91           33273    4270
#> 153   2018-10-11    nat 2018-09-23 2016-12-25  91           33273    4270
#> 154   2018-10-11    nat 2018-09-23 2016-12-25  91           33273    4270
#> 155   2018-10-11    nat 2018-09-23 2016-12-25  91           33273    4270
#> 156   2018-10-12    nat 2018-09-23 2016-12-25  91           33273    4270
#> 157   2018-10-08    nat 2018-09-23 2017-01-01  90           35028    4288
#> 158   2018-10-08    nat 2018-09-23 2017-01-01  90           35028    4288
#> 159   2018-10-09    nat 2018-09-23 2017-01-01  90           35028    4288
#> 160   2018-10-09    nat 2018-09-23 2017-01-01  90           35028    4288
#> 161   2018-10-09    nat 2018-09-23 2017-01-01  90           35028    4288
#> 162   2018-10-10    nat 2018-09-23 2017-01-01  90           35028    4288
#> 163   2018-10-10    nat 2018-09-23 2017-01-01  90           35028    4288
#> 164   2018-10-10    nat 2018-09-23 2017-01-01  90           35028    4288
#> 165   2018-10-11    nat 2018-09-23 2017-01-01  90           35028    4288
#> 166   2018-10-11    nat 2018-09-23 2017-01-01  90           35028    4288
#> 167   2018-10-11    nat 2018-09-23 2017-01-01  90           35028    4288
#> 168   2018-10-12    nat 2018-09-23 2017-01-01  90           35028    4288
#>     total_b percent_positive percent_a percent_b
#> 1        90          1.56951  0.896861  0.672646
#> 2        90          1.56951  0.896861  0.672646
#> 3        90          1.56951  0.896861  0.672646
#> 4        90          1.56951  0.896861  0.672646
#> 5        90          1.56951  0.896861  0.672646
#> 6        90          1.56951  0.896861  0.672646
#> 7        90          1.56951  0.896861  0.672646
#> 8        90          1.56951  0.896861  0.672646
#> 9        90          1.56951  0.896861  0.672646
#> 10       90          1.56951  0.896861  0.672646
#> 11       90          1.56951  0.896861  0.672646
#> 12       90          1.56951  0.896861  0.672646
#> 13       83          1.35914  0.768519  0.590621
#> 14       83          1.35914  0.768519  0.590621
#> 15       83          1.35914  0.768519  0.590621
#> 16       83          1.35914  0.768519  0.590621
#> 17       83          1.35914  0.768519  0.590621
#> 18       83          1.35914  0.768519  0.590621
#> 19       83          1.35914  0.768519  0.590621
#> 20       83          1.35914  0.768519  0.590621
#> 21       83          1.35914  0.768519  0.590621
#> 22       83          1.35914  0.768519  0.590621
#> 23       83          1.35914  0.768519  0.590621
#> 24       83          1.35914  0.768519  0.590621
#> 25       97          1.40304  0.761085  0.641959
#> 26       97          1.40304  0.761085  0.641959
#> 27       97          1.40304  0.761085  0.641959
#> 28       97          1.40304  0.761085  0.641959
#> 29       97          1.40304  0.761085  0.641959
#> 30       97          1.40304  0.761085  0.641959
#> 31       97          1.40304  0.761085  0.641959
#> 32       97          1.40304  0.761085  0.641959
#> 33       97          1.40304  0.761085  0.641959
#> 34       97          1.40304  0.761085  0.641959
#> 35       97          1.40304  0.761085  0.641959
#> 36       97          1.40304  0.761085  0.641959
#> 37       88          1.50862  0.933908  0.574713
#> 38       88          1.50862  0.933908  0.574713
#> 39       88          1.50862  0.933908  0.574713
#> 40       88          1.50862  0.933908  0.574713
#> 41       88          1.50862  0.933908  0.574713
#> 42       88          1.50862  0.933908  0.574713
#> 43       88          1.50862  0.933908  0.574713
#> 44       88          1.50862  0.933908  0.574713
#> 45       88          1.50862  0.933908  0.574713
#> 46       88          1.50862  0.933908  0.574713
#> 47       88          1.50862  0.933908  0.574713
#> 48       88          1.50862  0.933908  0.574713
#> 49      118          1.91569  1.207060  0.708624
#> 50      118          1.91569  1.207060  0.708624
#> 51      118          1.91569  1.207060  0.708624
#> 52      118          1.91569  1.207060  0.708624
#> 53      118          1.91569  1.207060  0.708624
#> 54      118          1.91569  1.207060  0.708624
#> 55      118          1.91569  1.207060  0.708624
#> 56      118          1.91569  1.207060  0.708624
#> 57      118          1.91569  1.207060  0.708624
#> 58      118          1.91569  1.207060  0.708624
#> 59      118          1.91569  1.207060  0.708624
#> 60      118          1.91569  1.207060  0.708624
#> 61      121          2.20089  1.521530  0.679355
#> 62      121          2.20089  1.521530  0.679355
#> 63      121          2.20089  1.521530  0.679355
#> 64      121          2.20089  1.521530  0.679355
#> 65      121          2.20089  1.521530  0.679355
#> 66      121          2.20089  1.521530  0.679355
#> 67      121          2.20089  1.521530  0.679355
#> 68      121          2.20089  1.521530  0.679355
#> 69      121          2.20089  1.521530  0.679355
#> 70      121          2.20089  1.521530  0.679355
#> 71      121          2.20089  1.521530  0.679355
#> 72      121          2.20089  1.521530  0.679355
#> 73      144          2.59130  1.831330  0.759975
#> 74      144          2.59130  1.831330  0.759975
#> 75      144          2.59130  1.831330  0.759975
#> 76      144          2.59130  1.831330  0.759975
#> 77      144          2.59130  1.831330  0.759975
#> 78      144          2.59130  1.831330  0.759975
#> 79      144          2.59130  1.831330  0.759975
#> 80      144          2.59130  1.831330  0.759975
#> 81      144          2.59130  1.831330  0.759975
#> 82      144          2.59130  1.831330  0.759975
#> 83      144          2.59130  1.831330  0.759975
#> 84      144          2.59130  1.831330  0.759975
#> 85      169          3.34773  2.435210  0.912527
#> 86      169          3.34773  2.435210  0.912527
#> 87      169          3.34773  2.435210  0.912527
#> 88      169          3.34773  2.435210  0.912527
#> 89      169          3.34773  2.435210  0.912527
#> 90      169          3.34773  2.435210  0.912527
#> 91      169          3.34773  2.435210  0.912527
#> 92      169          3.34773  2.435210  0.912527
#> 93      169          3.34773  2.435210  0.912527
#> 94      169          3.34773  2.435210  0.912527
#> 95      169          3.34773  2.435210  0.912527
#> 96      169          3.34773  2.435210  0.912527
#> 97      183          3.31910  2.469590  0.849503
#> 98      183          3.31910  2.469590  0.849503
#> 99      183          3.31910  2.469590  0.849503
#> 100     183          3.31910  2.469590  0.849503
#> 101     183          3.31910  2.469590  0.849503
#> 102     183          3.31910  2.469590  0.849503
#> 103     183          3.31910  2.469590  0.849503
#> 104     183          3.31910  2.469590  0.849503
#> 105     183          3.31910  2.469590  0.849503
#> 106     183          3.31910  2.469590  0.849503
#> 107     183          3.31910  2.469590  0.849503
#> 108     183          3.31910  2.469590  0.849503
#> 109     180          4.25987  3.384560  0.875316
#> 110     180          4.25987  3.384560  0.875316
#> 111     180          4.25987  3.384560  0.875316
#> 112     180          4.25987  3.384560  0.875316
#> 113     180          4.25987  3.384560  0.875316
#> 114     180          4.25987  3.384560  0.875316
#> 115     180          4.25987  3.384560  0.875316
#> 116     180          4.25987  3.384560  0.875316
#> 117     180          4.25987  3.384560  0.875316
#> 118     180          4.25987  3.384560  0.875316
#> 119     180          4.25987  3.384560  0.875316
#> 120     180          4.25987  3.384560  0.875316
#> 121     210          6.68342  5.788890  0.894531
#> 122     210          6.68342  5.788890  0.894531
#> 123     210          6.68342  5.788890  0.894531
#> 124     210          6.68342  5.788890  0.894531
#> 125     210          6.68342  5.788890  0.894531
#> 126     210          6.68342  5.788890  0.894531
#> 127     210          6.68342  5.788890  0.894531
#> 128     210          6.68342  5.788890  0.894531
#> 129     210          6.68342  5.788890  0.894531
#> 130     210          6.68342  5.788890  0.894531
#> 131     210          6.68342  5.788890  0.894531
#> 132     210          6.68342  5.788890  0.894531
#> 133     339         10.81230  9.546220  1.266110
#> 134     339         10.81230  9.546220  1.266110
#> 135     339         10.81230  9.546220  1.266110
#> 136     339         10.81230  9.546220  1.266110
#> 137     339         10.81230  9.546220  1.266110
#> 138     339         10.81230  9.546220  1.266110
#> 139     339         10.81230  9.546220  1.266110
#> 140     339         10.81230  9.546220  1.266110
#> 141     339         10.81230  9.546220  1.266110
#> 142     339         10.81230  9.546220  1.266110
#> 143     339         10.81230  9.546220  1.266110
#> 144     339         10.81230  9.546220  1.266110
#> 145     388         13.99930 12.833200  1.166110
#> 146     388         13.99930 12.833200  1.166110
#> 147     388         13.99930 12.833200  1.166110
#> 148     388         13.99930 12.833200  1.166110
#> 149     388         13.99930 12.833200  1.166110
#> 150     388         13.99930 12.833200  1.166110
#> 151     388         13.99930 12.833200  1.166110
#> 152     388         13.99930 12.833200  1.166110
#> 153     388         13.99930 12.833200  1.166110
#> 154     388         13.99930 12.833200  1.166110
#> 155     388         13.99930 12.833200  1.166110
#> 156     388         13.99930 12.833200  1.166110
#> 157     386         13.34360 12.241600  1.101980
#> 158     386         13.34360 12.241600  1.101980
#> 159     386         13.34360 12.241600  1.101980
#> 160     386         13.34360 12.241600  1.101980
#> 161     386         13.34360 12.241600  1.101980
#> 162     386         13.34360 12.241600  1.101980
#> 163     386         13.34360 12.241600  1.101980
#> 164     386         13.34360 12.241600  1.101980
#> 165     386         13.34360 12.241600  1.101980
#> 166     386         13.34360 12.241600  1.101980
#> 167     386         13.34360 12.241600  1.101980
#> 168     386         13.34360 12.241600  1.101980
#> 
#> $result
#> [1] 1
#> 
#> $message
#> [1] "success"
#> 
# }