fetch Google Flu Trends data

Obtains estimates of inluenza activity based on volume of certain search queries from Google.

gft(locations, epiweeks)

Arguments

locations

A character string that specifies a location to be fetched.

epiweeks

Takes in the form epirange(startdate,enddate), where startdate and enddate are of the form YYYYMMDD (can be passed as string or numeric).

Value

an instance of epidata_call

Details

Google has discontinues Flu Trends, and this is now a static endpoint. Possibile input for locations can be found in https://github.com/cmu-delphi/delphi-epidata/blob/main/labels/regions.txt, https://github.com/cmu-delphi/delphi-epidata/blob/main/labels/states.txt, and https://github.com/cmu-delphi/delphi-epidata/blob/main/labels/cities.txt.

References

API Documentation: https://cmu-delphi.github.io/delphi-epidata/api/gft.html

Examples

call <- gft(locations = "hhs1", epiweeks = epirange(201201, 202001))
fetch_classic(call)
#> $epidata
#>     location    epiweek   num
#> 1       hhs1 2012-01-01  1567
#> 2       hhs1 2012-01-08  1339
#> 3       hhs1 2012-01-15  1220
#> 4       hhs1 2012-01-22  1271
#> 5       hhs1 2012-01-29  1181
#> 6       hhs1 2012-02-05   994
#> 7       hhs1 2012-02-12  1091
#> 8       hhs1 2012-02-19  1030
#> 9       hhs1 2012-02-26   856
#> 10      hhs1 2012-03-04   799
#> 11      hhs1 2012-03-11   788
#> 12      hhs1 2012-03-18   722
#> 13      hhs1 2012-03-25   620
#> 14      hhs1 2012-04-01   696
#> 15      hhs1 2012-04-08   806
#> 16      hhs1 2012-04-15   782
#> 17      hhs1 2012-04-22   641
#> 18      hhs1 2012-04-29   628
#> 19      hhs1 2012-05-06   690
#> 20      hhs1 2012-05-13   649
#> 21      hhs1 2012-05-20   598
#> 22      hhs1 2012-05-27   463
#> 23      hhs1 2012-06-03   483
#> 24      hhs1 2012-06-10   439
#> 25      hhs1 2012-06-17   410
#> 26      hhs1 2012-06-24   318
#> 27      hhs1 2012-07-01   280
#> 28      hhs1 2012-07-08   294
#> 29      hhs1 2012-07-15   279
#> 30      hhs1 2012-07-22   261
#> 31      hhs1 2012-07-29   264
#> 32      hhs1 2012-08-05   279
#> 33      hhs1 2012-08-12   293
#> 34      hhs1 2012-08-19   344
#> 35      hhs1 2012-08-26   364
#> 36      hhs1 2012-09-02   428
#> 37      hhs1 2012-09-09   553
#> 38      hhs1 2012-09-16   732
#> 39      hhs1 2012-09-23   926
#> 40      hhs1 2012-09-30   992
#> 41      hhs1 2012-10-07  1056
#> 42      hhs1 2012-10-14  1030
#> 43      hhs1 2012-10-21   988
#> 44      hhs1 2012-10-28   809
#> 45      hhs1 2012-11-04   999
#> 46      hhs1 2012-11-11  1183
#> 47      hhs1 2012-11-18  1201
#> 48      hhs1 2012-11-25  1498
#> 49      hhs1 2012-12-02  1983
#> 50      hhs1 2012-12-09  2656
#> 51      hhs1 2012-12-16  3797
#> 52      hhs1 2012-12-23  5758
#> 53      hhs1 2012-12-30  6675
#> 54      hhs1 2013-01-06 12461
#> 55      hhs1 2013-01-13 14211
#> 56      hhs1 2013-01-20  7412
#> 57      hhs1 2013-01-27  4763
#> 58      hhs1 2013-02-03  3134
#> 59      hhs1 2013-02-10  2120
#> 60      hhs1 2013-02-17  1905
#> 61      hhs1 2013-02-24  1672
#> 62      hhs1 2013-03-03  1423
#> 63      hhs1 2013-03-10  1343
#> 64      hhs1 2013-03-17  1297
#> 65      hhs1 2013-03-24  1357
#> 66      hhs1 2013-03-31  1305
#> 67      hhs1 2013-04-07  1125
#> 68      hhs1 2013-04-14   807
#> 69      hhs1 2013-04-21   694
#> 70      hhs1 2013-04-28   584
#> 71      hhs1 2013-05-05   491
#> 72      hhs1 2013-05-12   439
#> 73      hhs1 2013-05-19   392
#> 74      hhs1 2013-05-26   412
#> 75      hhs1 2013-06-02   369
#> 76      hhs1 2013-06-09   371
#> 77      hhs1 2013-06-16   357
#> 78      hhs1 2013-06-23   316
#> 79      hhs1 2013-06-30   315
#> 80      hhs1 2013-07-07   300
#> 81      hhs1 2013-07-14   303
#> 82      hhs1 2013-07-21   330
#> 83      hhs1 2013-07-28   362
#> 84      hhs1 2013-08-04   366
#> 85      hhs1 2013-08-11   317
#> 86      hhs1 2013-08-18   381
#> 87      hhs1 2013-08-25   343
#> 88      hhs1 2013-09-01   354
#> 89      hhs1 2013-09-08   450
#> 90      hhs1 2013-09-15   567
#> 91      hhs1 2013-09-22   582
#> 92      hhs1 2013-09-29   627
#> 93      hhs1 2013-10-06   562
#> 94      hhs1 2013-10-13   580
#> 95      hhs1 2013-10-20   579
#> 96      hhs1 2013-10-27   529
#> 97      hhs1 2013-11-03   654
#> 98      hhs1 2013-11-10   680
#> 99      hhs1 2013-11-17   669
#> 100     hhs1 2013-11-24   604
#> 101     hhs1 2013-12-01   753
#> 102     hhs1 2013-12-08   741
#> 103     hhs1 2013-12-15   897
#> 104     hhs1 2013-12-22  1512
#> 105     hhs1 2013-12-29  1918
#> 106     hhs1 2014-01-05  2114
#> 107     hhs1 2014-01-12  2294
#> 108     hhs1 2014-01-19  1896
#> 109     hhs1 2014-01-26  1967
#> 110     hhs1 2014-02-02  1811
#> 111     hhs1 2014-02-09  1683
#> 112     hhs1 2014-02-16  1281
#> 113     hhs1 2014-02-23  1133
#> 114     hhs1 2014-03-02  1127
#> 115     hhs1 2014-03-09  1111
#> 116     hhs1 2014-03-16  1024
#> 117     hhs1 2014-03-23  1202
#> 118     hhs1 2014-03-30  1182
#> 119     hhs1 2014-04-06  1289
#> 120     hhs1 2014-04-13  1224
#> 121     hhs1 2014-04-20  1102
#> 122     hhs1 2014-04-27   812
#> 123     hhs1 2014-05-04   700
#> 124     hhs1 2014-05-11   588
#> 125     hhs1 2014-05-18   480
#> 126     hhs1 2014-05-25   400
#> 127     hhs1 2014-06-01   491
#> 128     hhs1 2014-06-08   377
#> 129     hhs1 2014-06-15   530
#> 130     hhs1 2014-06-22   404
#> 131     hhs1 2014-06-29   381
#> 132     hhs1 2014-07-06   409
#> 133     hhs1 2014-07-13   367
#> 134     hhs1 2014-07-20   464
#> 135     hhs1 2014-07-27   352
#> 136     hhs1 2014-08-03   362
#> 137     hhs1 2014-08-10   411
#> 138     hhs1 2014-08-17   403
#> 139     hhs1 2014-08-24   326
#> 140     hhs1 2014-08-31   418
#> 141     hhs1 2014-09-07   540
#> 142     hhs1 2014-09-14   506
#> 143     hhs1 2014-09-21   616
#> 144     hhs1 2014-09-28   759
#> 145     hhs1 2014-10-05   971
#> 146     hhs1 2014-10-12  1118
#> 147     hhs1 2014-10-19   954
#> 148     hhs1 2014-10-26   779
#> 149     hhs1 2014-11-02   864
#> 150     hhs1 2014-11-09   706
#> 151     hhs1 2014-11-16   908
#> 152     hhs1 2014-11-23   861
#> 153     hhs1 2014-11-30   988
#> 154     hhs1 2014-12-07  1364
#> 155     hhs1 2014-12-14  1831
#> 156     hhs1 2014-12-21  2778
#> 157     hhs1 2014-12-28  2904
#> 158     hhs1 2015-01-04  2433
#> 159     hhs1 2015-01-11  3376
#> 160     hhs1 2015-01-18  4789
#> 161     hhs1 2015-01-25  3998
#> 162     hhs1 2015-02-01  2850
#> 163     hhs1 2015-02-08  2660
#> 164     hhs1 2015-02-15  2597
#> 165     hhs1 2015-02-22  1829
#> 166     hhs1 2015-03-01  1881
#> 167     hhs1 2015-03-08  2111
#> 168     hhs1 2015-03-15  1848
#> 169     hhs1 2015-03-22  1508
#> 170     hhs1 2015-03-29  1584
#> 171     hhs1 2015-04-05  1597
#> 172     hhs1 2015-04-12  1231
#> 173     hhs1 2015-04-19   696
#> 174     hhs1 2015-04-26  1102
#> 175     hhs1 2015-05-03   950
#> 176     hhs1 2015-05-10   880
#> 177     hhs1 2015-05-17   490
#> 178     hhs1 2015-05-24   924
#> 179     hhs1 2015-05-31   281
#> 180     hhs1 2015-06-07   870
#> 181     hhs1 2015-06-14   504
#> 182     hhs1 2015-06-21   448
#> 183     hhs1 2015-06-28   732
#> 184     hhs1 2015-07-05   899
#> 185     hhs1 2015-07-12   581
#> 186     hhs1 2015-07-19   487
#> 187     hhs1 2015-07-26   454
#> 188     hhs1 2015-08-02   323
#> 189     hhs1 2015-08-09   487
#> 
#> $result
#> [1] 1
#> 
#> $message
#> [1] "success"
#>