Fetch NIDSS flu data

Obtains information on outpatient inluenza-like-illness from Taiwan National Infectious Disease Statistical System.

nidss_flu(regions, epiweeks, issues = NULL, lag = NULL)

Arguments

regions

character vector. The regions to fetch.

epiweeks

epirange. The epiweeks to fetch.

issues

epirange. Optionally, the issues to fetch. If not set, the most recent issue is returned. Mutually exclusive with lag.

lag

integer. Optionally, the lag of the issues to fetch. If not set, the most recent issue is returned. Mutually exclusive with issues.

Value

epidata_call

Details

API docs: https://cmu-delphi.github.io/delphi-epidata/api/nidss_flu.html

Examples

# \donttest{
nidss_flu(regions = "taipei", epiweeks = epirange(201201, 202001)) %>% fetch_classic()
#> $epidata
#>     release_date region issue    epiweek lag visits  ili
#> 1     2015-08-04 Taipei  <NA> 2012-01-01 186  31836 1.92
#> 2     2015-08-04 Taipei  <NA> 2012-01-08 185  37001 2.12
#> 3     2015-08-04 Taipei  <NA> 2012-01-15 184  40658 1.99
#> 4     2015-08-04 Taipei  <NA> 2012-01-22 183  13552 2.73
#> 5     2015-08-04 Taipei  <NA> 2012-01-29 182  27974 1.68
#> 6     2015-08-04 Taipei  <NA> 2012-02-05 181  23039 1.53
#> 7     2015-08-04 Taipei  <NA> 2012-02-12 180  22780 1.43
#> 8     2015-08-04 Taipei  <NA> 2012-02-19 179  21102 1.36
#> 9     2015-08-04 Taipei  <NA> 2012-02-26 178  18484 1.27
#> 10    2015-08-04 Taipei  <NA> 2012-03-04 177  18939 1.17
#> 11    2015-08-04 Taipei  <NA> 2012-03-11 176  18527 1.15
#> 12    2015-08-04 Taipei  <NA> 2012-03-18 175  18626 1.18
#> 13    2015-08-04 Taipei  <NA> 2012-03-25 174  20113 1.18
#> 14    2015-08-04 Taipei  <NA> 2012-04-01 173  16555 1.08
#> 15    2015-08-04 Taipei  <NA> 2012-04-08 172  15191 0.97
#> 16    2015-08-04 Taipei  <NA> 2012-04-15 171  11995 0.98
#> 17    2015-08-04 Taipei  <NA> 2012-04-22 170  11381 0.94
#> 18    2015-08-04 Taipei  <NA> 2012-04-29 169  13401 0.89
#> 19    2015-08-04 Taipei  <NA> 2012-05-06 168  14026 0.90
#> 20    2015-08-04 Taipei  <NA> 2012-05-13 167  13116 0.89
#> 21    2015-08-04 Taipei  <NA> 2012-05-20 166  13782 0.91
#> 22    2015-08-04 Taipei  <NA> 2012-05-27 165  12767 0.86
#> 23    2015-08-04 Taipei  <NA> 2012-06-03 164  12910 0.85
#> 24    2015-08-04 Taipei  <NA> 2012-06-10 163  12445 0.87
#> 25    2015-08-04 Taipei  <NA> 2012-06-17 162  12489 0.94
#> 26    2015-08-04 Taipei  <NA> 2012-06-24 161  14435 0.94
#> 27    2015-08-04 Taipei  <NA> 2012-07-01 160  14830 0.98
#> 28    2015-08-04 Taipei  <NA> 2012-07-08 159  14191 0.98
#> 29    2015-08-04 Taipei  <NA> 2012-07-15 158  12661 0.94
#> 30    2015-08-04 Taipei  <NA> 2012-07-22 157  12252 0.87
#> 31    2015-08-04 Taipei  <NA> 2012-07-29 156  10930 0.87
#> 32    2015-08-04 Taipei  <NA> 2012-08-05 155  11139 0.78
#> 33    2015-08-04 Taipei  <NA> 2012-08-12 154  10707 0.76
#> 34    2015-08-04 Taipei  <NA> 2012-08-19 153   9510 0.76
#> 35    2015-08-04 Taipei  <NA> 2012-08-26 152  10265 0.76
#> 36    2015-08-04 Taipei  <NA> 2012-09-02 151   9987 0.73
#> 37    2015-08-04 Taipei  <NA> 2012-09-09 150  11622 0.77
#> 38    2015-08-04 Taipei  <NA> 2012-09-16 149  12481 0.79
#> 39    2015-08-04 Taipei  <NA> 2012-09-23 148  11804 0.79
#> 40    2015-08-04 Taipei  <NA> 2012-09-30 147  12876 0.78
#> 41    2015-08-04 Taipei  <NA> 2012-10-07 146  11863 0.80
#> 42    2015-08-04 Taipei  <NA> 2012-10-14 145  12169 0.77
#> 43    2015-08-04 Taipei  <NA> 2012-10-21 144  11403 0.75
#> 44    2015-08-04 Taipei  <NA> 2012-10-28 143  11017 0.74
#> 45    2015-08-04 Taipei  <NA> 2012-11-04 142  11939 0.76
#> 46    2015-08-04 Taipei  <NA> 2012-11-11 141  12219 0.79
#> 47    2015-08-04 Taipei  <NA> 2012-11-18 140  12546 0.82
#> 48    2015-08-04 Taipei  <NA> 2012-11-25 139  11492 0.80
#> 49    2015-08-04 Taipei  <NA> 2012-12-02 138  12115 0.79
#> 50    2015-08-04 Taipei  <NA> 2012-12-09 137  12032 0.80
#> 51    2015-08-04 Taipei  <NA> 2012-12-16 136  12810 0.83
#> 52    2015-08-04 Taipei  <NA> 2012-12-23 135  13929 0.85
#> 53    2015-08-04 Taipei  <NA> 2012-12-30 134  12715 0.93
#> 54    2015-08-04 Taipei  <NA> 2013-01-06 133  14107 0.90
#> 55    2015-08-04 Taipei  <NA> 2013-01-13 132  14720 0.93
#> 56    2015-08-04 Taipei  <NA> 2013-01-20 131  15601 0.95
#> 57    2015-08-04 Taipei  <NA> 2013-01-27 130  16313 0.96
#> 58    2015-08-04 Taipei  <NA> 2013-02-03 129  17160 0.95
#> 59    2015-08-04 Taipei  <NA> 2013-02-10 128   7407 1.27
#> 60    2015-08-04 Taipei  <NA> 2013-02-17 127  14342 0.93
#> 61    2015-08-04 Taipei  <NA> 2013-02-24 126  13059 0.97
#> 62    2015-08-04 Taipei  <NA> 2013-03-03 125  16661 0.98
#> 63    2015-08-04 Taipei  <NA> 2013-03-10 124  16838 1.04
#> 64    2015-08-04 Taipei  <NA> 2013-03-17 123  16051 1.04
#> 65    2015-08-04 Taipei  <NA> 2013-03-24 122  15273 1.06
#> 66    2015-08-04 Taipei  <NA> 2013-03-31 121  15373 1.11
#> 67    2015-08-04 Taipei  <NA> 2013-04-07 120  17113 1.04
#> 68    2015-08-04 Taipei  <NA> 2013-04-14 119  15580 1.00
#> 69    2015-08-04 Taipei  <NA> 2013-04-21 118  15704 0.99
#> 70    2015-08-04 Taipei  <NA> 2013-04-28 117  16428 1.01
#> 71    2015-08-04 Taipei  <NA> 2013-05-05 116  13089 0.96
#> 72    2015-08-04 Taipei  <NA> 2013-05-12 115  11900 0.87
#> 73    2015-08-04 Taipei  <NA> 2013-05-19 114  13089 0.87
#> 74    2015-08-04 Taipei  <NA> 2013-05-26 113  13172 0.86
#> 75    2015-08-04 Taipei  <NA> 2013-06-02 112  12629 0.84
#> 76    2015-08-04 Taipei  <NA> 2013-06-09 111  11835 0.85
#> 77    2015-08-04 Taipei  <NA> 2013-06-16 110  12161 0.80
#> 78    2015-08-04 Taipei  <NA> 2013-06-23 109  11502 0.78
#> 79    2015-08-04 Taipei  <NA> 2013-06-30 108  10886 0.76
#> 80    2015-08-04 Taipei  <NA> 2013-07-07 107   9936 0.75
#> 81    2015-08-04 Taipei  <NA> 2013-07-14 106  10187 0.68
#> 82    2015-08-04 Taipei  <NA> 2013-07-21 105   9580 0.67
#> 83    2015-08-04 Taipei  <NA> 2013-07-28 104   9474 0.67
#> 84    2015-08-04 Taipei  <NA> 2013-08-04 103   9566 0.67
#> 85    2015-08-04 Taipei  <NA> 2013-08-11 102   9368 0.66
#> 86    2015-08-04 Taipei  <NA> 2013-08-18 101   8602 0.67
#> 87    2015-08-04 Taipei  <NA> 2013-08-25 100   9579 0.66
#> 88    2015-08-04 Taipei  <NA> 2013-09-01  99  10218 0.67
#> 89    2015-08-04 Taipei  <NA> 2013-09-08  98  10567 0.69
#> 90    2015-08-04 Taipei  <NA> 2013-09-15  97   9684 0.75
#> 91    2015-08-04 Taipei  <NA> 2013-09-22  96  11071 0.69
#> 92    2015-08-04 Taipei  <NA> 2013-09-29  95  11072 0.71
#> 93    2015-08-04 Taipei  <NA> 2013-10-06  94  10808 0.74
#> 94    2015-08-04 Taipei  <NA> 2013-10-13  93  11901 0.73
#> 95    2015-08-04 Taipei  <NA> 2013-10-20  92  12350 0.77
#> 96    2015-08-04 Taipei  <NA> 2013-10-27  91  12719 0.78
#> 97    2015-08-04 Taipei  <NA> 2013-11-03  90  11969 0.77
#> 98    2015-08-04 Taipei  <NA> 2013-11-10  89  11298 0.74
#> 99    2015-08-04 Taipei  <NA> 2013-11-17  88  12661 0.78
#> 100   2015-08-04 Taipei  <NA> 2013-11-24  87  12312 0.79
#> 101   2015-08-04 Taipei  <NA> 2013-12-01  86  13912 0.83
#> 102   2015-08-04 Taipei  <NA> 2013-12-08  85  13856 0.87
#> 103   2015-08-04 Taipei  <NA> 2013-12-15  84  13394 0.91
#> 104   2015-08-04 Taipei  <NA> 2013-12-22  83  15468 1.00
#> 105   2015-08-04 Taipei  <NA> 2013-12-29  82  19984 1.25
#> 106   2015-08-04 Taipei  <NA> 2014-01-05  81  24895 1.40
#> 107   2015-08-04 Taipei  <NA> 2014-01-12  80  26901 1.56
#> 108   2015-08-04 Taipei  <NA> 2014-01-19  79  31442 1.62
#> 109   2015-08-04 Taipei  <NA> 2014-01-26  78  22722 1.68
#> 110   2015-08-04 Taipei  <NA> 2014-02-02  77  19883 1.48
#> 111   2015-08-04 Taipei  <NA> 2014-02-09  76  19267 1.26
#> 112   2015-08-04 Taipei  <NA> 2014-02-16  75  21537 1.30
#> 113   2015-08-04 Taipei  <NA> 2014-02-23  74  22184 1.39
#> 114   2015-08-04 Taipei  <NA> 2014-03-02  73  20477 1.25
#> 115   2015-08-04 Taipei  <NA> 2014-03-09  72  20459 1.21
#> 116   2015-08-04 Taipei  <NA> 2014-03-16  71  20647 1.20
#> 117   2015-08-04 Taipei  <NA> 2014-03-23  70  19914 1.19
#> 118   2015-08-04 Taipei  <NA> 2014-03-30  69  17565 1.19
#> 119   2015-08-04 Taipei  <NA> 2014-04-06  68  18498 1.07
#> 120   2015-08-04 Taipei  <NA> 2014-04-13  67  17003 1.02
#> 121   2015-08-04 Taipei  <NA> 2014-04-20  66  15472 0.99
#> 122   2015-08-04 Taipei  <NA> 2014-04-27  65  14797 0.95
#> 123   2015-08-04 Taipei  <NA> 2014-05-04  64  14666 0.91
#> 124   2015-08-04 Taipei  <NA> 2014-05-11  63  14511 0.91
#> 125   2015-08-04 Taipei  <NA> 2014-05-18  62  13334 0.90
#> 126   2015-08-04 Taipei  <NA> 2014-05-25  61  13482 0.86
#> 127   2015-08-04 Taipei  <NA> 2014-06-01  60  12154 0.82
#> 128   2015-08-04 Taipei  <NA> 2014-06-08  59  12832 0.79
#> 129   2015-08-04 Taipei  <NA> 2014-06-15  58  12037 0.78
#> 130   2015-08-04 Taipei  <NA> 2014-06-22  57  11781 0.78
#> 131   2015-08-04 Taipei  <NA> 2014-06-29  56  11363 0.75
#> 132   2015-08-04 Taipei  <NA> 2014-07-06  55  11272 0.75
#> 133   2015-08-04 Taipei  <NA> 2014-07-13  54  11194 0.76
#> 134   2015-08-04 Taipei  <NA> 2014-07-20  53  10483 0.79
#> 135   2015-08-04 Taipei  <NA> 2014-07-27  52  10382 0.72
#> 136   2015-08-04 Taipei  <NA> 2014-08-03  51  10621 0.72
#> 137   2015-08-11 Taipei  <NA> 2014-08-10  51  10736 0.73
#> 138   2015-08-18 Taipei  <NA> 2014-08-17  51  11194 0.75
#> 139   2015-08-25 Taipei  <NA> 2014-08-24  51  10729 0.70
#> 140   2015-09-01 Taipei  <NA> 2014-08-31  51  10937 0.70
#> 141   2015-09-09 Taipei  <NA> 2014-09-07  51  10060 0.72
#> 142   2015-09-15 Taipei  <NA> 2014-09-14  51  10992 0.71
#> 143   2015-09-22 Taipei  <NA> 2014-09-21  51  10540 0.71
#> 144   2015-09-29 Taipei  <NA> 2014-09-28  51  10958 0.69
#> 145   2015-09-29 Taipei  <NA> 2014-10-05  50  10317 0.72
#> 146   2015-10-13 Taipei  <NA> 2014-10-12  51  11899 0.73
#> 147   2015-10-20 Taipei  <NA> 2014-10-19  51  11026 0.72
#> 148   2015-10-27 Taipei  <NA> 2014-10-26  51  10942 0.71
#> 149   2015-11-03 Taipei  <NA> 2014-11-02  51  10715 0.70
#> 150   2015-11-10 Taipei  <NA> 2014-11-09  51   9308 0.71
#> 151   2015-11-17 Taipei  <NA> 2014-11-16  51  10935 0.73
#> 152   2015-11-24 Taipei  <NA> 2014-11-23  51  11715 0.76
#> 153   2015-12-01 Taipei  <NA> 2014-11-30  51  11266 0.75
#> 154   2015-12-08 Taipei  <NA> 2014-12-07  51  11642 0.77
#> 155   2015-12-15 Taipei  <NA> 2014-12-14  51  10320 0.78
#> 156   2015-12-22 Taipei  <NA> 2014-12-21  51  11249 0.85
#> 157   2015-12-29 Taipei  <NA> 2014-12-28  51  13713 0.96
#> 158   2016-01-05 Taipei  <NA> 2015-01-04  51  16698 0.96
#> 159   2016-01-12 Taipei  <NA> 2015-01-11  51  17146 1.03
#> 160   2016-01-19 Taipei  <NA> 2015-01-18  51  17817 1.07
#> 161   2016-01-26 Taipei  <NA> 2015-01-25  51  18168 1.13
#> 162   2016-02-02 Taipei  <NA> 2015-02-01  51  18321 1.11
#> 163   2016-02-02 Taipei  <NA> 2015-02-08  50  19799 1.08
#> 164   2016-02-02 Taipei  <NA> 2015-02-15  49  11879 1.25
#> 165   2016-02-02 Taipei  <NA> 2015-02-22  48  17397 1.19
#> 166   2016-02-02 Taipei  <NA> 2015-03-01  47  17849 1.07
#> 167   2016-02-02 Taipei  <NA> 2015-03-08  46  17441 1.08
#> 168   2016-02-02 Taipei  <NA> 2015-03-15  45  19602 1.16
#> 169   2016-02-02 Taipei  <NA> 2015-03-22  44  19486 1.22
#> 170   2016-02-02 Taipei  <NA> 2015-03-29  43  20278 1.28
#> 171   2016-02-02 Taipei  <NA> 2015-04-05  42  20101 1.22
#> 172   2016-02-02 Taipei  <NA> 2015-04-12  41  17121 1.22
#> 173   2016-02-02 Taipei  <NA> 2015-04-19  40  20012 1.23
#> 174   2016-02-02 Taipei  <NA> 2015-04-26  39  19555 1.20
#> 175   2016-02-02 Taipei  <NA> 2015-05-03  38  17579 1.09
#> 176   2016-02-02 Taipei  <NA> 2015-05-10  37  17169 1.09
#> 177   2016-02-02 Taipei  <NA> 2015-05-17  36  16811 1.09
#> 178   2016-02-02 Taipei  <NA> 2015-05-24  35  16877 1.07
#> 179   2016-02-02 Taipei  <NA> 2015-05-31  34  17403 1.07
#> 180   2016-02-02 Taipei  <NA> 2015-06-07  33  16855 1.05
#> 181   2016-02-02 Taipei  <NA> 2015-06-14  32  14886 1.07
#> 182   2016-02-02 Taipei  <NA> 2015-06-21  31  15050 0.97
#> 183   2016-02-02 Taipei  <NA> 2015-06-28  30  13495 0.90
#> 184   2016-02-02 Taipei  <NA> 2015-07-05  29  11872 0.88
#> 185   2016-02-02 Taipei  <NA> 2015-07-12  28  11691 0.78
#> 186   2016-07-19 Taipei  <NA> 2015-07-19  51  10815 0.75
#> 187   2016-07-25 Taipei  <NA> 2015-07-26  51  10848 0.76
#> 188   2016-08-01 Taipei  <NA> 2015-08-02  51   9323 0.74
#> 189   2016-08-08 Taipei  <NA> 2015-08-09  51  10473 0.71
#> 190   2016-08-15 Taipei  <NA> 2015-08-16  51   9993 0.69
#> 191   2016-08-22 Taipei  <NA> 2015-08-23  51  10008 0.71
#> 192   2016-08-30 Taipei  <NA> 2015-08-30  51   8641 0.70
#> 193   2016-09-05 Taipei  <NA> 2015-09-06  51  11516 0.73
#> 194   2016-09-12 Taipei  <NA> 2015-09-13  51  12068 0.76
#> 195   2016-09-19 Taipei  <NA> 2015-09-20  51  12443 0.77
#> 196   2016-09-26 Taipei  <NA> 2015-09-27  51  10438 0.76
#> 197   2016-10-03 Taipei  <NA> 2015-10-04  51  11061 0.74
#> 198   2016-10-10 Taipei  <NA> 2015-10-11  51  12744 0.76
#> 199   2016-10-17 Taipei  <NA> 2015-10-18  51  11771 0.76
#> 200   2016-10-24 Taipei  <NA> 2015-10-25  51  11645 0.76
#> 201   2016-10-31 Taipei  <NA> 2015-11-01  51  11934 0.76
#> 202   2016-11-07 Taipei  <NA> 2015-11-08  51  12227 0.78
#> 203   2016-11-14 Taipei  <NA> 2015-11-15  51  12764 0.83
#> 204   2016-11-21 Taipei  <NA> 2015-11-22  51  12741 0.82
#> 205   2016-11-28 Taipei  <NA> 2015-11-29  51  12850 0.83
#> 206   2016-12-05 Taipei  <NA> 2015-12-06  51  13884 0.89
#> 207   2016-12-12 Taipei  <NA> 2015-12-13  51  14251 0.90
#> 208   2016-12-19 Taipei  <NA> 2015-12-20  51  15404 0.97
#> 209   2016-12-26 Taipei  <NA> 2015-12-27  51  21459 1.53
#> 210   2018-01-01 Taipei  <NA> 2016-01-03 103  15545 0.96
#> 211   2018-01-01 Taipei  <NA> 2016-01-10 102  16244 1.04
#> 212   2018-01-01 Taipei  <NA> 2016-01-17 101  18985 1.21
#> 213   2018-01-01 Taipei  <NA> 2016-01-24 100  23503 1.36
#> 214   2018-01-01 Taipei  <NA> 2016-01-31  99  32923 1.55
#> 215   2018-01-01 Taipei  <NA> 2016-02-07  98  14944 2.59
#> 216   2018-01-01 Taipei  <NA> 2016-02-14  97  36468 1.88
#> 217   2018-01-01 Taipei  <NA> 2016-02-21  96  39657 2.18
#> 218   2018-01-01 Taipei  <NA> 2016-02-28  95  39561 2.10
#> 219   2018-01-01 Taipei  <NA> 2016-03-06  94  36668 2.07
#> 220   2018-01-01 Taipei  <NA> 2016-03-13  93  31792 1.92
#> 221   2018-01-01 Taipei  <NA> 2016-03-20  92  26674 1.76
#> 222   2018-01-01 Taipei  <NA> 2016-03-27  91  27377 1.54
#> 223   2018-01-01 Taipei  <NA> 2016-04-03  90  20809 1.42
#> 224   2018-01-01 Taipei  <NA> 2016-04-10  89  19036 1.14
#> 225   2018-01-01 Taipei  <NA> 2016-04-17  88  17085 1.04
#> 226   2018-01-01 Taipei  <NA> 2016-04-24  87  15943 0.96
#> 227   2018-01-01 Taipei  <NA> 2016-05-01  86  13407 0.84
#> 228   2018-01-01 Taipei  <NA> 2016-05-08  85  12197 0.74
#> 229   2018-01-01 Taipei  <NA> 2016-05-15  84  11262 0.71
#> 230   2018-01-01 Taipei  <NA> 2016-05-22  83  10901 0.69
#> 231   2018-01-01 Taipei  <NA> 2016-05-29  82   9825 0.63
#> 232   2018-01-01 Taipei  <NA> 2016-06-05  81   8722 0.65
#> 233   2018-01-01 Taipei  <NA> 2016-06-12  80   9090 0.58
#> 234   2018-01-01 Taipei  <NA> 2016-06-19  79   8818 0.57
#> 235   2018-01-01 Taipei  <NA> 2016-06-26  78   8515 0.57
#> 236   2018-01-01 Taipei  <NA> 2016-07-03  77   7655 0.57
#> 237   2018-01-01 Taipei  <NA> 2016-07-10  76   8166 0.53
#> 238   2018-01-01 Taipei  <NA> 2016-07-17  75   8094 0.54
#> 239   2018-01-01 Taipei  <NA> 2016-07-24  74   7783 0.54
#> 240   2018-01-01 Taipei  <NA> 2016-07-31  73   7763 0.52
#> 241   2018-01-01 Taipei  <NA> 2016-08-07  72   7963 0.53
#> 242   2018-01-01 Taipei  <NA> 2016-08-14  71   7818 0.52
#> 243   2018-01-01 Taipei  <NA> 2016-08-21  70   8025 0.53
#> 244   2018-01-01 Taipei  <NA> 2016-08-28  69   8266 0.54
#> 245   2018-01-01 Taipei  <NA> 2016-09-04  68   7880 0.51
#> 246   2018-01-01 Taipei  <NA> 2016-09-11  67   7265 0.59
#> 247   2018-01-01 Taipei  <NA> 2016-09-18  66   9154 0.53
#> 248   2018-01-01 Taipei  <NA> 2016-09-25  65   7680 0.57
#> 249   2018-01-01 Taipei  <NA> 2016-10-02  64   8935 0.53
#> 250   2018-01-01 Taipei  <NA> 2016-10-09  63   8478 0.59
#> 251   2018-01-01 Taipei  <NA> 2016-10-16  62   9645 0.60
#> 252   2018-01-01 Taipei  <NA> 2016-10-23  61  11350 0.70
#> 253   2018-01-01 Taipei  <NA> 2016-10-30  60  13047 0.79
#> 254   2018-01-01 Taipei  <NA> 2016-11-06  59  13153 0.81
#> 255   2018-01-01 Taipei  <NA> 2016-11-13  58  14045 0.84
#> 256   2018-01-01 Taipei  <NA> 2016-11-20  57  13597 0.86
#> 257   2018-01-01 Taipei  <NA> 2016-11-27  56  14130 0.85
#> 258   2018-01-01 Taipei  <NA> 2016-12-04  55  14242 0.87
#> 259   2018-01-01 Taipei  <NA> 2016-12-11  54  14398 0.89
#> 260   2018-01-01 Taipei  <NA> 2016-12-18  53  14085 0.87
#> 261   2018-01-01 Taipei  <NA> 2016-12-25  52  13844 0.83
#> 262   2018-03-12 Taipei  <NA> 2017-01-01  61  13063 0.84
#> 263   2018-03-12 Taipei  <NA> 2017-01-08  60  12558 0.79
#> 264   2018-03-12 Taipei  <NA> 2017-01-15  59  13033 0.77
#> 265   2018-03-12 Taipei  <NA> 2017-01-22  58  12520 0.81
#> 266   2018-03-12 Taipei  <NA> 2017-01-29  57   8534 0.85
#> 267   2018-03-12 Taipei  <NA> 2017-02-05  56  12074 0.72
#> 268   2018-03-12 Taipei  <NA> 2017-02-12  55  11906 0.72
#> 269   2018-03-12 Taipei  <NA> 2017-02-19  54  11839 0.74
#> 270   2018-03-12 Taipei  <NA> 2017-02-26  53  10785 0.73
#> 271   2018-03-12 Taipei  <NA> 2017-03-05  52  11264 0.69
#> 272   2018-03-12 Taipei  <NA> 2017-03-12  51  12422 0.75
#> 273   2018-03-12 Taipei  <NA> 2017-03-19  50  13285 0.79
#> 274   2018-03-12 Taipei  <NA> 2017-03-26  49  14072 0.83
#> 275   2018-03-12 Taipei  <NA> 2017-04-02  48  12076 0.81
#> 276   2018-03-12 Taipei  <NA> 2017-04-09  47  12301 0.72
#> 277   2018-03-12 Taipei  <NA> 2017-04-16  46  12252 0.76
#> 278   2018-03-12 Taipei  <NA> 2017-04-23  45  13471 0.84
#> 279   2018-03-12 Taipei  <NA> 2017-04-30  44  13744 0.83
#> 280   2018-03-12 Taipei  <NA> 2017-05-07  43  14196 0.86
#> 281   2018-03-12 Taipei  <NA> 2017-05-14  42  15509 0.96
#> 282   2018-03-12 Taipei  <NA> 2017-05-21  41  16896 1.04
#> 283   2018-03-12 Taipei  <NA> 2017-05-28  40  16996 1.27
#> 284   2018-03-12 Taipei  <NA> 2017-06-04  39  20627 1.24
#> 285   2018-03-12 Taipei  <NA> 2017-06-11  38  23005 1.48
#> 286   2018-03-12 Taipei  <NA> 2017-06-18  37  30730 1.85
#> 287   2018-03-12 Taipei  <NA> 2017-06-25  36  33730 2.09
#> 288   2018-03-12 Taipei  <NA> 2017-07-02  35  31380 1.91
#> 289   2018-03-12 Taipei  <NA> 2017-07-09  34  29000 1.80
#> 290   2018-03-12 Taipei  <NA> 2017-07-16  33  25470 1.63
#> 291   2018-03-12 Taipei  <NA> 2017-07-23  32  19667 1.31
#> 292   2018-03-12 Taipei  <NA> 2017-07-30  31  16065 1.06
#> 293   2018-03-12 Taipei  <NA> 2017-08-06  30  14068 0.93
#> 294   2018-03-12 Taipei  <NA> 2017-08-13  29  11637 0.77
#> 295   2018-03-12 Taipei  <NA> 2017-08-20  28  10423 0.71
#> 296   2018-03-12 Taipei  <NA> 2017-08-27  27   9879 0.66
#> 297   2018-03-12 Taipei  <NA> 2017-09-03  26   9870 0.64
#> 298   2018-03-12 Taipei  <NA> 2017-09-10  25  10090 0.64
#> 299   2018-03-12 Taipei  <NA> 2017-09-17  24  10391 0.64
#> 300   2018-03-12 Taipei  <NA> 2017-09-24  23  10026 0.62
#> 301   2018-03-12 Taipei  <NA> 2017-10-01  22   9319 0.61
#> 302   2018-03-12 Taipei  <NA> 2017-10-08  21   8087 0.62
#> 303   2018-03-12 Taipei  <NA> 2017-10-15  20   9010 0.56
#> 304   2018-03-12 Taipei  <NA> 2017-10-22  19  10496 0.63
#> 305   2018-03-12 Taipei  <NA> 2017-10-29  18  10600 0.63
#> 306   2018-03-12 Taipei  <NA> 2017-11-05  17  11103 0.67
#> 307   2018-03-12 Taipei  <NA> 2017-11-12  16  10998 0.69
#> 308   2018-03-12 Taipei  <NA> 2017-11-19  15  11541 0.73
#> 309   2018-03-12 Taipei  <NA> 2017-11-26  14  12873 0.78
#> 310   2018-03-12 Taipei  <NA> 2017-12-03  13  14404 0.89
#> 311   2018-03-12 Taipei  <NA> 2017-12-10  12  17101 1.02
#> 312   2018-03-12 Taipei  <NA> 2017-12-17  11  21401 1.26
#> 313   2018-03-12 Taipei  <NA> 2017-12-24  10  28742 1.58
#> 314   2018-03-12 Taipei  <NA> 2017-12-31   9  29452 1.85
#> 315   2018-03-12 Taipei  <NA> 2018-01-07   8  28854 1.74
#> 316   2018-03-12 Taipei  <NA> 2018-01-14   7  35775 2.03
#> 317   2018-03-12 Taipei  <NA> 2018-01-21   6  34517 1.98
#> 318   2018-03-12 Taipei  <NA> 2018-01-28   5  29093 1.82
#> 319   2018-03-12 Taipei  <NA> 2018-02-04   4  28327 1.64
#> 320   2018-03-12 Taipei  <NA> 2018-02-11   3  23024 1.79
#> 321   2018-03-12 Taipei  <NA> 2018-02-18   2  25857 1.80
#> 322   2018-03-12 Taipei  <NA> 2018-02-25   1  27138 1.59
#> 323   2018-03-12 Taipei  <NA> 2018-03-04   0  24911 1.40
#> 
#> $result
#> [1] 1
#> 
#> $message
#> [1] "success"
#> 
# }