Fetch PAHO Dengue

API docs: https://cmu-delphi.github.io/delphi-epidata/api/paho_dengue.html

paho_dengue(regions, epiweeks, issues = NULL, lag = NULL)

Arguments

regions

character vector. The regions to fetch.

epiweeks

timeset. The epiweeks to fetch.

issues

timeset. Optionally, the issues to fetch. If not set, the most recent issue is returned. Mutually exclusive with lag.

lag

integer. Optionally, the lag of the issues to fetch. If not set, the most recent issue is returned. Mutually exclusive with issues.

Value

epidata_call

Examples

# \donttest{
paho_dengue(regions = "ca", epiweeks = epirange(201401, 201501)) %>% fetch_classic()
#> $epidata
#>    release_date region serotype      issue    epiweek lag total_pop num_dengue
#> 1    2020-08-07     CA          2020-08-02 2013-12-29 344         0          0
#> 2    2020-08-07     CA          2020-08-02 2014-01-05 343         0          0
#> 3    2020-08-07     CA          2020-08-02 2014-01-12 342         0          0
#> 4    2020-08-07     CA          2020-08-02 2014-01-19 341         0          0
#> 5    2020-08-07     CA          2020-08-02 2014-01-26 340         0          0
#> 6    2020-08-07     CA          2020-08-02 2014-02-02 339         0          0
#> 7    2020-08-07     CA          2020-08-02 2014-02-09 338         0          0
#> 8    2020-08-07     CA          2020-08-02 2014-02-16 337         0          0
#> 9    2020-08-07     CA          2020-08-02 2014-02-23 336         0          0
#> 10   2020-08-07     CA          2020-08-02 2014-03-02 335         0          0
#> 11   2020-08-07     CA          2020-08-02 2014-03-09 334         0          0
#> 12   2020-08-07     CA          2020-08-02 2014-03-16 333         0          0
#> 13   2020-08-07     CA          2020-08-02 2014-03-23 332         0          0
#> 14   2020-08-07     CA          2020-08-02 2014-03-30 331         0          0
#> 15   2020-08-07     CA          2020-08-02 2014-04-06 330         0          0
#> 16   2020-08-07     CA          2020-08-02 2014-04-13 329         0          0
#> 17   2020-08-07     CA          2020-08-02 2014-04-20 328         0          0
#> 18   2020-08-07     CA          2020-08-02 2014-04-27 327         0          0
#> 19   2020-08-07     CA          2020-08-02 2014-05-04 326         0          0
#> 20   2020-08-07     CA          2020-08-02 2014-05-11 325         0          0
#> 21   2020-08-07     CA          2020-08-02 2014-05-18 324         0          0
#> 22   2020-08-07     CA          2020-08-02 2014-05-25 323         0          0
#> 23   2020-08-07     CA          2020-08-02 2014-06-01 322         0          0
#> 24   2020-08-07     CA          2020-08-02 2014-06-08 321         0          0
#> 25   2020-08-07     CA          2020-08-02 2014-06-15 320         0          0
#> 26   2020-08-07     CA          2020-08-02 2014-06-22 319         0          0
#> 27   2020-08-07     CA          2020-08-02 2014-06-29 318         0          0
#> 28   2020-08-07     CA          2020-08-02 2014-07-06 317         0          0
#> 29   2020-08-07     CA          2020-08-02 2014-07-13 316         0          0
#> 30   2020-08-07     CA          2020-08-02 2014-07-20 315         0          0
#> 31   2020-08-07     CA          2020-08-02 2014-07-27 314         0          0
#> 32   2020-08-07     CA          2020-08-02 2014-08-03 313         0          0
#> 33   2020-08-07     CA          2020-08-02 2014-08-10 312         0          0
#> 34   2020-08-07     CA          2020-08-02 2014-08-17 311         0          0
#> 35   2020-08-07     CA          2020-08-02 2014-08-24 310         0          0
#> 36   2020-08-07     CA          2020-08-02 2014-08-31 309         0          0
#> 37   2020-08-07     CA          2020-08-02 2014-09-07 308         0          0
#> 38   2020-08-07     CA          2020-08-02 2014-09-14 307         0          0
#> 39   2020-08-07     CA          2020-08-02 2014-09-21 306         0          0
#> 40   2020-08-07     CA          2020-08-02 2014-09-28 305         0          0
#> 41   2020-08-07     CA          2020-08-02 2014-10-05 304         0          0
#> 42   2020-08-07     CA          2020-08-02 2014-10-12 303         0          0
#> 43   2020-08-07     CA          2020-08-02 2014-10-19 302         0          0
#> 44   2020-08-07     CA          2020-08-02 2014-10-26 301         0          0
#> 45   2020-08-07     CA          2020-08-02 2014-11-02 300         0          0
#> 46   2020-08-07     CA          2020-08-02 2014-11-09 299         0          0
#> 47   2020-08-07     CA          2020-08-02 2014-11-16 298         0          0
#> 48   2020-08-07     CA          2020-08-02 2014-11-23 297         0          0
#> 49   2020-08-07     CA          2020-08-02 2014-11-30 296         0          0
#> 50   2020-08-07     CA          2020-08-02 2014-12-07 295         0          0
#> 51   2020-08-07     CA          2020-08-02 2014-12-14 294         0          0
#> 52   2019-01-19     CA          2019-01-13 2014-12-21 212         0          0
#> 53   2020-08-07     CA          2020-08-02 2014-12-28 292         0          0
#> 54   2020-08-07     CA          2020-08-02 2015-01-04 291         0          0
#>    num_severe num_deaths incidence_rate
#> 1           0          0              0
#> 2           0          0              0
#> 3           0          0              0
#> 4           0          0              0
#> 5           0          0              0
#> 6           0          0              0
#> 7           0          0              0
#> 8           0          0              0
#> 9           0          0              0
#> 10          0          0              0
#> 11          0          0              0
#> 12          0          0              0
#> 13          0          0              0
#> 14          0          0              0
#> 15          0          0              0
#> 16          0          0              0
#> 17          0          0              0
#> 18          0          0              0
#> 19          0          0              0
#> 20          0          0              0
#> 21          0          0              0
#> 22          0          0              0
#> 23          0          0              0
#> 24          0          0              0
#> 25          0          0              0
#> 26          0          0              0
#> 27          0          0              0
#> 28          0          0              0
#> 29          0          0              0
#> 30          0          0              0
#> 31          0          0              0
#> 32          0          0              0
#> 33          0          0              0
#> 34          0          0              0
#> 35          0          0              0
#> 36          0          0              0
#> 37          0          0              0
#> 38          0          0              0
#> 39          0          0              0
#> 40          0          0              0
#> 41          0          0              0
#> 42          0          0              0
#> 43          0          0              0
#> 44          0          0              0
#> 45          0          0              0
#> 46          0          0              0
#> 47          0          0              0
#> 48          0          0              0
#> 49          0          0              0
#> 50          0          0              0
#> 51          0          0              0
#> 52          0          0              0
#> 53          0          0              0
#> 54          0          0              0
#> 
#> $result
#> [1] 1
#> 
#> $message
#> [1] "success"
#> 
# }