Read EpiData epx-Files

Reads EpiData epx-files into a data.frame or a list of data.frames.

read.EpiData(x, convert = "all")

Arguments

x

An epx-file created by EpiData.

convert

c("all", "none", "missings", "labels", "both") "missings" replaces defined missing values with NA; "labels" replaces value codes with value labels, "both" replaces missing values and value codes and "all" (default) also applies variable classes according to the field types defined in EpiData.

Value

A simple data base will be returned as a data.frame. A relational data base will be returned as a list of data.frames.

Study informations are stored as attributes(...)$study.info and variable labels are stored as attributes(...)$variable.labels. For relational data bases, key variables and parent data sets are stored as attributes(...)$info.relations.

Details

By default, object classes of the variables will be set according to the field type defined in EpiData. Coded variables will be returned as factors with the value labels defined in EpiData. Values that are defined as missing values in EpiData will be set to NA.

Examples

# Read the example data set "marathon.epx"
# (The helper function epx.example is only needed to identify the path
# of the example file in the loaded package).
dat <- read.EpiData(epx.example("SomeFakeData.epx"))
dat
#>     status ID       date age gender height weight             notes
#> 1 Verified  1 2023-11-03  65   male    189     78          No notes
#> 2  Deleted  2 2023-11-10  NA   <NA>    160     52              <NA>
#> 3   Normal  3 2023-10-30  18 female     NA     60              <NA>
#> 4   Normal  4       <NA>  39   male    170     81 This is fake data

# Extract study information
attributes(dat)$study.info
#>                      [,1]                                                                  
#> Abstract             "Anhand dieses Beispielprojektes erkläre ich das Arbeiten mit EpiData"
#> GeographicalCoverage "Hamburg"                                                             
#> Title                "Beispielprojekt"                                                     
#> UnitOfObservation    "Eine Person"                                                         
#> Author               "Johann Popp"                                                         
#> Agency               ""                                                                    
#> Created              "2019/05/21 20.20.55"                                                 
#> Identifier           ""                                                                    
#> Keywords             ""                                                                    
#> Modified             "2019/05/21 20.20.55"                                                 
#> Notes                ""                                                                    
#> Version              "1"