Fetch COVID hospitalization data for specific facilities

Obtains the COVID-19 reported patient impact and hospital capacity data by facility. This dataset is provided by the US Department of Health & Human Services via healthdata.gov.

covid_hosp_facility(hospital_pks, collection_weeks, publication_dates = NULL)

Arguments

hospital_pks

string. A character string of facility unique identifiers.

collection_weeks

epirange. Epiweeks to fetch.

publication_dates

epirange. Publication dates to fetch.

Value

epidata_call

See also the official description and data dictionary at <healthdata.gov> for more information.

Details

API docs: https://cmu-delphi.github.io/delphi-epidata/api/covid_hosp_facility.html

Starting October 1, 2022, some facilities are only required to report annually. The companion function covid_hosp_facility_lookup() can be used to look up facility identifiers in a variety of ways.

See also

Examples

# \donttest{
covid_hosp_facility(
  hospital_pks = "100075",
  collection_weeks = epirange(20200101, 20200501)
) %>% fetch_tbl()
#> Warning: No API key found. You will be limited to non-complex queries and encounter rate
#> limits if you proceed.
#> To avoid this, you can get your key [here] and then:
#> ℹ set the environment variable DELPHI_EPIDATA_KEY
#> ℹ set the option 'delphi.epidata.key'
#> 
#> To save your key for later sessions (and hide it from your code), you can edit
#> your .Renviron file with:
#> ℹ usethis::edit_r_environ()
#> This warning is displayed once every 8 hours.
#> # A tibble: 6 × 113
#>   hospital_pk state ccn    hospital_name    address city  zip   hospital_subtype
#>   <chr>       <chr> <chr>  <chr>            <chr>   <chr> <chr> <chr>           
#> 1 100075      FL    100075 ST JOSEPHS HOSP… 3001 W… TAMPA 33677 Short Term      
#> 2 100075      FL    100075 ST JOSEPHS HOSP… 3001 W… TAMPA 33677 Short Term      
#> 3 100075      FL    100075 ST JOSEPHS HOSP… 3001 W… TAMPA 33677 Short Term      
#> 4 100075      FL    100075 ST JOSEPHS HOSP… 3001 W… TAMPA 33677 Short Term      
#> 5 100075      FL    100075 ST JOSEPHS HOSP… 3001 W… TAMPA 33677 Short Term      
#> 6 100075      FL    100075 ST JOSEPHS HOSP… 3001 W… TAMPA 33677 Short Term      
#> # ℹ 105 more variables: fips_code <chr>, geocoded_hospital_address <chr>,
#> #   hhs_ids <chr>, publication_date <date>, collection_week <date>,
#> #   is_metro_micro <lgl>, total_beds_7_day_sum <int>,
#> #   all_adult_hospital_beds_7_day_sum <int>,
#> #   all_adult_hospital_inpatient_beds_7_day_sum <int>,
#> #   inpatient_beds_used_7_day_sum <int>,
#> #   all_adult_hospital_inpatient_bed_occupied_7_day_sum <int>, …
# }