Fetch COVID Hospitalization Data by State

Obtains the COVID-19 reported patient impact and hospital capacity data by state. This dataset is provided by the US Department of Health & Human Services via healthdata.gov.

covid_hosp_state_timeseries(states, dates, issues = NULL)

Arguments

states

character vector. Two letter state abbreviations.

dates

timeset. Dates to fetch.

issues

timeset. Optionally, the issues to fetch. If not set, the most recent issue is returned.

Value

epidata_call

See also the official description and data dictionary at <healthdata.gov> for more information.

Details

API docs: https://cmu-delphi.github.io/delphi-epidata/api/covid_hosp.html.

Starting October 1, 2022, some facilities are only required to report annually.

Examples

# \donttest{
covid_hosp_state_timeseries(
  states = "fl",
  dates = epirange(20200101, 20200501)
) %>% fetch_tbl()
#> Warning: One or more parsing issues, call `problems()` on your data frame for details,
#> e.g.:
#>   dat <- vroom(...)
#>   problems(dat)
#> # A tibble: 51 × 118
#>    state geocoded_state issue      date       critical_staffing_shortage_today…¹
#>    <chr> <lgl>          <date>     <date>     <lgl>                             
#>  1 FL    NA             2023-05-02 2020-03-12 FALSE                             
#>  2 FL    NA             2023-05-02 2020-03-13 FALSE                             
#>  3 FL    NA             2023-05-02 2020-03-14 FALSE                             
#>  4 FL    NA             2023-05-02 2020-03-15 FALSE                             
#>  5 FL    NA             2023-05-02 2020-03-16 FALSE                             
#>  6 FL    NA             2023-05-02 2020-03-17 FALSE                             
#>  7 FL    NA             2023-05-02 2020-03-18 FALSE                             
#>  8 FL    NA             2023-05-02 2020-03-19 FALSE                             
#>  9 FL    NA             2023-05-02 2020-03-20 FALSE                             
#> 10 FL    NA             2023-05-02 2020-03-21 FALSE                             
#> # ℹ 41 more rows
#> # ℹ abbreviated name: ¹​critical_staffing_shortage_today_yes
#> # ℹ 113 more variables: critical_staffing_shortage_today_no <lgl>,
#> #   critical_staffing_shortage_today_not_reported <lgl>,
#> #   critical_staffing_shortage_anticipated_within_week_yes <lgl>,
#> #   critical_staffing_shortage_anticipated_within_week_no <lgl>,
#> #   critical_staffing_shortage_anticipated_within_week_not_reported <lgl>, …
# }