Fetch covidcast data

API docs: https://cmu-delphi.github.io/delphi-epidata/api/covidcast.html

covidcast(
  data_source,
  signals,
  geo_type,
  time_type,
  geo_values,
  time_values,
  as_of = NULL,
  issues = NULL,
  lag = NULL
)

Arguments

data_source

string. The data source to query (see: https://cmu-delphi.github.io/delphi-epidata/api/covidcast_signals.html).

signals

string. The signals to query from a specific source (see: https://cmu-delphi.github.io/delphi-epidata/api/covidcast-signals.html).

geo_type

string. The geographic resolution of the data (see: https://cmu-delphi.github.io/delphi-epidata/api/covidcast_geography.html).

time_type

string. The temporal resolution of the data (either "day" or "week", depending on signal).

geo_values

character vector. The geographies to return. "*" fetches all. (See: https://cmu-delphi.github.io/delphi-epidata/api/covidcast_geography.html.)

time_values

timeset. The dates to fetch.

as_of

Date. Optionally, the as of date for the issues to fetch. If not specified, the most recent data is returned. Mutually exclusive with issues or lag.

issues

timeset. Optionally, the issue of the data to fetch. If not specified, the most recent issue is returned. Mutually exclusive with as_of or lag.

lag

integer. Optionally, the lag of the issues to fetch. If not set, the most recent issue is returned. Mutually exclusive with as_of or issues.

Value

epidata_call

Details

COVIDcast public dashboard: https://delphi.cmu.edu/covidcast/

See also

Examples

# \donttest{
covidcast(
  data_source = "jhu-csse",
  signals = "confirmed_7dav_incidence_prop",
  geo_type = "state",
  time_type = "day",
  geo_values = "ca,fl",
  time_values = epirange(20200601, 20200801)
) %>% fetch()
#> # A tibble: 124 × 15
#>    geo_value signal    source geo_type time_type time_value direction issue     
#>    <chr>     <chr>     <chr>  <fct>    <fct>     <date>         <dbl> <date>    
#>  1 ca        confirme… jhu-c… state    day       2020-06-01        NA 2023-03-10
#>  2 fl        confirme… jhu-c… state    day       2020-06-01        NA 2023-03-03
#>  3 ca        confirme… jhu-c… state    day       2020-06-02        NA 2023-03-10
#>  4 fl        confirme… jhu-c… state    day       2020-06-02        NA 2023-03-03
#>  5 ca        confirme… jhu-c… state    day       2020-06-03        NA 2023-03-10
#>  6 fl        confirme… jhu-c… state    day       2020-06-03        NA 2023-03-03
#>  7 ca        confirme… jhu-c… state    day       2020-06-04        NA 2023-03-10
#>  8 fl        confirme… jhu-c… state    day       2020-06-04        NA 2023-03-03
#>  9 ca        confirme… jhu-c… state    day       2020-06-05        NA 2023-03-10
#> 10 fl        confirme… jhu-c… state    day       2020-06-05        NA 2023-03-03
#> # ℹ 114 more rows
#> # ℹ 7 more variables: lag <int>, missing_value <int>, missing_stderr <int>,
#> #   missing_sample_size <int>, value <dbl>, stderr <dbl>, sample_size <dbl>
covidcast(
  data_source = "jhu-csse",
  signals = "confirmed_7dav_incidence_prop",
  geo_type = "state",
  time_type = "day",
  geo_values = "*",
  time_values = epirange(20200601, 20200801)
) %>% fetch()
#> # A tibble: 3,472 × 15
#>    geo_value signal    source geo_type time_type time_value direction issue     
#>    <chr>     <chr>     <chr>  <fct>    <fct>     <date>         <dbl> <date>    
#>  1 ak        confirme… jhu-c… state    day       2020-06-01        NA 2023-03-03
#>  2 al        confirme… jhu-c… state    day       2020-06-01        NA 2023-03-03
#>  3 ar        confirme… jhu-c… state    day       2020-06-01        NA 2023-03-03
#>  4 as        confirme… jhu-c… state    day       2020-06-01        NA 2023-03-03
#>  5 az        confirme… jhu-c… state    day       2020-06-01        NA 2023-03-03
#>  6 ca        confirme… jhu-c… state    day       2020-06-01        NA 2023-03-10
#>  7 co        confirme… jhu-c… state    day       2020-06-01        NA 2023-03-03
#>  8 ct        confirme… jhu-c… state    day       2020-06-01        NA 2023-03-03
#>  9 dc        confirme… jhu-c… state    day       2020-06-01        NA 2023-03-03
#> 10 de        confirme… jhu-c… state    day       2020-06-01        NA 2023-03-03
#> # ℹ 3,462 more rows
#> # ℹ 7 more variables: lag <int>, missing_value <int>, missing_stderr <int>,
#> #   missing_sample_size <int>, value <dbl>, stderr <dbl>, sample_size <dbl>
# }