Shared Documentation for epidatr Parameters

This is a central text for parameter documentation

Arguments

auth

string. Your restricted access key (not the same as API key).

locations

character. List of locations to fetch.

states

character. List of states to fetch, formatted as two letter state abbreviations.

regions

character. List of regions to fetch.

epiweeks

timeset. Epiweeks to fetch. Supports epirange() and defaults to all ("*") dates. Format as epirange(startweek, endweek), where startweek and endweek are of the form YYYYWW (string or numeric).

time_type

string. The temporal resolution of the data (either "day" or "week", depending on signal).

names

character. Sensor names to fetch.

dates

timeset. Dates to fetch. Supports epirange() and defaults to all ("*") dates.

time_values

timeset. Dates or epiweeks to fetch. Supports epirange() and defaults to all ("*") dates.

as_of

Date. Optionally, the as-of date for the issues to fetch. See the "Data Versioning" section for details.

issues

timeset. Optionally, the issue(s) of the data to fetch. See the "Data Versioning" section for details.

lag

integer. Optionally, the lag of the issues to fetch. See the "Data Versioning" section for details.

fetch_args

fetch_args_list(). Additional arguments to pass to fetch(). See fetch_args_list() for details.

...

not used for values, forces later arguments to bind by name

Data Versioning

Several endpoints support retrieving historical versions of the data. The following parameters control this and are mutually exclusive (only one can be provided at a time).

as_of: (Date) Retrieve the data as it was on this date.

issues: timeset Retrieve data from a specific issue date or range of dates.

lag: (integer) Retrieve data with a specific lag from its issue date.

If none of these is specified, the most recent version of the data is returned.

See vignette("versioned-data") for details and more ways to specify versioned data.

See also

For example queries showing how to discover signals and build calls, see vignette("signal-discovery", package = "epidatr").