Fetch ECDC data

Obtain information on influenza-like-illness from the European Centre for Disease Prevention and Control.

ecdc_ili(regions, epiweeks, issues = NULL, lag = NULL)

Arguments

regions

character vector. The regions to fetch.

epiweeks

timeset. The epiweeks to fetch.

issues

timeset. Optionally, the issues to fetch. If not set, the most recent issue is returned. Mutually exclusive with lag.

lag

integer. Optionally, the lag of the issues to fetch. If not set, the most recent issue is returned. Mutually exclusive with issues.

Value

epidata_call

Details

API docs: https://cmu-delphi.github.io/delphi-epidata/api/ecdc_ili.html.

The list of location argument can be found in https://github.com/cmu-delphi/delphi-epidata/blob/main/labels/ecdc_regions.txt.

Examples

# \donttest{
ecdc_ili(regions = "austria", epiweeks = epirange(201901, 202001)) %>% fetch()
#> # A tibble: 28 × 6
#>    release_date region  issue      epiweek      lag incidence_rate
#>    <date>       <chr>   <date>     <date>     <int>          <dbl>
#>  1 NA           Austria 2020-03-15 2019-01-06    62           787.
#>  2 NA           Austria 2020-03-15 2019-01-13    61           855.
#>  3 NA           Austria 2020-03-15 2019-01-20    60          1022.
#>  4 NA           Austria 2020-03-15 2019-01-27    59          1265.
#>  5 NA           Austria 2020-03-15 2019-02-03    58          1367.
#>  6 NA           Austria 2020-03-15 2019-02-10    57          1209.
#>  7 NA           Austria 2020-03-15 2019-02-17    56          1156.
#>  8 NA           Austria 2020-03-15 2019-02-24    55          1122.
#>  9 NA           Austria 2020-03-15 2019-03-03    54           948.
#> 10 NA           Austria 2020-03-15 2019-03-10    53           703.
#> # ℹ 18 more rows
# }