An abstraction that holds information needed to make an epidata request

epidata_call objects are generated by endpoint functions like covidcast; should be piped into the fetch function to fetch and format the data. For most endpoints this will return a tibble, but a few non-COVIDCAST endpoints only support will return a JSON-like list instead.

create_epidata_call is the constructor for epidata_call objects, but you should not need to use it directly; instead, use an endpoint function, e.g., covidcast, to generate an epidata_call for the data of interest.

There are some other functions available for debugging and advanced usage:

request_url (for debugging): outputs the request URL from which data would be fetched (note additional parameters below)

with_base_url (advanced): outputs an epidata_call that requests data from a different Epidata server

fetch usually returns the data in tibble format, but a few of the endpoints only support the JSON classic format (delphi, pvt_meta_norostat, and meta). In that case a JSON-like nested list structure is returned instead.

create_epidata_call(
  endpoint,
  params,
  meta = NULL,
  only_supports_classic = FALSE
)

fetch(
  epidata_call,
  fields = NULL,
  disable_date_parsing = FALSE,
  return_empty = FALSE,
  timeout_seconds = 30
)

with_base_url(epidata_call, base_url)

Arguments

endpoint

the epidata endpoint to call

params

the parameters to pass to the epidata endpoint

meta

meta data to attach to the epidata call

only_supports_classic

if true only classic format is supported

epidata_call

an instance of epidata_call

fields

a list of epidata fields to return, or NULL to return all fields (default) e.g. c("time_value", "value") to return only the time_value and value fields or c("-direction") to return everything except the direction field

disable_date_parsing

disable automatic date parsing

return_empty

boolean that allows returning an empty tibble if there is no data

timeout_seconds

the maximum amount of time to wait for a response

base_url

base URL to use

Value

For create_epidata_call: an epidata_call object

For fetch: a tibble or a JSON-like list

For with_base_url: an epidata_call object

Examples

covidcast(
  data_source = "jhu-csse",
  signals = "confirmed_7dav_incidence_prop",
  time_type = "day",
  geo_type = "state",
  time_values = epirange(20200601, 20200801),
  geo_values = c("ca", "fl")
) %>% fetch()
#> # A tibble: 124 × 15
#>    geo_value signal    source geo_type time_type time_value direction issue     
#>    <chr>     <chr>     <chr>  <fct>    <fct>     <date>         <dbl> <date>    
#>  1 ca        confirme… jhu-c… state    day       2020-06-01        NA 2023-03-10
#>  2 fl        confirme… jhu-c… state    day       2020-06-01        NA 2023-03-03
#>  3 ca        confirme… jhu-c… state    day       2020-06-02        NA 2023-03-10
#>  4 fl        confirme… jhu-c… state    day       2020-06-02        NA 2023-03-03
#>  5 ca        confirme… jhu-c… state    day       2020-06-03        NA 2023-03-10
#>  6 fl        confirme… jhu-c… state    day       2020-06-03        NA 2023-03-03
#>  7 ca        confirme… jhu-c… state    day       2020-06-04        NA 2023-03-10
#>  8 fl        confirme… jhu-c… state    day       2020-06-04        NA 2023-03-03
#>  9 ca        confirme… jhu-c… state    day       2020-06-05        NA 2023-03-10
#> 10 fl        confirme… jhu-c… state    day       2020-06-05        NA 2023-03-03
#> # ℹ 114 more rows
#> # ℹ 7 more variables: lag <int>, missing_value <int>, missing_stderr <int>,
#> #   missing_sample_size <int>, value <dbl>, stderr <dbl>, sample_size <dbl>