Fetch FluSurv hospitalization data

Obtain information on flu hospitalization rates from the Center of Disease Control.

flusurv(locations, epiweeks, issues = NULL, lag = NULL)

Arguments

locations

character vector. Character strings indicating location.

epiweeks

timeset. The epiweeks to fetch.

issues

timeset. Optionally, the issues to fetch. If not set, the most recent issue is returned. Mutually exclusive with lag.

lag

integer. Optionally, the lag of the issues to fetch. If not set, the most recent issue is returned. Mutually exclusive with issues.

Value

epidata_call

Details

API docs: https://cmu-delphi.github.io/delphi-epidata/api/flusurv.html. See also https://gis.cdc.gov/GRASP/Fluview/FluHospRates.html.

The list of location argument can be found in https://github.com/cmu-delphi/delphi-epidata/blob/main/labels/flusurv_locations.txt.

Examples

# \donttest{
flusurv(locations = "CA", epiweeks = epirange(201701, 201801)) %>% fetch_classic()
#> $epidata
#>    release_date location      issue    epiweek lag rate_age_0 rate_age_1
#> 1    2018-01-12       CA 2017-12-31 2017-01-01  52        3.4        0.6
#> 2    2018-01-19       CA 2018-01-07 2017-01-08  52        2.9        1.7
#> 3    2018-01-26       CA 2018-01-14 2017-01-15  52        3.4        1.0
#> 4    2018-02-02       CA 2018-01-21 2017-01-22  52        2.0        0.6
#> 5    2018-02-09       CA 2018-01-28 2017-01-29  52        2.0        0.4
#> 6    2018-02-16       CA 2018-02-04 2017-02-05  52        2.0        1.2
#> 7    2018-02-14       CA 2018-02-11 2017-02-12  52        2.0        0.2
#> 8    2018-03-02       CA 2018-02-18 2017-02-19  52        0.5        0.8
#> 9    2018-03-09       CA 2018-02-25 2017-02-26  52        2.0        0.2
#> 10   2018-03-16       CA 2018-03-04 2017-03-05  52        0.5        0.4
#> 11   2018-03-23       CA 2018-03-11 2017-03-12  52        0.0        0.0
#> 12   2018-03-30       CA 2018-03-18 2017-03-19  52        1.0        0.0
#> 13   2018-04-06       CA 2018-03-25 2017-03-26  52        1.0        0.0
#> 14   2018-04-13       CA 2018-04-01 2017-04-02  52        1.5        0.2
#> 15   2018-04-20       CA 2018-04-08 2017-04-09  52        0.0        0.0
#> 16   2018-04-27       CA 2018-04-15 2017-04-16  52        1.0        0.0
#> 17   2018-05-04       CA 2018-04-22 2017-04-23  52        0.5        0.0
#> 18   2018-10-12       CA 2018-09-30 2017-10-01  52        0.0        0.0
#> 19   2018-10-19       CA 2018-10-07 2017-10-08  52        0.0        0.0
#> 20   2018-10-26       CA 2018-10-14 2017-10-15  52        0.0        0.0
#> 21   2018-11-02       CA 2018-10-21 2017-10-22  52        0.0        0.0
#> 22   2018-11-09       CA 2018-10-28 2017-10-29  52        0.5        0.2
#> 23   2018-11-16       CA 2018-11-04 2017-11-05  52        0.0        0.2
#> 24   2018-11-26       CA 2018-11-11 2017-11-12  52        1.0        0.4
#> 25   2018-11-30       CA 2018-11-18 2017-11-19  52        2.5        0.0
#> 26   2018-12-07       CA 2018-11-25 2017-11-26  52        1.0        0.0
#> 27   2018-12-14       CA 2018-12-02 2017-12-03  52        0.0        0.6
#> 28   2018-12-21       CA 2018-12-09 2017-12-10  52        1.5        0.4
#> 29   2018-12-28       CA 2018-12-16 2017-12-17  52        2.5        0.2
#> 30   2019-01-04       CA 2018-12-23 2017-12-24  52        3.9        1.1
#> 31   2019-01-11       CA 2018-12-30 2017-12-31  52        4.4        1.7
#>    rate_age_2 rate_age_3 rate_age_4 rate_overall
#> 1         2.0       10.1       63.0         12.0
#> 2         2.1       10.0       52.1         10.6
#> 3         1.2        8.0       34.6          7.3
#> 4         1.1        3.4       24.1          4.8
#> 5         1.1        4.1       22.9          4.7
#> 6         0.5        2.7       15.0          3.1
#> 7         0.6        1.7        8.2          1.9
#> 8         0.2        1.7        6.4          1.5
#> 9         0.2        1.0        6.0          1.3
#> 10        0.3        1.0        5.4          1.2
#> 11        0.1        0.6        3.3          0.6
#> 12        0.1        0.4        3.7          0.7
#> 13        0.2        0.4        3.3          0.7
#> 14        0.1        0.7        3.5          0.8
#> 15        0.1        0.6        2.7          0.5
#> 16        0.1        0.9        1.6          0.5
#> 17        0.1        0.4        2.7          0.5
#> 18        0.1        0.1        1.1          0.2
#> 19        0.0        0.1        0.7          0.1
#> 20        0.0        0.4        2.6          0.5
#> 21        0.1        0.9        2.4          0.5
#> 22        0.1        0.3        1.7          0.4
#> 23        0.3        0.3        2.4          0.6
#> 24        0.2        0.3        2.1          0.5
#> 25        0.2        0.7        3.4          0.9
#> 26        0.1        0.6        3.2          0.7
#> 27        0.2        0.6        5.2          1.0
#> 28        0.5        2.3        9.1          2.1
#> 29        1.4        5.3       26.5          5.7
#> 30        2.7       12.9       74.9         15.0
#> 31        2.7       18.7       99.7         19.8
#> 
#> $result
#> [1] 1
#> 
#> $message
#> [1] "success"
#> 
# }