Fetch NIDSS flu data

Obtains information on outpatient inluenza-like-illness from Taiwan National Infectious Disease Statistical System.

nidss_flu(regions, epiweeks, issues = NULL, lag = NULL)

Arguments

regions

character vector. The regions to fetch.

epiweeks

timeset. The epiweeks to fetch.

issues

timeset. Optionally, the issues to fetch. If not set, the most recent issue is returned. Mutually exclusive with lag.

lag

integer. Optionally, the lag of the issues to fetch. If not set, the most recent issue is returned. Mutually exclusive with issues.

Value

epidata_call

Details

API docs: https://cmu-delphi.github.io/delphi-epidata/api/nidss_flu.html

Examples

# \donttest{
nidss_flu(regions = "taipei", epiweeks = epirange(201501, 201601)) %>% fetch_classic()
#> $epidata
#>    release_date region      issue    epiweek lag visits  ili
#> 1    2016-01-05 Taipei 2015-12-27 2015-01-04  51  16698 0.96
#> 2    2016-01-12 Taipei 2016-01-03 2015-01-11  51  17146 1.03
#> 3    2016-01-19 Taipei 2016-01-10 2015-01-18  51  17817 1.07
#> 4    2016-01-26 Taipei 2016-01-17 2015-01-25  51  18168 1.13
#> 5    2016-02-02 Taipei 2016-01-24 2015-02-01  51  18321 1.11
#> 6    2016-02-02 Taipei 2016-01-24 2015-02-08  50  19799 1.08
#> 7    2016-02-02 Taipei 2016-01-24 2015-02-15  49  11879 1.25
#> 8    2016-02-02 Taipei 2016-01-24 2015-02-22  48  17397 1.19
#> 9    2016-02-02 Taipei 2016-01-24 2015-03-01  47  17849 1.07
#> 10   2016-02-02 Taipei 2016-01-24 2015-03-08  46  17441 1.08
#> 11   2016-02-02 Taipei 2016-01-24 2015-03-15  45  19602 1.16
#> 12   2016-02-02 Taipei 2016-01-24 2015-03-22  44  19486 1.22
#> 13   2016-02-02 Taipei 2016-01-24 2015-03-29  43  20278 1.28
#> 14   2016-02-02 Taipei 2016-01-24 2015-04-05  42  20101 1.22
#> 15   2016-02-02 Taipei 2016-01-24 2015-04-12  41  17121 1.22
#> 16   2016-02-02 Taipei 2016-01-24 2015-04-19  40  20012 1.23
#> 17   2016-02-02 Taipei 2016-01-24 2015-04-26  39  19555 1.20
#> 18   2016-02-02 Taipei 2016-01-24 2015-05-03  38  17579 1.09
#> 19   2016-02-02 Taipei 2016-01-24 2015-05-10  37  17169 1.09
#> 20   2016-02-02 Taipei 2016-01-24 2015-05-17  36  16811 1.09
#> 21   2016-02-02 Taipei 2016-01-24 2015-05-24  35  16877 1.07
#> 22   2016-02-02 Taipei 2016-01-24 2015-05-31  34  17403 1.07
#> 23   2016-02-02 Taipei 2016-01-24 2015-06-07  33  16855 1.05
#> 24   2016-02-02 Taipei 2016-01-24 2015-06-14  32  14886 1.07
#> 25   2016-02-02 Taipei 2016-01-24 2015-06-21  31  15050 0.97
#> 26   2016-02-02 Taipei 2016-01-24 2015-06-28  30  13495 0.90
#> 27   2016-02-02 Taipei 2016-01-24 2015-07-05  29  11872 0.88
#> 28   2016-02-02 Taipei 2016-01-24 2015-07-12  28  11691 0.78
#> 29   2016-07-19 Taipei 2016-07-10 2015-07-19  51  10815 0.75
#> 30   2016-07-25 Taipei 2016-07-17 2015-07-26  51  10848 0.76
#> 31   2016-08-01 Taipei 2016-07-24 2015-08-02  51   9323 0.74
#> 32   2016-08-08 Taipei 2016-07-31 2015-08-09  51  10473 0.71
#> 33   2016-08-15 Taipei 2016-08-07 2015-08-16  51   9993 0.69
#> 34   2016-08-22 Taipei 2016-08-14 2015-08-23  51  10008 0.71
#> 35   2016-08-30 Taipei 2016-08-21 2015-08-30  51   8641 0.70
#> 36   2016-09-05 Taipei 2016-08-28 2015-09-06  51  11516 0.73
#> 37   2016-09-12 Taipei 2016-09-04 2015-09-13  51  12068 0.76
#> 38   2016-09-19 Taipei 2016-09-11 2015-09-20  51  12443 0.77
#> 39   2016-09-26 Taipei 2016-09-18 2015-09-27  51  10438 0.76
#> 40   2016-10-03 Taipei 2016-09-25 2015-10-04  51  11061 0.74
#> 41   2016-10-10 Taipei 2016-10-02 2015-10-11  51  12744 0.76
#> 42   2016-10-17 Taipei 2016-10-09 2015-10-18  51  11771 0.76
#> 43   2016-10-24 Taipei 2016-10-16 2015-10-25  51  11645 0.76
#> 44   2016-10-31 Taipei 2016-10-23 2015-11-01  51  11934 0.76
#> 45   2016-11-07 Taipei 2016-10-30 2015-11-08  51  12227 0.78
#> 46   2016-11-14 Taipei 2016-11-06 2015-11-15  51  12764 0.83
#> 47   2016-11-21 Taipei 2016-11-13 2015-11-22  51  12741 0.82
#> 48   2016-11-28 Taipei 2016-11-20 2015-11-29  51  12850 0.83
#> 49   2016-12-05 Taipei 2016-11-27 2015-12-06  51  13884 0.89
#> 50   2016-12-12 Taipei 2016-12-04 2015-12-13  51  14251 0.90
#> 51   2016-12-19 Taipei 2016-12-11 2015-12-20  51  15404 0.97
#> 52   2016-12-26 Taipei 2016-12-18 2015-12-27  51  21459 1.53
#> 53   2018-01-01 Taipei 2017-12-24 2016-01-03 103  15545 0.96
#> 
#> $result
#> [1] 1
#> 
#> $message
#> [1] "success"
#> 
# }