Various COVID and flu signals via the COVIDcast endpoint

API docs: https://cmu-delphi.github.io/delphi-epidata/api/covidcast_signals.html

The primary endpoint for fetching COVID-19 data, providing access to a wide variety of signals from a wide variety of sources. See the API documentation link above for more. Delphi's COVIDcast public dashboard is powered by this endpoint.

Usage

pub_covidcast(
  source,
  signals,
  geo_type,
  time_type,
  geo_values,
  time_values,
  ...,
  as_of = NULL,
  issues = NULL,
  lag = NULL,
  fetch_args = fetch_args_list()
)

Arguments

source

string. The data source to query (see: https://cmu-delphi.github.io/delphi-epidata/api/covidcast_signals.html).

signals

string. The signals to query from a specific source (see: https://cmu-delphi.github.io/delphi-epidata/api/covidcast_signals.html).

geo_type

string. The geographic resolution of the data (see: https://cmu-delphi.github.io/delphi-epidata/api/covidcast_geography.html).

time_type

string. The temporal resolution of the data (either "day" or "week", depending on signal).

geo_values

character. The geographies to return. "*" fetches all. (See: https://cmu-delphi.github.io/delphi-epidata/api/covidcast_geography.html.)

time_values

timeset. Dates to fetch.

...

not used for values, forces later arguments to bind by name

as_of

Date. Optionally, the as of date for the issues to fetch. If not specified, the most recent data is returned. Mutually exclusive with issues or lag.

issues

timeset. Optionally, the issue of the data to fetch. If not specified, the most recent issue is returned. Mutually exclusive with as_of or lag.

lag

integer. Optionally, the lag of the issues to fetch. If not set, the most recent issue is returned. Mutually exclusive with as_of or issues.

fetch_args

fetch_args. Additional arguments to pass to fetch().

Value

tibble::tibble

See also

Examples

if (FALSE) {
pub_covidcast(
  source = "jhu-csse",
  signals = "confirmed_7dav_incidence_prop",
  geo_type = "state",
  time_type = "day",
  geo_values = c("ca", "fl"),
  time_values = epirange(20200601, 20200801)
)
pub_covidcast(
  source = "jhu-csse",
  signals = "confirmed_7dav_incidence_prop",
  geo_type = "state",
  time_type = "day",
  geo_values = "*",
  time_values = epirange(20200601, 20200801)
)
}