Wikipedia access data

API docs: https://cmu-delphi.github.io/delphi-epidata/api/wiki.html

pub_wiki(
  articles,
  ...,
  dates = NULL,
  epiweeks = NULL,
  hours = NULL,
  language = "en",
  fetch_args = fetch_args_list()
)

Arguments

articles

character. Articles to fetch.

...

not used for values, forces later arguments to bind by name

dates

timeset. Dates to fetch. Mutually exclusive with epiweeks.

epiweeks

timeset. Epiweeks to fetch. Mutually exclusive with dates.

hours

integer. Optionally, the hours to fetch.

language

string. Language to fetch.

fetch_args

fetch_args. Additional arguments to pass to fetch().

Value

epidata_call

Delphi's COVID-19 Data Sources and Signals

Delphi's COVID-19 Surveillance Streams data includes the following data sources. Data from most of these sources is typically updated daily. You can use the covidcast_meta endpoint to get summary information about the ranges of the different attributes for the different data sources.

The API for retrieving data from these sources is described in the COVIDcast endpoint documentation. Changes and corrections to data from this endpoint are listed in the changelog.

To obtain many of these signals and update them daily, Delphi has written extensive software to obtain data from various sources, aggregate the data, calculate statistical estimates, and format the data to be shared through the COVIDcast endpoint of the Delphi Epidata API. This code is open source and available on GitHub, and contributions are welcome.

COVIDcast Dashboard Signals

The following signals are currently displayed on the public COVIDcast dashboard:

Kind	Name	Source	Signal

Public Behavior	People Wearing Masks	fb-survey	smoothed_wwearing_mask_7d

Public Behavior	Vaccine Acceptance	fb-survey	smoothed_wcovid_vaccinated_appointment_or_accept

Public Behavior	COVID Symptom Searches on Google	google-symptoms	sum_anosmia_ageusia_smoothed_search

Early Indicators	COVID-Like Symptoms	fb-survey	smoothed_wcli

Early Indicators	COVID-Like Symptoms in Community	fb-survey	smoothed_whh_cmnty_cli

Early Indicators	COVID-Related Doctor Visits	doctor-visits	smoothed_adj_cli

Cases and Testing	COVID Cases	jhu-csse	confirmed_7dav_incidence_prop

Late Indicators	COVID Hospital Admissions	hhs	confirmed_admissions_covid_1d_prop_7dav

Late Indicators	Deaths	jhu-csse	deaths_7dav_incidence_prop

All Available Sources and Signals

Beyond the signals available on the COVIDcast dashboard, numerous other signals are available through our data export tool or directly through the API:

Examples

if (FALSE) {
pub_wiki(articles = "avian_influenza", epiweeks = epirange(201501, 201601))
}